% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{AIC}
\alias{AIC}
\title{Akaike's Information Criterion}
\usage{
AIC(object, ..., k)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Optionally more objects of class \code{RprobitB_fit}.}

\item{k}{A numeric, the penalty per parameter. The default is \code{k = 2} for the
classical AIC.}
}
\value{
Either a numeric value (if just one object is provided) or a numeric vector.
}
\description{
This function calculates Akaike's Information Criterion (AIC) for an
\code{RprobitB_fit} object.
}
\details{
The AIC is computed as
\deqn{-2 \cdot \text{LL} + k \cdot \text{npar},}
where \eqn{\text{LL}} is the model's log-likelihood value at the estimated
parameters, \eqn{k} is the penalty per parameter (\eqn{k = 2} for the
classical AIC), and \eqn{npar} is the number of parameters in the fitted model.
The AIC quantifies the trade-off between over- and under-fitting, where
smaller values are preferred.
}
\examples{
data("model_train", package = "RprobitB")
AIC(model_train)

}
