% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{as_draws.RprobitB_fit}
\alias{as_draws.RprobitB_fit}
\title{Convert a fitted model to posterior draws}
\usage{
\method{as_draws}{RprobitB_fit}(x, ...)
}
\arguments{
\item{x}{[\code{RprobitB_fit}]\cr
Fitted choice model.}

\item{...}{Currently not used.}
}
\value{
A \code{draws_array} with dimensions iteration, chain, and variable.
}
\description{
Returns the canonical retained posterior draws.
}
\examples{
set.seed(1)
model <- fit(
  choice ~ x | 0, dgp_parameters = list(beta = c(x = 1)), chains = 1
)
posterior::as_draws(model)
}
\keyword{models}
