% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berserk_choice}
\alias{berserk_choice}
\title{Choice of berserking}
\format{
A data frame containing berserking choices of the White and the Black player
in 65811 bullet (1+0) games. It consists of the following columns:
\itemize{
\item \code{id_white} and \code{id_black}, unique lichess usernames of the White and the Black players for each game,
\item \code{rating_white} and \code{rating_black}, both player's lichess bullet rating at the start of each game,
\item \code{berserk_white} and \code{berserk_black}, a boolean whether the players berserked,
\item \code{winner}, the winner of the game, either \code{'white'}, \code{'black'} or \code{'draw'},
\item \code{minutes_remaining}, the number of minutes left in the tournament,
\item \code{game_id}, the game's id on lichess (each game can be accessed via \url{https://lichess.org/<game_id>}).
}
}
\source{
The data was obtained via the lichess API \url{https://lichess.org/api}
with the tournament id 'RibHfoX6' on 2022-01-25.
}
\usage{
data(berserk_choice)
}
\description{
This dataset includes the binary berserking choice of participants in the yearly
bullet arena on the online chess platform \url{https://lichess.org}.
The tournament startet at 2022-01-10 17:00:25 and lasted 240 minutes.
}
\details{
To 'berserk' is a feature on the online chess platform \url{https://lichess.org}.
Before the game starts, each player can click a 'berserk' button, after which
they lose half of their clock time, but the win is worth one extra tournament
point. See \url{https://lichess.org/tournament/help?system=arena}
for more information.
}
\keyword{dataset}
