% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{BIC}
\alias{BIC}
\title{Bayesian Information Criterion}
\usage{
BIC(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Optionally more objects of class \code{RprobitB_fit}.}
}
\value{
Either a numeric value (if just one object is provided) or a numeric vector.
}
\description{
This function calculates the Bayesian Information Criterion (BIC) or
Schwarz Information Criterion for an \code{RprobitB_fit} object.
}
\details{
The BIC is computed as
\deqn{-2 \cdot \text{LL} + \text{npar} \cdot \ln{\text{nobs}},}
where \eqn{\text{LL}} is the model's log-likelihood value at the estimated
parameters, \eqn{npar} is the number of parameters in the fitted model,
and \eqn{\text{nobs}} is the number of data points. The BIC quantifies the
trade-off between over- and under-fitting, where smaller values are preferred.
}
\examples{
data("model_train", package = "RprobitB")
BIC(model_train)

}
