% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check \code{data}}
\usage{
check_data(data, model, parm, norm)
}
\arguments{
\item{data}{A list of data information.}

\item{model}{A list of model information.}

\item{parm}{A list of true parameter values.}

\item{norm}{A list of normalization information.}
}
\value{
A list of \code{data} and \code{parm}.
}
\description{
Function that checks the input \code{data}.
}
