% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_init.R
\name{check_init}
\alias{check_init}
\title{Check \code{init}}
\usage{
check_init(init, model, parm, lcus)
}
\arguments{
\item{init}{A list of initial values for the Gibbs sampler.}

\item{model}{A list of model information.}

\item{parm}{A list of true parameter values.}

\item{lcus}{A list of latent class updating scheme parameters.}
}
\value{
\code{init}
}
\description{
Function that checks the input \code{init} and sets default values.
}
\details{
If \code{lcus$do_lcus = TRUE}, initialisation at true parameter
values is not possible.
}
