% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_latent_classes.R
\name{check_latent_classes}
\alias{check_latent_classes}
\title{Check \code{latent_classes}}
\usage{
check_latent_classes(latent_classes)
}
\arguments{
\item{latent_classes}{A list of parameters specifying the number and updating scheme of latent
classes:
\itemize{
\item \code{C}:
The number (greater or equal 1) of latent classes. Set to 1 per default.
\item \code{update}:
A boolean, determining whether to update \code{C}.
\item \code{Cmax}:
The maximum number of latent classes. Ignored if \code{update = FALSE}.
\item \code{buffer}:
The updating buffer (number of iterations to wait before the next update).
Ignored if \code{update = FALSE}.
\item \code{epsmin}:
The threshold weight for removing latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{epsmax}:
The threshold weight for splitting latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{distmin}:
The threshold difference in means for joining latent classes
(non-negative). Ignored if \code{update = FALSE}.
}}
}
\value{
The checked input \code{latent_classes}
}
\description{
Function that checks the input \code{latent_classes} and sets missing values
to default values.
}
\examples{
check_latent_classes(latent_classes = NULL)
check_latent_classes(latent_classes = list("C" = 2, "update" = TRUE,
                                           "Cmax" = 10, "buffer" = 100,
                                           "epsmin" = 0.01, "epsmax" = 0.99,
                                           "distmin" = 0.1))
}
