% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lcus.R
\name{check_lcus}
\alias{check_lcus}
\title{Check \code{lcus}}
\usage{
check_lcus(lcus, model)
}
\arguments{
\item{lcus}{A list of latent class updating scheme parameters.}

\item{model}{A list of model information.}
}
\value{
\code{lcus}
}
\description{
Function that checks the input \code{lcus} and sets default values.
}
