% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mcmc.R
\name{check_mcmc}
\alias{check_mcmc}
\title{Check \code{mcmc}}
\usage{
check_mcmc(mcmc)
}
\arguments{
\item{mcmc}{A list of Markov chain Monte Carlo parameters.}
}
\value{
\code{mcmc}
}
\description{
Function that checks the input \code{mcmc} and sets default values.
}
