% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check \code{model}}
\usage{
check_model(model, data)
}
\arguments{
\item{model}{A list of model information.}

\item{data}{A list data information.}
}
\value{
\code{model}
}
\description{
Function that checks the input \code{model} and sets default values.
}
