% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_norm.R
\name{check_norm}
\alias{check_norm}
\title{Check \code{norm}}
\usage{
check_norm(norm, model)
}
\arguments{
\item{norm}{A list of normalization information.}

\item{model}{A list of model information.}
}
\value{
\code{norm}
}
\description{
Function that checks the input \code{norm} and sets default values.
}
