% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_out.R
\name{check_out}
\alias{check_out}
\title{Check \code{out}}
\usage{
check_out(out)
}
\arguments{
\item{out}{A list of output settings.}
}
\value{
\code{out}
}
\description{
Function that checks the input \code{out} and sets default values.
}
