% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parm.R
\name{check_parm}
\alias{check_parm}
\title{Check \code{parm}}
\usage{
check_parm(parm, P_f, P_r, J)
}
\arguments{
\item{parm}{A named list of true parameter values.
\itemize{
\item \code{alpha}:
the fixed coefficient vector of length \code{P_f}
\item \code{C:}
the number (greater or equal 1) of latent classes of decision makers
(if \code{P_r = 0}, then then \code{C} is ignored)
\item \code{s}:
the vector of class weights of length \code{C}
\item \code{b}:
the matrix of class means as columns of dimension \code{P_r} x \code{C}
\item \code{Omega}:
the matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}
\item \code{Sigma}:
the error term covariance matrix of dimension \code{J} x \code{J}
}}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}
}
\value{
The checked input \code{parm}
}
\description{
Function that checks \code{parm} and draws missing parameter values.
}
\examples{
check_parm(parm = list("alpha" = 1:3), P_f = 3, P_r = 1, J = 2)
}
