% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prior.R
\name{check_prior}
\alias{check_prior}
\title{Check \code{prior}}
\usage{
check_prior(prior, model)
}
\arguments{
\item{prior}{A list of prior parameters.}

\item{model}{A list of model information.}
}
\value{
\code{prior}
}
\description{
Function that checks the input \code{prior} and sets default values.
}
