% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_saving.R
\name{check_saving}
\alias{check_saving}
\title{Saving checks}
\usage{
check_saving(out, objects)
}
\arguments{
\item{out}{A list of output settings.}

\item{objects}{A list of R objects.}
}
\value{
No return value. Saves rds-files of elements in \code{objects} in folder "\code{out[["rdir"]]/out[["id"]]}".
}
\description{
Function that checks overwriting and saves model objects.
}
\details{
Model results with \code{model[["id"]]="test"} will be overwritting.
}
