% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_scale}
\alias{check_scale}
\title{Check \code{scale}}
\usage{
check_scale(scale, P_f, J)
}
\arguments{
\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}
}
\value{
The checked input \code{scale}
}
\description{
Function that checks the input \code{scale}.
}
\details{
Per default, the first error-term variance is fixed to \code{1}, i.e.
\code{scale = list("parameter" = "s", "index" = 1, "value" = 1)}. Note that
you can set \code{"parameter" = "a"} only if the model has parameters with a
fixed coefficient (i.e. \code{P_f>0}).
}
\examples{
check_scale(scale = NULL, P_f = 2, J = 3)
}
