% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_probabilities.R
\name{choice_probabilities}
\alias{choice_probabilities}
\title{Return choice probabilities of an \code{RprobitB_fit}.}
\usage{
choice_probabilities(x, data = NULL, at_true = FALSE)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{data}{Either \code{NULL} or an object of class \code{RprobitB_data}. In the former
case, choice probabilities are computed for the data that was used for model
fitting. Alternatively, a new data set can be provided.}

\item{at_true}{If \code{TRUE}, choice probabilities are computed at the true parameter
values (if they are available).}
}
\value{
A data frame of choice probabilities with choice situations in rows and
alternatives in columns. The first two columns are the decider identifier
\code{"id"} and the choice situation identifier \code{"idc"}.
}
\description{
This function returns the choice probabilities of an \code{RprobitB_fit}.
}
\examples{
data <- simulate_choices(form = choice ~ covariate, N = 10, T = 10, J = 2)
x <- mcmc(data)
choice_probabilities(x)
}
