% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_probs.R
\name{choice_probs}
\alias{choice_probs}
\title{Compute choice probabilities of an \code{RprobitB_model}.}
\usage{
choice_probs(object, at_true = FALSE)
}
\arguments{
\item{object}{An object of class \code{RprobitB_model}.}

\item{at_true}{If \code{TRUE}, choice probabilities are computed at the true parameters.}
}
\value{
A data frame, choice situations in rows and alternatives in columns.
}
\description{
This function computes the choice probabilities of an \code{RprobitB_model}.
}
