% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{class_update_wb}
\alias{class_update_wb}
\title{Weight-based update of latent classes}
\usage{
class_update_wb(Cmax, epsmin, epsmax, distmin, s, m, b, Omega)
}
\arguments{
\item{Cmax}{}

\item{epsmin}{}

\item{epsmax}{}

\item{distmin}{}

\item{s}{The vector of class weights of length \code{C}.
Set to \code{NA} if \code{P_r = 0}.
For identifiability, the vector must be non-ascending.}

\item{b}{The matrix of class means as columns of dimension \code{P_r} x \code{C}.
Set to \code{NA} if \code{P_r = 0}.}

\item{Omega}{The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
Set to \code{NA} if \code{P_r = 0}.}
}
\value{
A list of updated values for \code{s}, \code{b}, \code{Omega},
}
\description{
This function updates the latent classes based on their class weights.
}
\examples{
euc_dist(c(0,0),c(1,1))
}
\keyword{internal}
