% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify deciders.}
\usage{
classify(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}
}
\value{
A data frame with the deciders id and the latent class number.
}
\description{
This function classifies deciders based on the estimated latent classes.
}
