% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Linear coefficients}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{RprobitB_coef}.
}
\description{
This function returns the estimated linear coefficients.
}
