% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare fitted models.}
\usage{
compare(...)
}
\arguments{
\item{...}{One or more objects of class \code{RprobitB_model}.}
}
\value{
A matrix with each model's
\itemize{
\item number of parameters,
\item log-likelihood,
\item AIC, and
\item BIC value.
}
}
\description{
This function computes model comparison criteria.
}
