% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_choice_probabilities.R
\name{compute_choice_probabilities}
\alias{compute_choice_probabilities}
\title{Compute probit choice probabilities for a single choice situation.}
\usage{
compute_choice_probabilities(X, parameter)
}
\arguments{
\item{X}{A matrix of covariates with one row per alternative.
The first \code{P_f} columns are connected to fixed coefficients, the last
\code{P_r} columns are connected to random coefficients.}

\item{parameter}{An object of class \code{RprobitB_parameter}.}
}
\value{
A probability vector of length \code{C}.
}
\description{
This function computes the probit choice probabilities for a single choice
situation.
}
\examples{
X = matrix(1,3)
parameter = RprobitB_parameter(P_f = 1, P_r = 0, J = 3,
                               alpha = 1, Sigma_full = diag(3))
prob = compute_choice_probabilities(X = X, parameter = parameter)
}
