% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_log_likelihood.R
\name{compute_log_likelihood}
\alias{compute_log_likelihood}
\title{Compute log-likelihood of an \code{RprobitB_model}.}
\usage{
compute_log_likelihood(object, at_true = TRUE)
}
\arguments{
\item{object}{An object of class \code{RprobitB_model}.}

\item{at_true}{If \code{TRUE}, choice probabilities are computed for the true parameter
values.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood of an \code{RprobitB_model}.
}
\examples{
p = simulate(form = choice ~ var | 0, N = 100, T = 10, J = 2, seed = 1)
m1 = mcmc(data = p, seed = 1)
compute_log_likelihood(m1)
}
