% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_number_of_linear_coefficients.R
\name{compute_number_of_linear_coefficients}
\alias{compute_number_of_linear_coefficients}
\title{Compute number of linear coefficients.}
\usage{
compute_number_of_linear_coefficients(vars, ASC, J, re)
}
\arguments{
\item{vars}{The element \code{vars} in the output of \link{check_form}.}

\item{ASC}{A boolean, determining whether the model has ASCs.}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}
}
\value{
A list with two elements:
\itemize{
\item \code{P_f}:
The number of linear coefficients with fixed effect (can be 0).
\item \code{P_r}:
The number of linear coefficients with random effect (can be 0).
}
}
\description{
This function computes the numbers \code{P_f} and \code{P_r} of linear
coefficients with fixed and random effects, respectively.
}
