% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{compute_p_si}
\alias{compute_p_si}
\title{Compute probability for each observation for different posterior samples}
\usage{
compute_p_si(x, S, print_progress = TRUE, ncores = 1)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{S}{The number of posterior samples used for the calculation of the WAIC value.
Must be greater or equal two for variance computation.}

\item{print_progress}{Set to \code{TRUE} to print computation progress.}

\item{ncores}{Computation is parallelized, set the number of cores.}
}
\value{
A matrix, observations in rows and posterior samples in columns.
}
\description{
This function computes the probability for each observation for different
samples from the posterior.
}
\keyword{internal}
