% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_parameter_statistics.R
\name{compute_parameter_statistics}
\alias{compute_parameter_statistics}
\title{Compute parameter statistics.}
\usage{
compute_parameter_statistics(gibbs_samples, P_f, P_r, J, C)
}
\arguments{
\item{gibbs_samples}{An object of class \code{RprobitB_gibbs_samples}, i.e. the output of
\code{\link{transform_gibbs_samples}}.}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{C}{The number (greater or equal 1) of latent classes.}
}
\value{
An object of class \code{RprobitB_parameter_statistics}, which is a list of
\itemize{
\item A list of statistics from the Gibbs samples for each model parameter
\code{alpha}, \code{s}, \code{b}, \code{Omega}, \code{Sigma} with:
\itemize{
\item the parameter indices as rows
\item \code{mean}, standard deviation \code{sd}, Tukey's
five number summary obtained from
\code{\link[stats]{fivenum}} (i.e. minimum \code{min},
lower-hinge \code{q.25}, \code{median},
upper-hinge \code{q.75}, maximum \code{max}), and the
Gelman-Rubin statistic \code{R_hat} as columns
}
\item \code{C_est}, the estimated number of latent classes.
}
}
\description{
This function computes parameter statistics of the Gibbs samples.
}
