% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pis.R
\name{compute_pis}
\alias{compute_pis}
\title{Compute probability for each observation for different samples from the posterior.}
\usage{
compute_pis(x, no_samples = 2)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}

\item{no_samples}{The number of different samples taken from the posterior.}
}
\value{
A matrix.
}
\description{
This function computes the probability for each observation for different samples
from the posterior.
}
