% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_point_estimates.R
\name{compute_point_estimates}
\alias{compute_point_estimates}
\title{Compute point estimates of an \code{RprobitB_model}.}
\usage{
compute_point_estimates(object)
}
\arguments{
\item{object}{An object of class \code{RprobitB_model}.}
}
\value{
An object of class \code{RprobitB_parameter}.
}
\description{
This function computes the point estimates of an \code{RprobitB_model} based
on the means of the Gibbs samples.
}
