% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_suff_statistics.R
\name{compute_suff_statistics}
\alias{compute_suff_statistics}
\title{Sufficient statistics}
\usage{
compute_suff_statistics(model, data)
}
\arguments{
\item{model}{A list of model identifications.}

\item{data}{A list of data information.}
}
\value{
A list of sufficient statistics.
}
\description{
Function that computes sufficient statistics for estimation.
}
