% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sufficient_statistics.R
\name{compute_sufficient_statistics}
\alias{compute_sufficient_statistics}
\title{Compute sufficient statistics.}
\usage{
compute_sufficient_statistics(data, normalization)
}
\arguments{
\item{data}{An object of class \code{RprobitB_data}.}

\item{normalization}{An object of class \code{RprobitB_normalization}.}
}
\value{
A list of sufficient statistics.
}
\description{
This function computes sufficient statistics from \code{data} for estimation.
}
\examples{
data = simulate(form = choice ~ var, N = 10, T = 10, J = 3, re = "ASC")
normalization = RprobitB_normalization(J = data$J, P_f = data$P_f)
compute_sufficient_statistics(data = data, normalization)
}
