% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic.R
\name{compute_waic}
\alias{compute_waic}
\title{WAIC}
\usage{
compute_waic(model, gibbs_samples, data)
}
\arguments{
\item{model}{A list of model information.}

\item{gibbs_samples}{A list of gibbs samples.}

\item{data}{A list of data information.}
}
\value{
WAIC value (numeric).
}
\description{
Function that computes the Widely Applicable Information Criterion (WAIC).
}
