% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{cov_mix}
\alias{cov_mix}
\title{Extract estimated covariance matrix of mixing distribution}
\usage{
cov_mix(x, cor = FALSE)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{cor}{If \code{TRUE}, returns the correlation matrix instead.}
}
\value{
The estimated covariance matrix of the mixing distribution. In case of
multiple classes, a list of matrices for each class.
}
\description{
This convenience function returns the estimated covariance matrix of the
mixing distribution.
}
