% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_data.R
\name{cov_vec_to_mat}
\alias{cov_vec_to_mat}
\title{Build covariate matrix from covariate vector}
\usage{
cov_vec_to_mat()
}
\description{
This helper function builds a covariate matrix from a covariate vector.
}
