% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labels.R
\name{create_labels}
\alias{create_labels}
\title{Create coefficient labels}
\usage{
create_labels(P_f, P_r, J, C, C_est, symmetric)
}
\arguments{
\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{C}{The number (greater or equal 1) of latent classes.}

\item{C_est}{The estimated number (greater or equal 1) of latent classes.}

\item{symmetric}{A boolean, determining whether labels for symmetric matrix elements should be
created.}
}
\value{
A list of labels for the model coefficients \code{alpha}, \code{s}, \code{b},
\code{Omega}, and \code{Sigma}.
}
\description{
Function that creates labels for the model coefficients.
}
\examples{
create_labels(P_f = 1, P_r = 2, J = 3, C = 4, C_est = 5, symmetric = TRUE)
}
