% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{create_labels_alpha}
\alias{create_labels_alpha}
\title{Create labels for \code{alpha}.}
\usage{
create_labels_alpha(P_f)
}
\value{
A vector of labels for the model parameter \code{alpha}.
}
\description{
This function creates labels for the model parameter \code{alpha}.
}
\keyword{internal}
