% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{create_labels_b}
\alias{create_labels_b}
\title{Create labels for \code{b}.}
\usage{
create_labels_b(P_r, C)
}
\arguments{
\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{C}{The number of latent classes.}
}
\value{
A vector of labels for the model parameter \code{b}.
}
\description{
This function creates labels for the model parameter \code{b}.
}
\keyword{internal}
