% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{create_labels_d}
\alias{create_labels_d}
\title{Create labels for \code{d}}
\usage{
create_labels_d(J, ordered)
}
\arguments{
\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{ordered}{A boolean, \code{FALSE} per default. If \code{TRUE}, the choice set
\code{alternatives} is assumed to be ordered from worst to best.}
}
\value{
A vector of labels for the model parameter \code{d} of length \code{J - 2} if
\code{ordered = TRUE} and \code{NULL} otherwise.
}
\description{
This function creates labels for the model parameter \code{d}.
}
\details{
Note that \code{J} must be greater or equal \code{3} in the ordered probit
model.
}
\examples{
RprobitB:::create_labels_d(5, TRUE)
}
\keyword{internal}
