% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{create_labels_s}
\alias{create_labels_s}
\title{Create labels for \code{s}.}
\usage{
create_labels_s(P_r, C)
}
\value{
A vector of labels for the model parameter \code{s}.
}
\description{
This function creates labels for the model parameter \code{s}.
}
\keyword{internal}
