% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{create_labels_Sigma}
\alias{create_labels_Sigma}
\title{Create labels for \code{Sigma}.}
\usage{
create_labels_Sigma(J, cov_sym)
}
\value{
A vector of labels for the model parameter \code{Sigma}.
}
\description{
This function creates labels for the model parameter \code{Sigma}.
}
\keyword{internal}
