% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_labels.R
\name{create_parameter_labels}
\alias{create_parameter_labels}
\title{Create parameters labels.}
\usage{
create_parameter_labels(P_f, P_r, J, C, cov_sym, drop_par)
}
\arguments{
\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{C}{The number of latent classes.}

\item{cov_sym}{If \code{TRUE} considers symmetric covariance matrix elements.}

\item{drop_par}{A vector of parameter names which get dropped.}
}
\value{
A list of labels for the model parameters.
}
\description{
This function creates labels for the model parameters.
}
\examples{
create_parameter_labels(P_f = 0, P_r = 2, J = 3, C = 4, cov_sym = TRUE, drop_par = "s")
}
\keyword{internal}
