% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_to_gamma}
\alias{d_to_gamma}
\title{Transform increments to thresholds}
\usage{
d_to_gamma(d)
}
\arguments{
\item{d}{[\code{numeric(J - 2)}]\cr
Threshold increments.}
}
\value{
The threshold vector of length \code{J + 1}.
}
\description{
Transform increments to thresholds
}
\examples{
d_to_gamma(c(0, 0, 0))

}
\keyword{gibbs_sampler}
