% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{determine_data_sizes}
\alias{determine_data_sizes}
\title{Determine data sizes}
\usage{
determine_data_sizes(dependent_variable, independent_variable)
}
\arguments{
\item{dependent_variable}{A \code{list} that contains the dependent variables.
It is of length \code{N}, where \code{N} is the number of individuals.
The \code{n}-th element of \code{dependent_variable} is a \code{vector} of
length \code{Tp[n]}, where the \code{t}-th element is an \code{integer} that
refers to the dependent variable from the underlying choice set.}

\item{independent_variable}{A \code{list} that contains the independent variables.
It is of length \code{N}, where \code{N} is the number of individuals.
The \code{n}-th element of \code{independent_variable} is a \code{list} of
length \code{Tp[n]}, where the \code{t}-th element is the design
\code{matrix} for time point \code{t}.}
}
\value{
A \code{list} with three elements:
\itemize{
\item \code{N}, the number of individuals
\item \code{Tp}, the \code{vector} of time points per individual
\item \code{cs_Tp}, the cumulative sum of \code{Tp}
}
}
\description{
Determine data sizes
}
