% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm_cpp}
\alias{dmvnorm_cpp}
\title{Compute density of multivariate normal distribution}
\usage{
dmvnorm_cpp(x, mean, Sigma, log = FALSE)
}
\arguments{
\item{x}{A \code{numeric}, a quantile vector of length \code{n}.}

\item{mean}{A \code{numeric}, the mean vector of length \code{n}.}

\item{Sigma}{A \code{matrix}, the covariance matrix of dimension \code{n} x \code{n}.}

\item{log}{A \code{logical}, if \code{TRUE} the logarithm of the density value is
returned.
By default, \code{log = FALSE}.}
}
\value{
A \code{numeric}, the density value.
}
\description{
This function computes the density of a multivariate normal distribution.
}
\details{
This function performs no input checks. See \code{\link{dmvnorm}}
for the version with input checks.
}
\examples{
x <- c(0,0)
mean <- c(0,0)
Sigma <- diag(2)
RprobitB:::dmvnorm_cpp(x = x, mean = mean, Sigma = Sigma)
RprobitB:::dmvnorm_cpp(x = x, mean = mean, Sigma = Sigma, log = TRUE)

}
\keyword{cpp}
\keyword{internal}
