% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dwishart}
\alias{dwishart}
\title{Compute density of (Inverse-) Wishart distribution}
\usage{
dwishart(x, df, scale, log = FALSE, inv = FALSE)
}
\arguments{
\item{x}{A \code{matrix}, a covariance matrix of dimension \code{n} x \code{n}.}

\item{df}{An \code{integer}, the degrees of freedom.
Must be greater or equal \code{n}.}

\item{scale}{A \code{matrix}, the scale matrix of dimension \code{n} x \code{n}.
Must be a covariance matrix.}

\item{log}{A \code{logical}, if \code{TRUE} the logarithm of the density value is
returned.
By default, \code{log = FALSE}.}

\item{inv}{A \code{logical}, if \code{TRUE} the density of the Inverse-Wishart
distribution is returned.
By default, \code{inv = FALSE}.}
}
\value{
A \code{numeric}, the density value.
}
\description{
This function computes the density of the (Inverse-) Wishart distribution.
}
\examples{
x <- diag(2)
df <- 4
scale <- diag(2)
dwishart(x = x, df = df, scale = scale)
dwishart(x = x, df = df, scale = scale, log = TRUE)
dwishart(x = x, df = df, scale = scale, inv = TRUE)

}
\keyword{internal}
\keyword{utils}
