% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{end_timer}
\alias{end_timer}
\title{Routine to terminate the timer for Gibbs sampling}
\usage{
end_timer(R)
}
\arguments{
\item{R}{An integer, total number of iterations}
}
\value{
No return value, called for on-screen information.
}
\description{
Routine to terminate the timer for Gibbs sampling
}
