% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance}
\usage{
euc_dist(a, b)
}
\arguments{
\item{a, b}{[\code{numeric()}]\cr
Numeric vectors of the same length.}
}
\value{
The Euclidean distance.
}
\description{
This helper function computes the Euclidean distance between two vectors.
}
\examples{
euc_dist(c(0, 1), c(1, 0))

}
\keyword{utils}
