% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_covariates.R
\name{expand_T}
\alias{expand_T}
\title{Expand \code{T}}
\usage{
expand_T(N, T = 1)
}
\arguments{
\item{N}{A positive \code{integer}, the number of deciders.}

\item{T}{A positive \code{integer}, the number of choice occasions per decider.
Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.
By default, \code{T = 1}.}
}
\value{
An \code{integer} \code{vector} of length \code{N}.
}
\description{
This function expands the number of choice occasions \code{T} to a
\code{vector} of length \code{N}.
}
\examples{
\dontrun{
expand_T(N = 10, T = 2)
expand_T(N = 10, T = 1:10)
}

}
\keyword{internal}
