% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gibbs_samples.R
\name{filter_gibbs_samples}
\alias{filter_gibbs_samples}
\title{Filter Gibbs samples.}
\usage{
filter_gibbs_samples(
  x,
  P_f,
  P_r,
  J,
  C,
  cov_sym,
  keep_par = c("s", "alpha", "b", "Omega", "Sigma"),
  drop_par = NULL
)
}
\arguments{
\item{x}{An object of class \code{RprobitB_gibbs_samples}.}
}
\value{
An object of class \code{RprobitB_gibbs_samples} filtered by the labels
\code{create_parameter_labels(P_f, P_r, J, C, cov_sym, keep_par, drop_par)}.
}
\description{
This is a helper function that filters Gibbs samples.
}
\keyword{internal}
