% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov.R
\name{get_cov}
\alias{get_cov}
\title{Extract covariates of choice occasion}
\usage{
get_cov(x, id = NULL, idc = NULL, id_label = NULL, idc_label = NULL)
}
\arguments{
\item{x}{Either an object of class \code{RprobitB_data} or \code{RprobitB_fit}.}

\item{id, idc}{[\code{integer()} | \code{NULL}]\cr
Identifiers for deciders and choice occasions.

If \code{NULL}, everything is returned.}

\item{id_label, idc_label}{[\code{character(1)} | \code{NULL}]\cr
The columns that contain the decider and choice occasion identifier.

If \code{NULL}, this information is extracted from \code{x}.}
}
\value{
A subset of the \code{choice_data} data frame specified in \code{prepare_data()}.
}
\description{
This helper function returns the covariates and the choices of specific
choice occasions.
}
\examples{
data <- simulate_choices(
  form = product ~ price,
  N = 10,
  T = 1:10,
  J = 3,
  ranked = TRUE
)
get_cov(data, id = 2)
}
