% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_estimates.R
\name{gibbs_estimates}
\alias{gibbs_estimates}
\title{Point estimates}
\usage{
gibbs_estimates(gibbs_samples, model, parm)
}
\arguments{
\item{gibbs_samples}{A list of Gibbs samples.}

\item{model}{A list of model information.}

\item{parm}{A list of true parameter values.}
}
\value{
A list of model estimates.
}
\description{
Function that computes and prints point estimates, standard deviations and 5\% and 95\% quantiles.
}
