% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampling}
\alias{gibbs_sampling}
\title{Gibbs sampler}
\usage{
gibbs_sampling(
  R,
  B,
  print_progress,
  N,
  J,
  P_f,
  P_r,
  latent_classes,
  suff_statistics,
  prior,
  init
)
}
\arguments{
\item{R}{The number of Gibbs sampler iterations.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{print_progress}{A boolean, determining whether to print the progress.}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{latent_classes}{A list of parameters specifying the number and updating scheme of latent
classes:
\itemize{
\item \code{C}:
The number (greater or equal 1) of latent classes. Set to 1 per default.
\item \code{update}:
A boolean, determining whether to update \code{C}.
\item \code{Cmax}:
The maximum number of latent classes. Ignored if \code{update = FALSE}.
\item \code{buffer}:
The updating buffer (number of iterations to wait before the next update).
Ignored if \code{update = FALSE}.
\item \code{epsmin}:
The threshold weight for removing latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{epsmax}:
The threshold weight for splitting latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{distmin}:
The threshold difference in means for joining latent classes
(non-negative). Ignored if \code{update = FALSE}.
}}

\item{suff_statistics}{The output of \code{\link{compute_suff_statistics}}.}

\item{prior}{A list of parameters for the prior distributions of the normalized
parameters:
\itemize{
\item \code{eta}:
The mean vector of length \code{P_f} of the normal prior for
\code{alpha}.
\item \code{Psi}:
The covariance matrix of dimension \code{P_f} x \code{P_f} of the
normal prior for \code{alpha}.
\item \code{delta}:
The concentration parameter of length 1 of the Dirichlet prior for
\code{s}.
\item \code{xi}:
The mean vector of length \code{P_r} of the normal prior for each
\code{b_c}.
\item \code{D}:
The covariance matrix of dimension \code{P_r} x \code{P_r} of the
normal prior for each \code{b_c}.
\item \code{nu}:
The degrees of freedom (a natural number greater than \code{P_r}) of
the Inverse Wishart prior for each \code{Omega_c}.
\item \code{Theta}:
The scale matrix of dimension \code{P_r} x \code{P_r} of the
Inverse Wishart prior for each \code{Omega_c}.
\item \code{kappa}:
The degrees of freedom (a natural number greater than \code{J-1}) of
the Inverse Wishart prior for \code{Sigma}.
\item \code{E}:
The scale matrix of dimension \code{J-1} x \code{J-1} of the
Inverse Wishart prior for \code{Sigma}.
}}

\item{init}{The output of \code{\link{set_init}}.}
}
\value{
A list of Gibbs samples
}
\description{
Function to perform Gibbs sampling for the LCMMNP model
}
