% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_trace.R
\name{gibbs_trace}
\alias{gibbs_trace}
\title{Trace plots}
\usage{
gibbs_trace(gibbs_samples, model, mcmc, out)
}
\arguments{
\item{gibbs_samples}{A list of Gibbs samples.}

\item{model}{A list of model information.}

\item{mcmc}{A list of Markov chain Monte Carlo parameters.}

\item{out}{A list of output settings.}
}
\value{
No return value. Creates pdf-file "trace.pdf" in folder "\code{out[["rdir"]]/out[["id"]]}".
}
\description{
Function that plots traces of the Gibbs samples.
}
