% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_transform.R
\name{gibbs_transform}
\alias{gibbs_transform}
\title{Transformation of Gibbs samples}
\usage{
gibbs_transform(gibbs_loop_out, model, mcmc, norm)
}
\arguments{
\item{gibbs_loop_out}{Output of \code{gibbs_loop}.}

\item{model}{A list of model information.}

\item{mcmc}{A list of Markov chain Monte Carlo parameters.}

\item{norm}{A list of normalization information.}
}
\value{
A list of normalized, burned and thinned Gibbs samples.
}
\description{
Function that normalizes, burnes and thins the Gibbs samples.
}
