% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_bool}
\alias{is_bool}
\title{check for boolean}
\usage{
is_bool(i)
}
\arguments{
\item{i}{any element}
}
\value{
\code{TRUE} if \code{i} is either \code{TRUE} or \code{FALSE}, \code{FALSE} else
}
\description{
check for boolean
}
