% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cov}
\alias{is_cov}
\title{check for covariance matrix}
\usage{
is_cov(M)
}
\arguments{
\item{M}{any element}
}
\value{
\code{TRUE} if \code{M} is a covariance matrix, \code{FALSE} else
}
\description{
check for covariance matrix
}
