% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cov_matrix}
\alias{is_cov_matrix}
\title{Check covariance matrix}
\usage{
is_cov_matrix(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A \code{matrix}.}

\item{tol}{A \code{numeric}, a numeric tolerance value.
Per default, \code{tol = sqrt(.Machine$double.eps)}.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is a proper covariance matrix,
\code{FALSE} else.
}
\description{
This function checks whether the input is a proper covariance matrix.
}
\details{
A proper covariance matrix is a square, symmetric, numeric matrix with
non-negative eigenvalues.
}
\examples{
x <- RprobitB:::sample_cov_matrix(dim = 3)
RprobitB:::is_cov_matrix(x)

}
\keyword{internal}
\keyword{utils}
