% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_covariance_matrix}
\alias{is_covariance_matrix}
\title{Check for covariance \code{matrix}}
\usage{
is_covariance_matrix(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A \code{matrix}.}

\item{tol}{A \code{numeric}, a tolerance value.
By default, \code{tol = sqrt(.Machine$double.eps)}.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is a proper covariance
\code{matrix}, \code{FALSE} else.
}
\description{
This function checks whether the input is a proper covariance \code{matrix}.
}
\details{
A proper covariance \code{matrix} is a square, symmetric, \code{numeric}
\code{matrix} with non-negative eigenvalues.
}
\examples{
\dontrun{
x <- sample_covariance_matrix(dim = 3)
is_covariance_matrix(x)
}

}
\keyword{internal}
\keyword{utils}
