% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_int}
\alias{is_int}
\title{check for positive integer}
\usage{
is_int(i)
}
\arguments{
\item{i}{any element}
}
\value{
\code{TRUE} if \code{i} is an integer, \code{FALSE} else
}
\description{
check for positive integer
}
