% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pos_int}
\alias{is_pos_int}
\title{Check positive integer}
\usage{
is_pos_int(i)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{i} is a single positive integer,
\code{FALSE} else.
}
\description{
This function checks whether the input is a single positive integer.
}
\examples{
RprobitB:::is_pos_int(-1.5)
RprobitB:::is_pos_int(1)

}
\keyword{internal}
\keyword{utils}
