% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_positive_integer}
\alias{is_positive_integer}
\title{Check for positive \code{integer}}
\usage{
is_positive_integer(i)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{i} is a single positive integer,
\code{FALSE} else.
}
\description{
This function checks whether the input is a single positive \code{integer}.
}
\examples{
\dontrun{
is_positive_integer(-1.5)
is_positive_integer(1)
is_positive_integer(1:2)
}

}
\keyword{internal}
\keyword{utils}
