% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_single_numeric}
\alias{is_single_numeric}
\title{Check for single \code{numeric}}
\usage{
is_single_numeric(i)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is a single \code{numeric},
\code{FALSE} else.
}
\description{
This function checks whether the input is a single \code{numeric} value.
}
\examples{
\dontrun{
is_single_numeric(1)
is_single_numeric("1")
is_single_numeric(NA_real_)
is_single_numeric(1:2)
}

}
\keyword{internal}
\keyword{utils}
