% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_sn}
\alias{is_sn}
\title{Check for single numeric}
\usage{
is_sn(i)
}
\arguments{
\item{x}{Any element.}
}
\value{
\code{TRUE} if \code{x} is a single numeric, \code{FALSE} else.
}
\description{
This function checks whether the input is a single numeric value.
}
\examples{
is_sn(1)
is_sn("1")
is_sn(NA_real_)
}
\keyword{utils}
