% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_true_false}
\alias{is_true_false}
\title{Check for \code{TRUE} or \code{FALSE}}
\usage{
is_true_false(x)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is either \code{TRUE} or
\code{FALSE}, \code{FALSE} else.
}
\description{
This function checks whether the input is \code{TRUE} or \code{FALSE}.
}
\examples{
\dontrun{
is_true_false("TRUE")
is_true_false(FALSE)
is_true_false(c(TRUE, FALSE))
}

}
\keyword{internal}
\keyword{utils}
