% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_d}
\alias{ll_d}
\title{Conditional choice probability in the ordered probit model}
\usage{
ll_d(d, y, mu, log)
}
\arguments{
\item{d}{A numeric vector of threshold increments.}

\item{y}{}

\item{mu}{}

\item{log}{A boolean, if \code{TRUE} the logged probability is returned.}
}
\value{
The choice probability.
}
\description{
This function computes the conditional probability of one choice occasion
given the threshold increments \code{d}.
}
\details{

}
\examples{
ll_d(c(0,0,0), 1, 1, FALSE)
}
\keyword{posterior}
