% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute log-likelihood of an \code{RprobitB_fit}.}
\usage{
log_likelihood(x, par_set = mean)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{par_set}{Specifying the parameter set for calculation and either
\itemize{
\item a function that computes a posterior point estimate (the default is
\code{mean()}),
\item \code{"true"} to select the true parameter set,
\item an object of class \code{RprobitB_parameter}.
}}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood of an \code{RprobitB_fit}.
}
\examples{
data <- simulate_choices(form = choice ~ covariate, N = 10, T = 10, J = 2)
x <- mcmc(data)
log_likelihood(x)
}
