% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{logLik}
\alias{logLik}
\title{Log-likelihood value}
\usage{
logLik(object, par_set, recompute, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{par_set}{Specifying the parameter set for calculation and either
\itemize{
\item a function that computes a posterior point estimate (the default is
\code{mean()}),
\item \code{"true"} to select the true parameter set,
\item an object of class \code{RprobitB_parameter}.
}}

\item{recompute}{Set to \code{TRUE} to recompute the log-likelihood value if it is already
saved in \code{object}.}

\item{...}{Ignored.}
}
\value{
A numeric.
}
\description{
This function computes the log-likelihood value of an \code{RprobitB_fit}
object.
}
\examples{
data("model_train", package = "RprobitB")
logLik(model_train)

}
