% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{M}
\alias{M}
\title{Matrix difference operator for ranked vectors}
\usage{
M(ranking)
}
\arguments{
\item{ranking}{A numeric vector of the ranking in decreasing order.}
}
\value{
A matrix of dimension \code{length(rank)-1} x \code{length(rank)}.
}
\description{
This function creates the difference operator matrix for differencing ranked
vector elements such that the resulting vector is negative.
}
\examples{
x <- c(-1,5,10,2)
ranking <- order(x, decreasing = TRUE)
RprobitB:::M(ranking) \%*\% x

}
\keyword{utils}
