% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels}
\alias{make_labels}
\title{Coefficient labels}
\usage{
make_labels(gibbs_samples, model, symm)
}
\arguments{
\item{gibbs_samples}{A list of Gibbs samples.}

\item{model}{A list of model information.}

\item{symm}{A boolean, determining whether labels for symmetric matrix elements should be added.}
}
\value{
A list of model coefficient labels.
}
\description{
Function that creates labels for the model coefficients.
}
