% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{mod_sel}
\alias{mod_sel}
\title{An example output of \code{model_selection()}.}
\format{
An object of class \code{RprobitB_model_selection} (inherits from \code{data.frame}) with 2 rows and 9 columns.
}
\usage{
mod_sel
}
\description{
An example output of \code{model_selection()}.
}
\keyword{internal}
