% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.frame.RprobitB_fit}
\alias{model.frame.RprobitB_fit}
\title{Extract the fitted data}
\usage{
\method{model.frame}{RprobitB_fit}(formula, ...)
}
\arguments{
\item{formula}{[\code{RprobitB_fit}]\cr
Fitted choice model.}

\item{...}{Currently not used.}
}
\value{
A \code{data.frame} containing the fitted choice data.
}
\description{
Returns the choice data stored in a fitted model as a \code{data.frame}.
}
\examples{
set.seed(1)
model <- fit(
  choice ~ x, dgp_parameters = list(beta = c(x = 1, ASC_B = -0.5)),
  chains = 1
)
head(model.frame(model))
}
\keyword{models}
