% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_berserk}
\alias{model_berserk}
\title{Latent class mixed probit model for binary berserking choice}
\format{
An object of class \code{RprobitB_fit}.
}
\usage{
data(model_berserk)
}
\description{
This object is a fitted mixed probit model to the \code{choice_berserk}
dataset of the {RprobitB} package with the formula
\code{berserk ~ 0 | white + rating + rating_diff + min_rem + streak + berserk.1 + lost.1 + 1}.
}
\details{
The model was estimated via
\preformatted{
choice_berserk <- create_lagged_cov(
  choice_data = RprobitB::choice_berserk,
  column = c("berserk","lost"),
  k = 1,
  id = "player_id"
)

data_berserk <- prepare_data(
  form = berserk ~ 0 | white + rating + rating_diff + min_rem + streak + berserk.1 + lost.1 + 1,
  re = c("rating_diff","lost.1"),
  choice_data = choice_berserk,
  id = "player_id",
  idc = "game_id",
  standardize = c("rating","rating_diff","min_rem"),
  impute = "zero"
)
model_berserk <- fit_model(
  data_berserk,
  latent_classes = list("dp_update" = TRUE, "C" = 10),
  R = 5000
)
}
}
\keyword{model}
