% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_chessop}
\alias{model_chessop}
\title{Latent class mixed probit model for chess opening choice}
\format{
An object of class \code{RprobitB_fit}.
}
\usage{
data(model_chessop)
}
\description{
This object is a fitted mixed probit model to the \code{choice_chess_opening}
dataset of the {RprobitB} package with the formula
\code{w1 ~ w1.1 | sex_w + byear_w + rating_w + rating_diff + 1}.
}
\details{
The model was estimated via
\preformatted{
choice_chess_opening <- create_lagged_cov(
  choice_data = RprobitB::choice_chess_opening,
  column = "w1",
  k = 1,
  id = "fideid_w")
choice_chess_opening <- fastDummies::dummy_cols(choice_chess_opening, "w1.1")
choice_chess_opening$rating_diff <- choice_chess_opening$rating_w - choice_chess_opening$rating_b
data <- prepare_data(
  form = w1 ~ w1.1 | sex_w + byear_w + rating_w + rating_diff + 1,
  choice_data = choice_chess_opening,
  id = "fideid_w",
  idc = "fideid_b",
  alternatives = c("e4","d4","b3"),
  standardize = c("byear_w","rating_w","rating_diff"),
  impute = "complete_cases"
)
model_chessop <- fit_model(
  data = data,
  re = "w1.1",
  latent_classes = list("dp_update" = TRUE, "C" = 2)
)
}
}
\keyword{model}
