% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Compare fitted models}
\usage{
model_selection(
  ...,
  criteria = c("npar", "LL", "AIC", "BIC"),
  add_form = FALSE
)
}
\arguments{
\item{...}{One or more objects of class \code{RprobitB_fit}.}

\item{criteria}{A vector of one or more of the following characters:
\itemize{
\item \code{"npar"} for the number of model parameters (see \code{\link{npar}}),
\item \code{"LL"} for the log-likelihood value (see \code{\link{logLik}}),
\item \code{"AIC"} for the AIC value (see \code{\link{AIC}}),
\item \code{"BIC"} for the BIC value (see \code{\link{BIC}}),
\item \code{"WAIC"} for the WAIC value (also shows its standard error
\code{sd(WAIC)} and the number \code{pWAIC} of effective model parameters,
see \code{\link{WAIC}}),
\item \code{"MMLL"} for the marginal model log-likelihood,
\item \code{"BF"} for the Bayes factor,
\item \code{"pred_acc"} for the prediction accuracy (see \code{\link{pred_acc}}).
}}

\item{add_form}{Set to \code{TRUE} to add the model formulas.}
}
\value{
A data frame, models in rows, criteria in columns.
}
\description{
This function returns a table with several criteria for model comparison.
}
\details{
See the vignette on model selection for more details.
}
\examples{
data("model_train", package = "RprobitB")
data("model_train_sparse", package = "RprobitB")
criteria <- c("npar", "LL", "AIC", "BIC", "WAIC", "MMLL", "BF", "pred_acc")
model_selection(model_train, model_train_sparse, criteria = criteria)
}
