% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_train}
\alias{model_train}
\title{Probit model for binary choice between Train trip alternatives}
\format{
An object of class \code{RprobitB_fit}.
}
\usage{
data(model_train)
}
\description{
This object is a fitted probit model to the Train data set of the {mlogit}
package with the model formula \code{choice ~ price + time + change + comfort | 0}.
}
\details{
The model was estimated via
\preformatted{
data("Train", package = "mlogit")
Train$price_A <- Train$price_A / 100 * 2.20371
Train$price_B <- Train$price_B / 100 * 2.20371
Train$time_A <- Train$time_A / 60
Train$time_B <- Train$time_B / 60
form <- choice ~ price + time + change + comfort | 0
data <- prepare_data(
  form = form,
  choice_data = Train,
  idc = "choiceid"
)
model_train <- fit_model(
  data = data,
  R = 10000,
  Q = 10,
  scale = price ~ -1)
)
}
}
\keyword{model}
