% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_train_sparse}
\alias{model_train_sparse}
\title{Probit model for binary choice between Train trip alternatives with the price
as the only explanatory variable}
\format{
An object of class \code{RprobitB_fit}.
}
\usage{
data(model_train_sparse)
}
\description{
This object is a fitted probit model to the Train data set of the {mlogit}
package with the model formula \code{choice ~ price | 0}.
}
\details{
The model was derived from the \code{\link{model_train}} object via
\preformatted{
data("model_train", package = "RprobitB")
nested_model(model_train, form = choice ~ price | 0)
}
}
\keyword{model}
