% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvn_draw}
\alias{mvn_draw}
\title{Sample from multivariate normal distribution.}
\usage{
mvn_draw(mu, Sigma)
}
\arguments{
\item{mu}{A numeric vector of length \code{n}.}

\item{Sigma}{A covariance matrix of dimension \code{n} x \code{n}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
This function returns a sample from a multivariate normal distribution
specified by its mean vector \code{mu} and its covariance matrix \code{Sigma}.
}
\examples{
mu <- c(1,0)
Sigma <- matrix(c(1,0.5,0.5,1),2,2)
RprobitB:::mvn_draw(mu = mu, Sigma = Sigma)

}
\keyword{utils}
