% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{nested_model}
\alias{nested_model}
\title{Estimating a nested model}
\usage{
nested_model(
  x,
  form,
  re,
  alternatives,
  id,
  idc,
  standardize,
  missing_data,
  scale,
  R,
  B,
  Q,
  print_progress,
  prior,
  latent_classes,
  seed
)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative due to identifiability).
They can be removed by adding \code{+0} in the second spot.
See the vignette on choice data for more details.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{id}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each decision maker. The default is \code{"id"}.}

\item{idc}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each choice situation of each decision maker.
The default is \code{NULL}, in which case these identifier are generated
automatically.}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{<covariate>_<alternative>}.
If \code{standardize = "all"}, all covariates get standardized.}

\item{missing_data}{Specifies how to handle missing entries (\code{NA, NaN, -Inf, Inf}) in
\code{choice_data}. The following options are available:
\itemize{
\item \code{"complete_cases"}, removes rows containing missing entries,
\item \code{"zero_out"}, replaces missing entries by zero,
\item \code{"mean"}, imputes missing entries by the covariate mean.
}}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}

\item{R}{The number of iterations of the Gibbs sampler.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{print_progress}{A boolean, determining whether to print the Gibbs sampler progress and the
estimated remaining computation time.}

\item{prior}{A named list of parameters for the prior distributions. See the documentation
of \code{\link{check_prior}} for details about which parameters can be specified.}

\item{latent_classes}{Either \code{NULL} (for no latent classes) or a list of parameters specifying
the number of latent classes and their updating scheme:
\itemize{
\item \code{C}: The fixed number (greater or equal 1) of latent classes,
which is set to 1 per default. If either \code{weight_update = TRUE}
or \code{dp_update = TRUE} (i.e. if classes are updated), \code{C}
equals the initial number of latent classes.
\item \code{weight_update}: A boolean, set to \code{TRUE} to weight-based
update the latent classes. See ... for details.
\item \code{dp_update}: A boolean, set to \code{TRUE} to update the latent
classes based on a Dirichlet process. See ... for details.
\item \code{Cmax}: The maximum number of latent classes.
\item \code{buffer}: The number of iterations to wait before a next
weight-based update of the latent classes.
\item \code{epsmin}: The threshold weight (between 0 and 1) for removing
a latent class in the weight-based updating scheme.
\item \code{epsmax}: The threshold weight (between 0 and 1) for splitting
a latent class in the weight-based updating scheme.
\item \code{distmin}: The (non-negative) threshold difference in class means
for joining two latent classes in the weight-based updating scheme.
}}

\item{seed}{Set a seed for the Gibbs sampling.}
}
\value{
An object of class \code{RprobitB_fit}.
}
\description{
This function is a wrapper for \code{\link{prepare_data}} and \code{\link{mcmc}}
to estimate a nested probit model based on a given \code{RprobitB_fit} object.
}
\details{
All parameters (except for \code{x})  are optional and if not specified
retrieved from the specification for \code{x}.
}
