% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_alternatives.R
\name{new_RprobitB_alternatives}
\alias{new_RprobitB_alternatives}
\title{Constructor for \code{RprobitB_alternatives} object}
\usage{
new_RprobitB_alternatives(
  alternatives = character(),
  base = tail(alternatives, 1),
  ordered = FALSE
)
}
\arguments{
\item{alternatives}{A character vector: the names of the choice alternatives. Must be at least of
length two.}

\item{base}{A character: the name of the base alternative for covariates that are not
alternative specific.
Ignored if the model has no alternative specific covariates (e.g., in the
ordered probit case).
Per default, \code{base} is the last element of \code{alternatives}.
The details of the base alternative are given under \sQuote{Details}.}

\item{ordered}{A boolean: \code{TRUE} for if the alternatives are ordered and \code{FALSE}
(default) else.}
}
\value{
An object of class \code{RprobitB_alternatives}.
}
\description{
This function constructs an S3 object of class \code{RprobitB_alternatives}.
}
\section{Details of the base alternative}{
The full collection of coefficients for covariates that are constant across
alternatives (including ASCs) is not identified, one coefficient is a linear
coefficient of the others, respectively. To achieve identifiability, the
coefficient of one alternative \code{base} is fixed to zero. The other
coefficients then have to be interpreted with respect to \code{base}.
}

\examples{
new_RprobitB_alternatives(LETTERS[1:3])

}
\keyword{internal}
