% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_formula.R
\name{new_RprobitB_formula}
\alias{new_RprobitB_formula}
\title{Constructor for \code{RprobitB_formula} object}
\usage{
new_RprobitB_formula(
  formula = choice ~ covariate,
  re = character(),
  ordered = FALSE
)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}:
a symbolic description of the model to be fitted.
The details of model specification are given under \sQuote{Details}.}

\item{re}{A character (vector): the names of covariates in \code{formula} that should
have a (log-) normal mixing distribution.
The details of random effects are given under \sQuote{Details}.}

\item{ordered}{A boolean, \code{FALSE} per default. If \code{TRUE}, the choice set
\code{alternatives} is assumed to be ordered from worst to best.}
}
\value{
An object of class \code{RprobitB_formula}.
}
\description{
This function constructs an S3 object of class \code{RprobitB_formula}.
}
\section{Details of model specification}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of alternative and choice situation specific
covariates with a coefficient that is constant across alternatives,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}

Multiple covariates of one type are separated by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot.
}

\section{Model specification in the ordered case}{
In the ordered probit model (\code{ordered = TRUE}), covariates cannot be
alternative specific. Therefore, the \code{formula} object does need any
separation via \code{|} anymore, and has the simple structure
\code{choice ~ A + B + C}. ASCs cannot be estimated in the ordered case.
}

\section{Details of random effects}{
TODO
}

\examples{
formula <- choice ~ A | B + 0 | C + D
re <- c("A", "D+")
new_RprobitB_formula(formula = formula, re = re, ordered = FALSE)
}
