% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{nobs}
\alias{nobs}
\alias{nobs.RprobitB_fit}
\title{Number of observations}
\usage{
nobs(object, ...)

\method{nobs}{RprobitB_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Ignored.}
}
\value{
An integer.
}
\description{
This function extracts the number of observations from an \code{RprobitB_fit}
object.
}
\examples{
data("model_train", package = "RprobitB")
nobs(model_train)

}
