% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{npar}
\alias{npar}
\title{Number of model parameters}
\usage{
npar(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Optionally more objects of class \code{RprobitB_fit}.}
}
\value{
Either a numeric value (if just one object is provided) or a numeric vector.
}
\description{
This function extracts the number of model parameters of an \code{RprobitB_fit}
object.
}
\examples{
data("model_train", package = "RprobitB")
npar(model_train)

}
