% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_effects.R
\name{number_covariates}
\alias{number_covariates}
\title{Compute number of covariates}
\usage{
number_covariates(formula, J, ordered = FALSE)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the model to be
fitted, see details.}

\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least 2.
If \code{ordered = TRUE}, must be at least 3.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}
}
\value{
An \code{integer}, the number of covariates per decider.
}
\description{
This helper function computes the number of covariates per decider.
}
\examples{
number_covariates(formula <- choice ~ cost | income | time, J = 3)

}
