% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_effects.R
\name{P}
\alias{P}
\alias{P_f}
\alias{P_r}
\title{Compute number of (fixed and random) model effects}
\usage{
P(formula, re, ordered = FALSE)

P_f(formula, re, ordered = FALSE)

P_r(formula, re, ordered = FALSE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}:
a symbolic description of the model to be fitted.
The details of model specification are given under \sQuote{Details}.}

\item{re}{A character (vector): the names of covariates in \code{formula} that should
have a random effect.
Set \code{re = NULL} (default) for no random effects.
The details of random effects are given under \sQuote{Details}.}

\item{ordered}{A boolean: \code{TRUE} for if the alternatives are ordered and \code{FALSE}
(default) else.}
}
\value{
An integer
}
\description{
\code{P()} computes the total number \code{P} of model effects.

\code{P_f()} computes the number \code{P_f} of fixed model effects.

\code{P_r()} computes the number \code{P_r} of random model effects.
}
\section{Details of model specification}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need any
separation via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Details of random effects}{
Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (for sign-restriction) can be specified via appending a \code{+}
to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
formula <- choice ~ A | B + 0 | C + D
re <- c("A", "D+")
P(formula, re)
P_f(formula, re)
P_r(formula, re)

}
