% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutations}
\alias{permutations}
\title{Permutations of a vector}
\usage{
permutations(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A list of all permutations of \code{x}.
}
\description{
This function returns all permutations of a given vector.
}
\examples{
RprobitB:::permutations(x = c("a","b","c"))

}
\references{
This function is a modified version of
\url{https://stackoverflow.com/a/20199902/15157768}.
}
\keyword{internal}
\keyword{utils}
