% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Autocorrelation plots}
\usage{
plot_acf(gibbs_samples, model, mcmc, out)
}
\arguments{
\item{gibbs_samples}{A list of Gibbs samples.}

\item{model}{A list of model information.}

\item{mcmc}{A list of Markov chain Monte Carlo parameters.}

\item{out}{A list of output settings.}
}
\value{
No return value. Creates pdf-file "acf.pdf" in folder "\code{out[["rdir"]]/out[["id"]]}".
}
\description{
Function that plots the autocorrelation of the Gibbs samples.
}
\details{
Function computes the \href{https://mc-stan.org/docs/2_18/reference-manual/effective-sample-size-section.html}{effective sample size}.
}
