% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_class_seq}
\alias{plot_class_seq}
\title{Visualizing the number of classes during Gibbs sampling}
\usage{
plot_class_seq(class_sequence, B)
}
\arguments{
\item{class_sequence}{The sequence of class numbers during Gibbs sampling of length \code{R}.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots the number of latent Glasses during Gibbs sampling
to visualize the class updating.
}
\keyword{internal}
