% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contour.R
\name{plot_contour}
\alias{plot_contour}
\title{Contour plots}
\usage{
plot_contour(x)
}
\arguments{
\item{x}{...}
}
\value{
...
}
\description{
Function that creates contour plot and progress contour plots of the Gibbs
samples.
}
