% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Visualizing the effects.}
\usage{
plot_effects(x, par_names, cov_names = NULL)
}
\arguments{
\item{x}{An object of class \code{RprobitB_gibbs_samples}.}

\item{par_names}{...}

\item{cov_names}{A character vector of covariate names.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function visualizes the effects of the covariates on the choices (i.e.
the linear coefficients) along with an uncertainty interval of plus and minus
one standard deviation.
}
