% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Visualizing the effects.}
\usage{
plot_estimates(x, par_names)
}
\arguments{
\item{x}{An object of class \code{RprobitB_gibbs_samples}.}

\item{par_names}{...}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function visualizes the effects of the covariates on the choices (i.e.
the linear coefficients) along with an uncertainty interval of plus and minus
one standard deviation.
}
