% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marginal.R
\name{plot_marginal}
\alias{plot_marginal}
\title{Marginal mixing distributions}
\usage{
plot_marginal(x, gibbs_samples, model, estimates, parm, out)
}
\arguments{
\item{x}{...}

\item{gibbs_samples}{A list of Gibbs samples.}

\item{model}{A list of model information.}

\item{estimates}{A list of model estimates.}

\item{parm}{A list of true parameter values.}

\item{out}{A list of output settings.}
}
\value{
No return value. Creates pdf-file "marginal.pdf" in folder "\code{out[["rdir"]]/out[["id"]]}".
}
\description{
Function that plots the estimated marginal mixing distributions.
}
\details{
Adds true mixing distribution if available.
}
