% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mixture_marginal}
\alias{plot_mixture_marginal}
\title{Plot marginal mixing distributions}
\usage{
plot_mixture_marginal(mean, cov, weights, name)
}
\arguments{
\item{mean}{}

\item{cov}{}

\item{weights}{}

\item{name}{}
}
\value{
GGplot TODO
}
\description{
This function plots an estimated marginal mixing distributions.
}
\examples{
mean <- list(1,2)
cov <- list(0.1,1)
weights <- c(0.3,0.7)
name <- "test"
RprobitB:::plot_mixture_marginal(mean, cov, weights, name)

}
\keyword{internal}
