% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve}
\usage{
plot_roc(..., reference = NULL)
}
\arguments{
\item{...}{One or more \code{RprobitB_fit} objects or \code{data.frame}s of choice
probability.}

\item{reference}{The reference alternative.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function draws receiver operating characteristic (ROC) curves.
}
\examples{
plot_roc(RprobitB::model_train)

}
