% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Visualizing the trace of Gibbs samples.}
\usage{
plot_trace(gibbs_samples, names, R, B, Q)
}
\arguments{
\item{gibbs_samples}{A matrix of Gibbs samples.}

\item{names}{The parameter names.}

\item{R}{The number of iterations of the Gibbs sampler.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}
}
\value{
...
}
\description{
This function plots traces of Gibbs samples.
}
