% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{pred_acc}
\alias{pred_acc}
\title{Prediction accuracy}
\usage{
pred_acc(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}
}
\value{
A numeric.
}
\description{
This function computes the prediction accuracy of an \code{RprobitB_fit}
object. Prediction accuracy means the share of choices that are correctly
predicted by the model, where prediction is based on the maximum choice
probability.
}
\examples{
data("model_train", package = "RprobitB")
pred_acc(model_train)
}
