% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Title}
\usage{
predict()
}
\value{
value
}
\description{
Funtion that...
}
\details{
For more details see the vignette "Prediction":
\code{vignette("prediction", package = "RprobitB")}
}
