% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_choices.R
\name{predict_choices}
\alias{predict_choices}
\title{Predict choices.}
\usage{
predict_choices(x, data = NULL, overview = TRUE)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{data}{Either \code{NULL} or an object of class \code{RprobitB_data}.}

\item{overview}{If \code{TRUE}, aggregate the prediction in a table.}
}
\value{
Either a table if \code{overview = TRUE} or a data frame otherwise.
}
\description{
This function predicts the choices of decision makers.
}
\examples{
data <- simulate_choices(form = choice ~ covariate, N = 10, T = 10, J = 2)
x <- mcmc(data)
predict_choices(x, overview = TRUE)
predict_choices(x, overview = FALSE)

}
