% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_classification.R
\name{preference_classification}
\alias{preference_classification}
\title{Classify preferences of deciders.}
\usage{
preference_classification(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}
}
\value{
A data frame with the deciders id and the latent class number.
}
\description{
This function classifies the preferences of deciders based on the estimated
latent classes.
}
