% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_flip.R
\name{preference_flip}
\alias{preference_flip}
\title{Check for flip in preferences after change in model scale.}
\usage{
preference_flip(model_old, model_new)
}
\arguments{
\item{model_old, model_new}{[\code{RprobitB_fit}]\cr
Two models, one before and one after a scale change.}
}
\value{
No return value, called for side-effects.
}
\description{
This function checks if a change in the model scale flipped the preferences.
}
\keyword{internal}
