% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare empirical choice data.}
\usage{
prepare(
  form,
  choice_data,
  alternatives = NULL,
  re = NULL,
  id = "id",
  standardize = NULL
)
}
\arguments{
\item{choice_data}{A data frame of choice data in "wide" format (i.e. each row represents
one choice occasion) with the following requirements:
\itemize{
\item It must contain columns named \code{id} (a unique identifier for each
decision maker) and \code{choice} (the chosen alternatives), where
\code{choice} is the name of the dependent variable in \code{form}.
\item For each alternative specific covariate \emph{p} (covariate of type 1 or
3) in \code{form} and each choice alternative \emph{j} in
\code{alternatives}, \code{choice_data} must contain a column named
\emph{p_j}.
\item For each covariate \emph{q} in \code{form} that is constant across
covariates (covariate of type 2), \code{choice_data} must contain a
column named \emph{q}.
}}

\item{alternatives}{A character vector with the names of the choice alternatives.}

\item{id}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each decision maker. The default is \code{"id"}.}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{covariate_alternative}.
If \code{standardize = "all"}, all covariates get standardized.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function prepares empirical choice data for the RprobitB package.
}
\details{
For more details see the vignette "Data management":
\code{vignette("data_management", package = "RprobitB")}.
}
\examples{
data("Train", package = "mlogit")
data = prepare(form = choice ~ price | 0 | time + comfort + change,
               choice_data = Train,
               re = c("price","time"),
               standardize = "all")
}
