% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.RprobitB_data}
\alias{print.RprobitB_data}
\title{Print method for \code{RprobitB_data}}
\usage{
\method{print}{RprobitB_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_data}}

\item{...}{ignored}
}
\description{
Print method for \code{RprobitB_data}
}
