% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.RprobitB_fit}
\alias{print.RprobitB_fit}
\title{Print a fitted choice model}
\usage{
\method{print}{RprobitB_fit}(x, ...)
}
\arguments{
\item{x}{[\code{RprobitB_fit}]\cr
Fitted choice model.}

\item{...}{Currently not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints the model formula, data size, and retained posterior sample size.
}
\examples{
model <- fit(choice ~ x, chains = 1)
print(model)
}
\keyword{models}
