% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RprobitB_gibbs_samples_statistics.R
\name{print.RprobitB_gibbs_samples_statistics}
\alias{print.RprobitB_gibbs_samples_statistics}
\title{Print method for \code{RprobitB_gibbs_samples_statistics}.}
\usage{
\method{print}{RprobitB_gibbs_samples_statistics}(x, true = NULL, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_gibbs_samples_statistics}.}

\item{true}{Either \code{NULL} or an object of class \code{RprobitB_true_parameter}.}

\item{digits}{The number of printed decimal places.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class
\code{RprobitB_gibbs_samples_statistics}.
}
