% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RprobitB_latent_classes.R
\name{print.RprobitB_latent_classes}
\alias{print.RprobitB_latent_classes}
\title{Print method for \code{RprobitB_latent_classes}.}
\usage{
\method{print}{RprobitB_latent_classes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_latent_classes}.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class
\code{RprobitB_latent_classes}.
}
