% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RprobitB_model.R
\name{print.RprobitB_model}
\alias{print.RprobitB_model}
\title{Print method for \code{RprobitB_model}.}
\usage{
\method{print}{RprobitB_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class \code{RprobitB_model}.
}
