% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.RprobitB_data.R
\name{print.summary.RprobitB_data}
\alias{print.summary.RprobitB_data}
\title{Print method for the summary of \code{RprobitB_data}}
\usage{
\method{print}{summary.RprobitB_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.RprobitB_data}}

\item{...}{ignored}
}
\description{
Print method for the summary of \code{RprobitB_data}
}
