% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{print.summary.RprobitB_fit}
\alias{print.summary.RprobitB_fit}
\title{Print a fitted model summary}
\usage{
\method{print}{summary.RprobitB_fit}(x, digits = 3L, ...)
}
\arguments{
\item{x}{[\code{summary.RprobitB_fit}]\cr
Model summary returned by \code{summary()}.}

\item{digits}{[\code{integer(1)}]\cr
Number of significant digits to print.}

\item{...}{Further arguments passed to \code{print.data.frame()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints sampling information followed by the posterior summary table.
}
\keyword{models}
