% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.RprobitB_model.R
\name{print.summary.RprobitB_model}
\alias{print.summary.RprobitB_model}
\title{Print method for the summary of \code{RprobitB_model}.}
\usage{
\method{print}{summary.RprobitB_model}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{summary.RprobitB_model}.}

\item{digits}{The number of printed decimal places.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class
\code{summary.RprobitB_model}.
}
