% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_settings.R
\name{print_settings}
\alias{print_settings}
\title{Model setting}
\usage{
print_settings(model, lcus, init, mcmc, norm, out)
}
\arguments{
\item{model}{A list of model information.}

\item{lcus}{A list of latent class updating scheme parameters.}

\item{init}{A list of initial values.}

\item{mcmc}{A list of Markov chain Monte Carlo parameters.}

\item{norm}{A list of normalization information.}

\item{out}{A list of output settings.}
}
\value{
No return value.
}
\description{
Function that prints the model settings.
}
