% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hat.R
\name{R_hat}
\alias{R_hat}
\title{Compute Gelman-Rubin statistic.}
\usage{
R_hat(samples, parts = 2)
}
\arguments{
\item{samples}{Normalized and burned but not thinned Gibbs samples.}

\item{parts}{The number of parts to divide the samples into sub-chains.}
}
\value{
The Gelman-Rubin statistic.
}
\description{
This function computes the Gelman-Rubin statistic \code{R_hat}.
}
\references{
\url{https://bookdown.org/rdpeng/advstatcomp/monitoring-convergence.html}
}
