% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Routine to draw from Dirichlet}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{A vector, the concentration parameter}
}
\value{
A vector, the sample from the Dirichlet distribution
}
\description{
Routine to draw from Dirichlet
}
