% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_cpp}
\alias{rdirichlet_cpp}
\title{Draw from Dirichlet distribution}
\usage{
rdirichlet_cpp(concentration)
}
\arguments{
\item{concentration}{A \code{numeric}, the concentration vector of length \code{n}.}
}
\value{
A \code{numeric} of length \code{n}, the random draw.
}
\description{
Function to draw from a Dirichlet distribution.
}
\details{
This function performs no input checks. See \code{\link{rdirichlet}}
for the version with input checks.
}
\examples{
RprobitB:::rdirichlet(concentration = 1:3)

}
\keyword{cpp}
\keyword{internal}
