% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{train_test}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{choicedata}{\code{\link[choicedata:train_test]{train_test()}}}
}}

