% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reference_normalization}
\alias{reference_normalization}
\title{Normalize independent variables with respect to level}
\usage{
reference_normalization(independent_variable, J, reference_level)
}
\arguments{
\item{independent_variable}{A \code{list} that contains the independent variables.
It is of length \code{N}, where \code{N} is the number of individuals.
The \code{n}-th element of \code{independent_variable} is a \code{list} of
length \code{Tp[n]}, where the \code{t}-th element is the design
\code{matrix} for time point \code{t}.}

\item{J}{An \code{integer}, the number of levels of the dependent variable.}

\item{reference_level}{An \code{integer} between \code{1} and \code{J} that determines the
reference level for level normalization. It is not relevant and ignored in
the ordered case.}
}
\value{
The input \code{independent_variable}, where for each design \code{matrix}
differences with respect to the level \code{reference_level} are taken.
}
\description{
Normalize independent variables with respect to level
}
