% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Draw from multivariate normal distribution}
\usage{
rmvnorm(mu, Sigma)
}
\arguments{
\item{mu}{The mean vector of length \code{n}.}

\item{Sigma}{The covariance matrix of dimension \code{n} x \code{n}.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
This function draws from a multivariate normal distribution.
}
\keyword{dist}
