% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm_cpp}
\alias{rmvnorm_cpp}
\title{Draw from multivariate normal distribution}
\usage{
rmvnorm_cpp(mean, Sigma, log = FALSE)
}
\arguments{
\item{mean}{A \code{numeric}, the mean vector of length \code{n}.}

\item{Sigma}{A \code{matrix}, the covariance matrix of dimension \code{n} x \code{n}.}

\item{log}{A \code{logical}, if \code{TRUE} the draw is taken from the log-normal
distribution.
By default, \code{log = FALSE}.}
}
\value{
A \code{numeric} of length \code{n}, the random draw.
}
\description{
This function draws from a multivariate normal distribution.
}
\details{
The function builds upon the following fact:
If \eqn{\epsilon = (\epsilon_1,\dots,\epsilon_n)},
where each \eqn{\epsilon_i} is drawn independently from a standard normal
distribution, then \eqn{\mu+L\epsilon} is a draw from the multivariate
normal distribution \eqn{N(\mu,\Sigma)}, where \eqn{L} is the lower
triangular factor of the Choleski decomposition of \eqn{\Sigma}.

This function performs no input checks. See \code{\link{rmvnorm}}
for the version with input checks.
}
\examples{
RprobitB:::rmvnorm_cpp(mean = c(0,0), Sigma = diag(2))
RprobitB:::rmvnorm_cpp(mean = c(0,0), Sigma = diag(2), log = TRUE)

}
\keyword{cpp}
\keyword{internal}
