% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{roc}
\alias{roc}
\title{Plot ROC curve}
\usage{
roc(..., reference = NULL, auc = TRUE)
}
\arguments{
\item{...}{One or more \code{RprobitB_fit} objects or \code{data.frame}s.}

\item{reference}{}

\item{auc}{}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function
}
\examples{
roc(RprobitB::model_train)

}
