% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_package.R
\docType{package}
\name{RprobitB}
\alias{RprobitB}
\alias{RprobitB-package}
\title{RprobitB: Bayes Estimation of Probit Models for Discrete Choice Analysis}
\description{
This package provides tools for Bayes estimation of probit models for
discrete choice analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/RprobitB/}
  \item Report bugs at \url{https://github.com/loelschlaeger/RprobitB/issues}
}

}
\author{
Lennart Oelschläger, \email{oelschlaeger.lennart@gmail.com}
}
\keyword{internal}
