% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_alternatives.R
\name{RprobitB_alternatives}
\alias{RprobitB_alternatives}
\alias{is.RprobitB_alternatives}
\alias{validate_RprobitB_alternatives}
\alias{print.RprobitB_alternatives}
\title{Define choice alternatives}
\usage{
RprobitB_alternatives(alternatives, base = alternatives[1], ordered = FALSE)

is.RprobitB_alternatives(x)

validate_RprobitB_alternatives(x)

\method{print}{RprobitB_alternatives}(x, ...)
}
\arguments{
\item{alternatives}{A \code{character}, the vector of names of the choice alternatives.
Must be at least of length two.}

\item{base}{A \code{character}, the name of the base alternative for covariates that are
not alternative specific, see details.
Ignored if the model has no alternative specific covariates (e.g., in the
ordered probit case).
By default, \code{base} is the first element of \code{alternatives}.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{An \code{RprobitB_alternatives} object.}
}
\value{
An \code{RprobitB_alternatives} object.
}
\description{
This function constructs an object of class \code{RprobitB_alternatives},
which contains the choice alternatives.
}
\section{Base alternative}{
The full collection of coefficients for covariates that are constant across
alternatives (including ASCs) is not identified, one coefficient is a linear
coefficient of the others, respectively. To achieve identifiability, the
coefficient of one alternative \code{base} is fixed to zero. The other
coefficients then have to be interpreted with respect to \code{base}.
}

\examples{
RprobitB_alternatives(LETTERS[1:3])

}
\keyword{internal}
