% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_demo.R
\name{RprobitB_demo}
\alias{RprobitB_demo}
\title{A demonstration of RprobitB}
\usage{
RprobitB_demo()
}
\value{
An object of class \code{RprobitB_fit}.
}
\description{
This function demonstrates the functionality of RprobitB.
}
