% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_formula.R
\name{RprobitB_formula}
\alias{RprobitB_formula}
\alias{is.RprobitB_formula}
\alias{validate_RprobitB_formula}
\alias{print.RprobitB_formula}
\title{Define probit model formula}
\usage{
RprobitB_formula(formula, re = NULL, ordered = FALSE)

is.RprobitB_formula(x)

validate_RprobitB_formula(x)

\method{print}{RprobitB_formula}(x, ...)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the model to be
fitted, see details.}

\item{re}{A \code{character}, the vector of names of covariates in \code{formula} that
should have a random effect, see details.
By default, \code{re = NULL}, i.e., no random effects.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{An \code{RprobitB_formula} object.}
}
\value{
An \code{RprobitB_formula} object.
}
\description{
This function constructs an object of class \code{RprobitB_formula}, which
contains the formula for a probit model.
}
\section{Model formula}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need any
separation via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{
Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (for sign-restriction) can be specified via appending \code{+}
to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
formula <- choice ~ A | B + 0 | C + D
re <- c("A", "D+")
RprobitB_formula(formula, re, ordered = FALSE)

}
\keyword{internal}
\keyword{object}
