% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_parameter.R
\name{RprobitB_parameter}
\alias{RprobitB_parameter}
\alias{print.RprobitB_parameter}
\title{Define probit model parameters}
\usage{
RprobitB_parameter(
  C = 1,
  s = rep(1, C)/C,
  alpha = NA,
  b = NA,
  Omega = NA,
  Sigma = NA,
  Sigma_diff = NA,
  diff_alt = 1,
  beta = NA,
  z = NA,
  d = NA
)

\method{print}{RprobitB_parameter}(x, ..., digits = 4)
}
\arguments{
\item{C}{The number (greater or equal 1) of latent classes of decision makers.
Per default, \code{C = 1}.}

\item{s}{The vector of class weights of length \code{C}.
For identifiability, the vector must be descending.
Per default, \code{s = rep(1,C)/C}.}

\item{alpha}{The matrix of fixed coefficients of dimension \code{P_f} x \code{C}.
The coefficients for class \code{c} are stored in column \code{c}.}

\item{b}{The matrix of class means of dimension \code{P_r} x \code{C}.
The mean vector for class \code{c} is stored in column \code{c}.}

\item{Omega}{The matrix of class covariance matrices of dimension \code{P_r^2} x \code{C}.
The covariance matrix for class \code{c} is stored as a vector in column
\code{c}.}

\item{Sigma}{The error term covariance matrix of dimension \code{J} x \code{J}.
In the ordered probit model (see details), \code{Sigma} is a single numeric.}

\item{Sigma_diff}{The differenced error term covariance matrix of dimension
\code{J-1} x \code{J-1}.
\code{Sigma_diff} is assumed to be differenced with respect to alternative
\code{diff_alt}, see details.
\code{Sigma_diff} is ignored in case of the ordered probit model
(see details) or if \code{Sigma} is specified.}

\item{diff_alt}{An integer from \code{1} to \code{J}, the reference alternative for utility
differencing that maps \code{Sigma} to \code{Sigma_diff}, see details.
Per default, \code{diff_alt = 1}.}

\item{beta}{The matrix of the decider-specific coefficient vectors of dimension
\code{P_r} x \code{N}.
The coefficient vector for decider \code{n} is stored in column \code{n}.}

\item{z}{The vector of the allocation variables of length \code{N}.
Entry \code{n} of \code{z} is an integer from \code{1} to \code{C} and
denotes the allocated class for decider \code{n}.}

\item{d}{The numeric vector of logarithmic increases of the utility thresholds of
length \code{J-2}, only relevant in the ordered probit model case
(see details) and ignored otherwise.}

\item{...}{Names of parameters to be printed. If not specified, all available parameters
are printed.}

\item{digits}{The number of decimal places of the parameters to be printed.
Per default, \code{digits = 4}.}
}
\value{
An object of class \code{RprobitB_parameter}.
}
\description{
This function creates an object of class \code{RprobitB_parameter}, which
contains the parameters of a probit model, see details.
}
\section{The probit model}{
Assume that we know the choices of \eqn{N} deciders choosing between
\eqn{J \geq 2} alternatives at each of \eqn{T} choice occasions.
Specific to each decider, alternative and choice occasion, we observe \eqn{P}
covariates, a linear combination of which eventually explains the latent
random utility:
\deqn{U_{ntj} = X_{ntj}' \tilde{\beta}_n + \epsilon_{ntj},}
\eqn{n=1,\dots,N}, \eqn{t=1,\dots,T}, and \eqn{j=1,\dots,J}.
Here, \eqn{X_{ntj}} is a (column) vector of \eqn{P} characteristics specific
to alternative \eqn{j} as faced by decider \eqn{n} at choice occasion
\eqn{t}, \eqn{\tilde{\beta}_n \in \mathbb{R}^{P}} is the coefficient vector
encoding the preferences of \eqn{n}, and
\eqn{(\epsilon_{nt:}) = (\epsilon_{nt1},\dots,\epsilon_{ntJ})' \sim
\text{MVN}_{J} (0,\Sigma)} is the model's error term vector for \eqn{n} at
\eqn{t}.

The value \eqn{U_{ntj}} can be interpreted as the decider's utility.
It is unobserved by the researcher, but we assume that the deciders know
their utilities for each alternative and make a choice which is consistent
with utility maximization. Therefore,
\deqn{y_{nt} = \operatorname*{argmax}_{j = 1,\dots,J} U_{ntj},}
where \eqn{y_{nt}=j} denotes the event that decider \eqn{n} chooses \eqn{j}
at her \eqn{t}-th choice occasion.

Entries of the decider-specific coefficient vector \eqn{\tilde{\beta}_n} can
be fixed across deciders, in which case the coefficient vector is of the form
\eqn{\tilde{\beta}_n' = (\alpha', \beta_n')'}, where
\eqn{\alpha \in \mathbb{R}^{P_f}} are \eqn{P_f} coefficients that are
constant across deciders and \eqn{\beta_n} are \eqn{P_r} decider-specific
coefficients, \eqn{P_f + P_r = P}.

The decider-specific coefficients are assumed to be realizations of an
underlying mixing distribution and to be independent of the characteristics
\eqn{X_{ntj}} and the errors \eqn{(\epsilon_{nt:})}.
This distribution characterizes heterogeneity among the deciders and allows
for individual sensitivities. As mixing distribution, we assume a mixture of
\eqn{P_r}-variate Gaussian densities \eqn{\phi_{P_r}} with mean vectors
\eqn{b = (b_c)_{c}} and covariance matrices \eqn{\Omega = (\Omega_c)_{c}}
using \eqn{C} components:
\deqn{\beta_n\mid b,\Omega \sim \sum_{c=1}^{C} s_c \phi_{P_r} (\cdot \mid
b_c,\Omega_c).}
Here, \eqn{(s_c)_{c}} are weights satisfying \eqn{0 < s_c\leq 1} for
\eqn{c=1,\dots,C} and \eqn{\sum_c s_c=1}.

One interpretation of the latent class model is obtained by introducing
variables \eqn{z=(z_n)_n}, allocating each decision maker \eqn{n} to class
\eqn{c} with probability \eqn{s_c}, i.e.,
\deqn{\text{Prob}(z_n=c)=s_c \land \beta_n \mid z,b,\Omega \sim
\phi_{P_r}(\cdot \mid b_{z_n},\Omega_{z_n}).}
}

\section{Ordered probit model}{
When the set of choice alternatives is ordered, the probit model has only a
single utility
\deqn{U_{nt} = X_{nt}' \tilde{\beta}_n + \epsilon_{nt},}
\eqn{\epsilon_{nt} \sim \text{MVN}_{1} (0,\Sigma)},
per decider \eqn{n} and choice occasion \eqn{t}. The utility can be
interpreted as the level of association that \eqn{n} has with the choice
question. It falls into discrete categories, which in turn are linked to the
ordered alternatives \eqn{j=1,\dots,J}. Formally,
\deqn{y_{nt} = \sum_{j = 1,\dots,J} j \cdot I(\gamma_{j-1} < U_{nt} \leq
\gamma_{j}),}
where \eqn{\gamma_0 = -\infty} and \eqn{\gamma_J = +\infty}. This implies
that alternative \eqn{j} is chosen, if the utility falls into the interval
\eqn{(\gamma_{j-1}, \gamma_j]}.
Monotonicity of the thresholds \eqn{(\gamma_j)_{j=1,\dots,J-1}} is ensured
by estimating logarithmic increments \eqn{d_j} with
\eqn{\gamma_j = \sum_{i\leq j} \exp{(d_i)}}, \eqn{j=1,\dots,J-1}.
For level normalization, we fix \eqn{\gamma_1 = 0}.
}

\section{Level and scale normalization}{
The probit model is invariant towards the level and scale of utility, hence
a transformation is required for identifiability.

For level normalization, we take utility differences:
\deqn{\tilde{U}_{ntj} = \tilde{X}_{ntj}' \tilde{\beta}_n +
\tilde{\epsilon}_{ntj},}
where (choosing some alternative \eqn{k \in \{1,\dots,J\}} as the reference,
also denoted by \code{diff_alt})
\eqn{\tilde{U}_{ntj} = U_{ntj} - U_{ntk}},
\eqn{\tilde{X}_{ntj} = X_{ntj} - X_{ntk}}, and
\eqn{\tilde{\epsilon}_{ntj} = \epsilon_{ntj} - \epsilon_{ntk}} for
\eqn{j\neq k}.
The error term differences \eqn{(\tilde{\epsilon}_{nt:})} again are
multivariate normally distributed with mean \eqn{0} but transformed
covariance matrix \eqn{\tilde{\Sigma}}, also denoted by \code{Sigma_diff}.
See \code{\link{Sigma_diff}} for computing \code{Sigma_diff} from
\code{Sigma}.

For level normalization in the ordered probit model, we fix
\eqn{\gamma_1 = 0}.

For scale normalization, we fix the top left element of \code{Sigma_diff} to
\eqn{1}. Other options exist, see \code{\link{transform}}.
}

\examples{
RprobitB_parameter()

}
