% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior}
\alias{RprobitB_prior}
\title{Define model prior}
\usage{
RprobitB_prior(formula, re = NULL, J, C = 1, ordered = FALSE, ...)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}:
a symbolic description of the model to be fitted.
The details of model specification are given under \sQuote{Details}.}

\item{re}{A character (vector): the names of covariates in \code{formula} that should
have a random effect.
Set \code{re = NULL} (default) for no random effects.
The details of random effects are given under \sQuote{Details}.}

\item{J}{An integer, the number of choice alternatives.}

\item{C}{An integer, the number of latent classes.}

\item{ordered}{A boolean: \code{TRUE} for if the alternatives are ordered and \code{FALSE}
(default) else.}

\item{...}{
  Arguments passed on to \code{\link[=RprobitB_prior_alpha]{RprobitB_prior_alpha}}
  \describe{
    \item{\code{P_f}}{An integer, the number of fixed model effects.}
    \item{\code{alpha_prior_mean}}{A numeric vector of length \code{P_f}, the mean vector for the conjugate
normal prior distribution of \code{alpha}.}
    \item{\code{alpha_prior_Sigma}}{A covariance matrix of dimension \code{P_f} x \code{P_f}, the covariance
matrix for the conjugate normal prior distribution of \code{alpha}.}
    \item{\code{alpha_prior_custom}}{A function, a custom prior density function for \code{alpha}.
The details of specifying a custom prior are given under \sQuote{Details}.}
    \item{\code{alpha_prior_custom_test_par}}{A numeric vector of length \code{P_f}, a test input for the custom prior
density function for \code{alpha}.}
  }}
}
\value{
An object of class \code{RprobitB_prior}.
}
\description{
This function defines the prior distributions for a probit model.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item \code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item \code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item \code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item \code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item \code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

Per default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha}
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b}
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega}
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma}
\itemize{
\item \code{Sigma_prior_df = if (ordered) 4 else J + 1}
\item \code{Sigma_prior_scale = if (ordered) diag(1) else diag(J-1)}
}
\item \code{d}
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

Per default, conjugate priors are used, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior at any point, see \sQuote{Examples}.
}
}

\section{Details of model specification}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need any
separation via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Details of random effects}{
Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (for sign-restriction) can be specified via appending a \code{+}
to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
### default conjugate prior distributions
RprobitB_prior(formula = choice ~ A, J = 2)

### conjugate prior distributions with custom parameters
RprobitB_prior(
  formula = choice ~ A | B | C, re = "C", J = 3, C = 2,
  alpha_prior_mean = numeric(5),
  s_prior_concentration = c(1,2)
)

### custom prior distributions
RprobitB_prior(
  formula = choice ~ A | B | C, re = "C", J = 3,
  alpha_prior_custom = function(x) {
    dnorm(x[1], mean = 0, sd = 1) * dunif(x[2], min = 0, max = 1)
  }
)

}
\keyword{specification}
