% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_alpha}
\alias{RprobitB_prior_alpha}
\alias{print.RprobitB_prior_alpha}
\title{Prior for \code{alpha}}
\usage{
RprobitB_prior_alpha(
  P_f,
  alpha_prior_mean = numeric(P_f),
  alpha_prior_Sigma = 10 * diag(P_f),
  alpha_prior_custom = NULL,
  alpha_prior_custom_test_par = numeric(P_f)
)

\method{print}{RprobitB_prior_alpha}(x, ...)
}
\arguments{
\item{P_f}{An integer, the number of fixed model effects.}

\item{alpha_prior_mean}{A numeric vector of length \code{P_f}, the mean vector for the conjugate
normal prior distribution of \code{alpha}.}

\item{alpha_prior_Sigma}{A covariance matrix of dimension \code{P_f} x \code{P_f}, the covariance
matrix for the conjugate normal prior distribution of \code{alpha}.}

\item{alpha_prior_custom}{A function, a custom prior density function for \code{alpha}.
The details of specifying a custom prior are given under \sQuote{Details}.}

\item{alpha_prior_custom_test_par}{A numeric vector of length \code{P_f}, a test input for the custom prior
density function for \code{alpha}.}

\item{x}{An object of class \code{RprobitB_prior_alpha}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{RprobitB_prior_alpha}.
}
\description{
This function defines the prior distributions for the probit model parameter
\code{alpha}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item \code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item \code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item \code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item \code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item \code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

Per default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha}
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b}
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega}
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma}
\itemize{
\item \code{Sigma_prior_df = if (ordered) 4 else J + 1}
\item \code{Sigma_prior_scale = if (ordered) diag(1) else diag(J-1)}
}
\item \code{d}
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

Per default, conjugate priors are used, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior at any point, see \sQuote{Examples}.
}
}

\examples{
### conjugate prior: alpha ~ Normal(0, 10*I)
RprobitB_prior_alpha(
  P_f = 2,
  alpha_prior_mean = numeric(P_f),
  alpha_prior_Sigma = 10 * diag(P_f)
)

### custom prior: alpha_1 ~ Normal(0,1), alpha_2 ~ Uniform(0,1)
RprobitB_prior_alpha(
  P_f = 2,
  alpha_prior_custom = function(x) {
    dnorm(x[1], mean = 0, sd = 1) * dunif(x[2], min = 0, max = 1)
  }
)

}
\keyword{internal}
