% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_s}
\alias{RprobitB_prior_s}
\alias{is.RprobitB_prior_s}
\alias{print.RprobitB_prior_s}
\title{Define \code{s} prior}
\usage{
RprobitB_prior_s(
  C = 1,
  s_prior_concentration = rep(1, C),
  s_prior_custom = NA,
  s_prior_custom_test_par = numeric(C)
)

is.RprobitB_prior_s(x)

\method{print}{RprobitB_prior_s}(x, coldots = 4, digits = 2, simplify = TRUE, details = FALSE, ...)
}
\arguments{
\item{C}{An \code{integer}, the number (greater or equal 1) of latent classes of
decision makers.
By default, \code{C = 1}.}

\item{s_prior_concentration}{A \code{numeric} of length \code{C}, the concentration vector for the
conjugate Dirichlet prior distribution of \code{s}.
By default, \code{s_prior_concentration = rep(1, C)}.}

\item{s_prior_custom}{A \code{function}, a custom prior density function for \code{s},
see details.
By default, \code{s_prior_custom = NULL}, i.e., no custom prior.}

\item{s_prior_custom_test_par}{A \code{numeric} of length \code{C}, a test input for the custom prior
density function \code{s_prior_custom}.
By default, \code{s_prior_custom_test_par = numeric(C)}.
Ignored if \code{s_prior_custom = NULL}.}

\item{x}{An \code{RprobitB_prior_s} object.}

\item{coldots}{An \code{integer}, the column number which is replaced by \code{...}.
By default, \code{coldots = 4}.}

\item{digits}{An \code{integer}, the number of decimal places to be printed.
By default, \code{digits = 2}.}

\item{simplify}{A \code{logical}, set to \code{TRUE} to simplify the output.
By default, \code{simplify = FALSE}.}

\item{details}{A \code{logical}, set to \code{TRUE} to print the type and
dimension of \code{x}.
By default, \code{details = !simplify}.}

\item{...}{Not used.}
}
\value{
An \code{RprobitB_prior_s} object.
}
\description{
This function defines the prior distributions for the probit model parameter
\code{s}. Only relevant if \code{C > 1}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item Only if \code{P_f > 0}:
\code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item Only if \code{C > 1}:
\code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item Only if \code{P_r > 0}:
\code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item Only if \code{P_r > 0}:
\code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item Only in the unordered case (\code{ordered = FALSE}):
\code{Sigma_diff ~ Inverse-Wishart(df = Sigma_diff_prior_df, scale = Sigma_diff_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha} (only if \code{P_f > 0})
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s} (only if \code{C > 1})
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b} (only if \code{P_r > 0})
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega} (only if \code{P_r > 0})
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma} (only if \code{ordered = TRUE})
\itemize{
\item \code{Sigma_prior_df = 3}
\item \code{Sigma_prior_scale = 1}
}
\item \code{Sigma_diff} (only if \code{ordered = FALSE})
\itemize{
\item \code{Sigma_diff_prior_df = J + 1}
\item \code{Sigma_diff_prior_scale = diag(J-1)}
}
\item \code{d} (only if \code{ordered = TRUE})
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are applied, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in {RprobitB}) can be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\examples{
### conjugate prior: s ~ Dirichlet(rep(1, C))
RprobitB_prior_s(C = 2)

### custom prior: s_1 ~ Uniform(0.5,1), s_2 := 1 - s_1
RprobitB_prior_s(
  C = 2,
  s_prior_custom = function(x) {
    stats::dunif(x[1], min = 0.5, max = 1) * (x[2] == 1 - x[1])
  },
  s_prior_custom_test_par = c(0.6, 0.4)
)

}
\keyword{internal}
