% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_Sigma}
\alias{RprobitB_prior_Sigma}
\alias{is.RprobitB_prior_Sigma}
\alias{print.RprobitB_prior_Sigma}
\title{Define \code{Sigma} prior}
\usage{
RprobitB_prior_Sigma(
  ordered = FALSE,
  J,
  Sigma_prior_df = 3,
  Sigma_prior_scale = 1,
  Sigma_prior_custom = NA,
  Sigma_prior_custom_test_par = matrix(1)
)

is.RprobitB_prior_Sigma(x)

\method{print}{RprobitB_prior_Sigma}(
  x,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = TRUE,
  details = FALSE,
  ...
)
}
\arguments{
\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least \code{2}.
If \code{ordered = TRUE}, must be at least \code{3}.}

\item{Sigma_prior_df}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Sigma}.
It must be greater or equal than \code{3} (to ensure a proper prior).
By default, \code{Sigma_prior_df = 3}.}

\item{Sigma_prior_scale}{A \code{matrix} of dimension \code{1} x \code{1}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of
\code{Sigma}.
Can also be a single \code{numeric}.
By default, \code{Sigma_prior_scale = 1}.}

\item{Sigma_prior_custom}{A \code{function}, a custom prior density function for \code{Sigma},
see details.
By default, \code{Sigma_prior_custom = NULL}, i.e., no custom prior.}

\item{Sigma_prior_custom_test_par}{A \code{matrix} of dimension \code{1} x \code{1}, a test input for
the custom prior density function \code{Sigma_prior_custom}.
Can also be a single \code{numeric}.
By default, \code{Sigma_prior_custom_test_par = matrix(1)}.
Ignored if \code{Sigma_prior_custom = NULL}.}

\item{x}{An \code{\link{RprobitB_prior_Sigma}} object.}

\item{...}{Currently not used.}
}
\value{
An \code{\link{RprobitB_prior_Sigma}} object or \code{NA} if
\code{ordered = FALSE}.

It contains the elements:
\describe{
\item{\code{conjugate}}{Is the prior conjugated?}
\item{\code{Sigma_prior_df}}{The prior degrees of freedom.}
\item{\code{Sigma_prior_scale}}{The prior scale matrix.}
\item{\code{Sigma_prior_custom}}{The custom prior function.}
}
}
\description{
This function defines the prior distributions for the probit model parameter
\code{Sigma}. Only relevant if \code{ordered = TRUE}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item Only if \code{P_f > 0}:
\code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item Only if \code{P_r > 0}:
\code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item Only if \code{P_r > 0}:
\code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item Only in the unordered case (\code{ordered = FALSE}):
\code{Sigma_diff ~ Inverse-Wishart(df = Sigma_diff_prior_df, scale = Sigma_diff_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha} (only if \code{P_f > 0})
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b} (only if \code{P_r > 0})
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega} (only if \code{P_r > 0})
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma} (only if \code{ordered = TRUE})
\itemize{
\item \code{Sigma_prior_df = 3}
\item \code{Sigma_prior_scale = 1}
}
\item \code{Sigma_diff} (only if \code{ordered = FALSE})
\itemize{
\item \code{Sigma_diff_prior_df = J + 1}
\item \code{Sigma_diff_prior_scale = diag(J-1)}
}
\item \code{d} (only if \code{ordered = TRUE})
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are applied, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in \{RprobitB\}) can be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\examples{
\dontrun{
### conjugate prior: Sigma ~ Inverse-Wishart(3, 1)
RprobitB_prior_Sigma(ordered = TRUE, J = 4)

### custom prior: Sigma ~ Uniform(0,1)
RprobitB_prior_Sigma(
  ordered = TRUE,
  J = 4,
  Sigma_prior_custom = function(x) dunif(x)
)
}

}
\keyword{internal}
\keyword{object}
