% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_Sigma_diff}
\alias{RprobitB_prior_Sigma_diff}
\alias{is.RprobitB_prior_Sigma_diff}
\alias{print.RprobitB_prior_Sigma_diff}
\title{Define \code{Sigma_diff} prior}
\usage{
RprobitB_prior_Sigma_diff(
  ordered = FALSE,
  J,
  Sigma_diff_prior_df = J + 1,
  Sigma_diff_prior_scale = diag(J - 1),
  Sigma_diff_prior_custom = NA,
  Sigma_diff_prior_custom_test_par = diag(J - 1)
)

is.RprobitB_prior_Sigma_diff(x)

\method{print}{RprobitB_prior_Sigma_diff}(
  x,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = TRUE,
  details = FALSE,
  ...
)
}
\arguments{
\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{J}{An \code{integer}, the number of choice alternatives.}

\item{Sigma_diff_prior_df}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Sigma_diff}.
It must be greater or equal than \code{J + 1} (to ensure a proper prior).
By default, \code{Sigma_diff_prior_df = J + 1}.}

\item{Sigma_diff_prior_scale}{A \code{matrix} of dimension \code{(J-1)} x \code{(J-1)}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of
\code{Sigma_diff}.
By default, \code{Sigma_diff_prior_scale = diag(J-1)}.}

\item{Sigma_diff_prior_custom}{A \code{function}, a custom prior density function for \code{Sigma_diff},
see details.
By default, \code{Sigma_diff_prior_custom = NULL}, i.e., no custom prior.}

\item{Sigma_diff_prior_custom_test_par}{A \code{matrix} of dimension \code{(J-1)} x \code{(J-1)}, a test input for
the custom prior density function \code{Sigma_diff_prior_custom}.
By default, \code{Sigma_diff_prior_custom_test_par = diag(J-1)}.
Ignored if \code{Sigma_diff_prior_custom = NULL}.}

\item{x}{An \code{RprobitB_prior_Sigma_diff} object.}

\item{rowdots}{An \code{integer}, the row number which is replaced by \code{...}.
By default, \code{rowdots = 4}.}

\item{coldots}{An \code{integer}, the column number which is replaced by \code{...}.
By default, \code{coldots = 4}.}

\item{digits}{An \code{integer}, the number of printed decimal places.
Only relevant if input \code{x} is numeric.
By default, \code{digits = 2}.}

\item{simplify}{A \code{logical}, set to \code{TRUE} to simplify the output.
By default, \code{simplify = FALSE}.}

\item{details}{A \code{logical}, set to \code{TRUE} to print the type and
dimension of \code{x}.
By default, \code{details = !simplify}.}

\item{...}{Not used.}
}
\value{
An \code{RprobitB_prior_Sigma_diff} object.
}
\description{
This function defines the prior distributions for the probit model parameter
\code{Sigma_diff}. Only relevant if \code{ordered = FALSE}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item Only if \code{P_f > 0}:
\code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item Only if \code{P_r > 0}:
\code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item Only if \code{P_r > 0}:
\code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item Only in the unordered case (\code{ordered = FALSE}):
\code{Sigma_diff ~ Inverse-Wishart(df = Sigma_diff_prior_df, scale = Sigma_diff_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha} (only if \code{P_f > 0})
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b} (only if \code{P_r > 0})
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega} (only if \code{P_r > 0})
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma} (only if \code{ordered = TRUE})
\itemize{
\item \code{Sigma_prior_df = 3}
\item \code{Sigma_prior_scale = 1}
}
\item \code{Sigma_diff} (only if \code{ordered = FALSE})
\itemize{
\item \code{Sigma_diff_prior_df = J + 1}
\item \code{Sigma_diff_prior_scale = diag(J-1)}
}
\item \code{d} (only if \code{ordered = TRUE})
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are applied, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in {RprobitB}) can be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\examples{
### conjugate prior: Sigma_diff ~ Inverse-Wishart(J + 1, I)
RprobitB_prior_Sigma_diff(J = 3)

### custom prior: Sigma_diff ~ Inverse-Wishart(J + 1, I), Sigma_diff diagonal
RprobitB_prior_Sigma_diff(
  J = 3,
  Sigma_diff_prior_custom = function(x) {
    dwishart(x, df = 4, scale = diag(2), inv = TRUE) *
    all(x[row(x) != col(x)] == 0)
  }
)

}
\keyword{internal}
