% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtnorm}
\alias{rtnorm}
\title{Draw from truncated normal}
\usage{
rtnorm(mu, sig, trunpt, above)
}
\arguments{
\item{mu}{The mean.}

\item{sig}{The standard deviation.}

\item{trunpt}{The truncation point.}

\item{above}{A boolean, if \code{TRUE} truncate from above, otherwise from below.}
}
\value{
A numeric value.
}
\description{
This function draws from a truncated univariate normal distribution.
}
\keyword{dist}
