% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rttnorm}
\alias{rttnorm}
\title{Draw from two-sided truncated normal}
\usage{
rttnorm(mu, sig, lower_bound, upper_bound)
}
\arguments{
\item{mu}{The mean.}

\item{sig}{The standard deviation.}

\item{lower_bound}{The lower truncation point.}

\item{upper_bound}{The upper truncation point.}
}
\value{
A numeric value.
}
\description{
This function draws from a two-sided truncated univariate normal
distribution.
}
\examples{
### samples from a standard normal truncated at -2 and 2
R <- 1e4
draws <- replicate(R, rttnorm(0,1,-2,2))
plot(density(draws))
}
\keyword{distribution}
