% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_allocation}
\alias{sample_allocation}
\title{Sample allocation}
\usage{
sample_allocation(prob)
}
\arguments{
\item{prob}{[\code{numeric(C)}]\cr
The vector of class probabilities.}
}
\value{
An integer \code{1:C}.
}
\description{
Sample allocation
}
\examples{
sample_allocation(c(0.5, 0.3, 0.2))

}
\keyword{gibbs_sampler}
