% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_cov_matrix}
\alias{sample_cov_matrix}
\title{Sample covariance matrix}
\usage{
sample_cov_matrix(dim, df = dim, scale = diag(dim))
}
\arguments{
\item{dim}{An \code{integer}, the matrix dimension.}

\item{df}{An \code{integer}, the degrees of freedom.
Must be at least \code{dim}.
Per default, \code{df = dim}.}

\item{scale}{A \code{matrix}, the scale matrix.
Must be a covariance matrix.
Per default, \code{scale = diag(dim)}.}
}
\value{
A \code{matrix}, a covariance matrix.
}
\description{
This function samples a covariance matrix from a Wishart distribution.
}
\examples{
RprobitB:::sample_cov_matrix(dim = 3)

}
\keyword{internal}
\keyword{utils}
