% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_covariance_matrix}
\alias{sample_covariance_matrix}
\title{Sample covariance \code{matrix}}
\usage{
sample_covariance_matrix(dim, df = dim, scale = diag(dim))
}
\arguments{
\item{dim}{An \code{integer}, the matrix dimension.}

\item{df}{An \code{integer}, the degrees of freedom for the Wishart distribution.
Must be at least \code{dim}.
By default, \code{df = dim}.}

\item{scale}{A \code{matrix}, the scale matrix for the Wishart distribution.
Must be a covariance \code{matrix}.
By default, \code{scale = diag(dim)}.}
}
\value{
A covariance \code{matrix}.
}
\description{
This function samples a covariance \code{matrix} from a Wishart distribution.
}
\examples{
\dontrun{
sample_covariance_matrix(dim = 3)
}

}
\keyword{internal}
\keyword{utils}
