% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setup_data}
\alias{setup_data}
\title{Function to setup data for MCMC}
\usage{
setup_data(
  data,
  J = 2L,
  P_f = 0L,
  P_r = 0L,
  reference_level = 1L,
  ordered = FALSE,
  ranked = FALSE
)
}
\arguments{
\item{data}{A \code{list} that contains the data.
It is of length \code{N}, where \code{N} is the number of individuals.
The \code{n}-th element of \code{data} contains the data for the \code{n}-th
individual and is a \code{list} with two elements:
\itemize{
\item \code{X} is a \code{list} of length \code{Tp[n]}, where the
\code{t}-th element is the design \code{matrix} of \code{n}-th independent
variables for time point \code{t}.
Each design \code{matrix} must have \code{P_f + P_r} columns.
In the ordered case, each design \code{matrix} must have a single row.
In the non-ordered case, each design \code{matrix} have \code{J} rows.
\item \code{y} is a \code{vector} of length \code{Tp[n]}, where the
\code{t}-th element is an \code{integer} that refers to the dependent
variable from the underlying choice set.
}}

\item{J}{An \code{integer}, the number of levels of the dependent variable.}

\item{reference_level}{An \code{integer} between \code{1} and \code{J} that determines the
reference level for level normalization. It is not relevant and ignored in
the ordered case.}

\item{ordered}{A \code{logical} that determines if the dependent variable is ordered.}

\item{ranked}{A \code{logical} that determines if the dependent variable is ranked.}
}
\value{
A \code{list}... TODO
}
\description{
\enumerate{
\item TODO
\item 
\item 
}
}
