% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate choice data.}
\usage{
simulate(
  form,
  N,
  T,
  J,
  re = NULL,
  alternatives = NULL,
  parm = NULL,
  distr = NULL,
  standardize = NULL,
  seed = NULL
)
}
\arguments{
\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{alternatives}{A character vector with the names of the choice alternatives.}

\item{parm}{A named list of true parameter values:
\itemize{
\item \code{alpha}:
The fixed coefficient vector of length \code{P_f}.
\item \code{C:}
The number (greater or equal 1) of latent classes of decision makers.
If \code{P_r = 0}, then \code{C} is ignored.
\item \code{s}:
The vector of class weights of length \code{C}.
\item \code{b}:
The matrix of class means as columns of dimension \code{P_r} x \code{C}.
\item \code{Omega}:
The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
\item \code{Sigma}:
The error term covariance matrix of dimension \code{J} x \code{J}.
Internally, \code{Sigma} gets differenced with respect to alternative
\code{J}, so it becomes a covariance matrix of dimension
\code{J-1} x \code{J-1}.
}}

\item{distr}{A named list of number generation functions from which the covariates are
drawn. Each element of \code{distr} must be of the form
\code{"cov" = list("name" = "<name of the number generation function>", ...)},
where \code{cov} is the name of the covariate and \code{...} are required
parameters for the number generation function.
Covariates for which no distribution is specified are drawn from a standard
normal distribution.
Possible number generation functions are
\itemize{
\item functions of the type \code{r*} from base R (e.g. \code{rnorm}) where
all required parameters (except for \code{n}) must be specified,
\item the function \code{sample}, where all required parameters
(except for \code{size}) must be specified.
}}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{covariate_alternative}.
If \code{standardize = "all"}, all covariates get standardized.}

\item{seed}{Set a seed for the simulation.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function simulates choice data for the RprobitB package.
}
\details{
For more details see the vignette "Data management":
\code{vignette("data_management", package = "RprobitB")}.
}
\examples{
form = choice ~ cost | income | travel_time
re = "cost"
N = 100
T = 10
J = 3
alternatives = c("car", "bus", "train")
parm = list("C" = 2, "s" = c(0.5,0.5))
distr = list("cost" = list("name" = "rnorm", sd = 3),
             "income" = list("name" = "sample", x = (1:10)*1e3,
                             replace = TRUE),
             "travel_time_car" = list("name" = "rlnorm", meanlog = 1),
             "travel_time_bus" = list("name" = "rlnorm", meanlog = 2))
standardize = c("income", "travel_time_car", "travel_time_bus",
                "travel_time_train")
data = simulate(form = form, N = N, T = T, J = J, re = re,
                alternatives = alternatives, parm = parm, distr = distr,
                standardize = standardize)
}
