% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(model, parm, sd = 3)
}
\arguments{
\item{model}{A list of model information.}

\item{parm}{A list of true parameter values.}

\item{sd}{A numeric value, standard deviation for the simulation of the regressors. Currently fixed to \code{sd=3}.}
}
\value{
A list of \code{data} and additional \code{parm}.
}
\description{
Function that simulates choice data from a LCMMNP model.
}
