% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_data}
\alias{split_data}
\title{Split data into dependent and independent variables}
\usage{
split_data(data)
}
\arguments{
\item{data}{A \code{list} that contains the data.
It is of length \code{N}, where \code{N} is the number of individuals.
The \code{n}-th element of \code{data} contains the data for the \code{n}-th
individual and is a \code{list} with two elements:
\itemize{
\item \code{X} is a \code{list} of length \code{Tp[n]}, where the
\code{t}-th element is the design \code{matrix} of \code{n}-th independent
variables for time point \code{t}.
Each design \code{matrix} must have \code{P_f + P_r} columns.
In the ordered case, each design \code{matrix} must have a single row.
In the non-ordered case, each design \code{matrix} have \code{J} rows.
\item \code{y} is a \code{vector} of length \code{Tp[n]}, where the
\code{t}-th element is an \code{integer} that refers to the dependent
variable from the underlying choice set.
}}
}
\value{
A \code{list} with two elements:
\itemize{
\item \code{dependent_variable} is a \code{list} that contains the dependent
variables. It is of length \code{N}, where \code{N} is the number of
individuals. The \code{n}-th element of \code{dependent_variable} is a
\code{vector} of length \code{Tp[n]}, where the \code{t}-th element is an
\code{integer} that refers to the dependent variable from the underlying
choice set.
\item \code{independent_variable} is a \code{list} that contains the independent
variables. It is of length \code{N}, where \code{N} is the number of
individuals. The \code{n}-th element of \code{independent_variable} is a
\code{list} of length \code{Tp[n]}, where the \code{t}-th element is the
design \code{matrix} for time point \code{t}.
}
}
\description{
Split data into dependent and independent variables
}
