% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{start_timer}
\alias{start_timer}
\title{Routine to initialize the timer for Gibbs sampling}
\usage{
start_timer()
}
\value{
No return value, called for on-screen information.
}
\description{
Routine to initialize the timer for Gibbs sampling
}
