% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RprobitB_data.R
\name{summary.RprobitB_data}
\alias{summary.RprobitB_data}
\title{Summary method for \code{RprobitB_data}.}
\usage{
\method{summary}{RprobitB_data}(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_data}.}

\item{...}{Ignored.}
}
\description{
This function is the summary method for an object of class
\code{RprobitB_data}.
}
