% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RprobitB_model.R
\name{summary.RprobitB_model}
\alias{summary.RprobitB_model}
\title{Summary method for \code{RprobitB_model}}
\usage{
\method{summary}{RprobitB_model}(object, digits = 2, ...)
}
\arguments{
\item{object}{description}

\item{digits}{description}

\item{...}{description}
}
\description{
Summary method for \code{RprobitB_model}
}
