% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{transform.RprobitB_fit}
\alias{transform.RprobitB_fit}
\title{Transform an \code{RprobitB_fit} object}
\usage{
\method{transform}{RprobitB_fit}(
  `_data`,
  B = NULL,
  Q = NULL,
  scale = NULL,
  check_preference_flip = TRUE,
  ...
)
}
\arguments{
\item{_data}{An object of class \code{\link{RprobitB_fit}}.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}

\item{check_preference_flip}{Set to \code{TRUE} to check for flip in preferences after new \code{scale}.}

\item{...}{Ignored.}
}
\value{
The transformed \code{RprobitB_fit} object.
}
\description{
Given an object of class \code{RprobitB_fit}, this function can:
\itemize{
\item change the length \code{B} of the burn-in period,
\item change the the thinning factor \code{Q} of the Gibbs samples,
\item change the model normalization \code{scale}.
}
}
\details{
See the vignette "Model fitting" for more details:
\code{vignette("v03_model_fitting", package = "RprobitB")}.
}
\examples{
data(model_train)

### change the length B of the burn-in period
transform(model_train, B = 1)

### change the thinning factor Q
transform(model_train, Q = 1)

### change the scale
new_scale <- list(parameter = "s", index = 1, value = 1)
transform(model_train, scale = new_scale)

}
