% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Data transformation}
\usage{
transform_data(data_raw, cov_col, cov_ord, cov_zst)
}
\arguments{
\item{data_raw}{A data frame of choice data in "wide" format.}

\item{cov_col}{A numeric vector, columns of data_raw with covariates.}

\item{cov_ord}{A character vector, order of covariates, where fixed-coefficient covariates come first.}

\item{cov_zst}{A boolean, if \code{TRUE} covariates get z-standardized.}
}
\value{
A list of transformed data.
}
\description{
Function that transforms empricial data for estimation.
}
\details{
\code{data_raw} must contain columns named "id" (unique identifier for each decision maker)
and "choice" (the chosen alternatives).
}
