% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gibbs_samples.R
\name{transform_gibbs_samples}
\alias{transform_gibbs_samples}
\title{Transformation of Gibbs samples}
\usage{
transform_gibbs_samples(gibbs_samples, R, B, Q, scale)
}
\arguments{
\item{gibbs_samples}{The output of \code{\link{gibbs_sampling}}, i.e. a list of Gibbs samples.}

\item{R}{The number of Gibbs sampler iterations.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{scale}{A list of three elements, determining the parameter normalization with
respect to the scale of utility:
\itemize{
\item \code{parameter}:
either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"})
\item \code{index}:
the index of the parameter that gets fixed
\item \code{value}:
the value for the fixed parameter
}}
}
\value{
A list of transformed Gibbs samples. Each element is a list, containing the
Gibbs samples for \code{s}, \code{alpha}, \code{b}, \code{Omega}, and
\code{Sigma} (if available):
\itemize{
\item \code{gibbs_samples}:
The function input \code{gibbs_samples}
\item \code{gibbs_samples_n}:
A list of normalized samples based on \code{scale}
\item \code{gibbs_samples_nb}:
A list of normalized and burned samples based on \code{scale} and \code{B}
\item \code{gibbs_samples_nt}:
A list of normalized and thinned samples based on \code{scale} and \code{Q}
\item \code{gibbs_samples_nbt}:
A list of normalized, burned and thinned samples based on \code{scale},
\code{B} and \code{Q}
}
}
\description{
Function that normalizes, burns and thins the Gibbs samples.
}
