% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_parm.R
\name{transform_parm}
\alias{transform_parm}
\title{Transformation of true parameter values}
\usage{
transform_parm(parm = parm, scale = scale)
}
\arguments{
\item{parm}{A named list of true parameter values:
\itemize{
\item \code{alpha}:
The fixed coefficient vector of length \code{P_f}.
\item \code{C:}
The number (greater or equal 1) of latent classes of decision makers.
If \code{P_r = 0}, then \code{C} is ignored.
\item \code{s}:
The vector of class weights of length \code{C}.
\item \code{b}:
The matrix of class means as columns of dimension \code{P_r} x \code{C}.
\item \code{Omega}:
The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
\item \code{Sigma}:
The error term covariance matrix of dimension \code{J} x \code{J}.
Internally, \code{Sigma} gets differenced with respect to alternative
\code{J}, so it becomes a covariance matrix of dimension
\code{J-1} x \code{J-1}.
}}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}
}
\value{
The transformed version of \code{parm}
}
\description{
Function that transforms \code{parm} based on scale.
}
\examples{
parm = check_parm(parm = NULL, P_f = 2, P_r = 2, J = 3)
scale = check_scale(scale = NULL, P_f = 2, J = 3)
transform_parm(parm = parm, scale = scale)
}
