% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_true_parameter}
\alias{transform_true_parameter}
\title{Transformation of true parameter values.}
\usage{
transform_true_parameter(true_parameter, scale)
}
\arguments{
\item{true_parameter}{An object of class \code{RprobitB_parameter} with true model parameters.}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}
}
\value{
An object of class \code{RprobitB_parameter}.
}
\description{
This function transforms the true parameter values 'true_parameter' based
on 'scale'.
}
\examples{
true_parameter = RprobitB_parameter(P_f = 2, P_r = 2, J = 3, N = 100)
scale = RprobitB_scale(scale = NULL, P_f = 2, J = 3)
transform_parm(true_parameter = true_parameter, scale = scale)
}
