% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undiff_Sigma.R
\name{undiff_Sigma}
\alias{undiff_Sigma}
\title{Transform differenced to non-differenced error term covariance matrix.}
\usage{
undiff_Sigma(Sigma, i)
}
\arguments{
\item{Sigma}{An error term covariance matrix of dimension \code{J-1} x \code{J-1} which
was differenced with respect to alternative \code{i}.}

\item{i}{An integer, the alternative number with respect to which \code{Sigma}
was differenced.}
}
\value{
A covariance matrix of dimension \code{J} x \code{J}. If this covariance
matrix gets differenced with respect to alternative \code{i}, the results is
again \code{Sigma}.
}
\description{
This function transforms the differenced error term covariance matrix
\code{Sigma} back to a non-differenced error term covariance matrix.
}
