% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_b}
\alias{update_b}
\title{Update class means}
\usage{
update_b(beta, Omega, z, m, Sigma_b_0_inv, mu_b_0)
}
\arguments{
\item{beta}{The matrix of the decision-maker specific coefficient vectors of dimension
\code{P_r} x \code{N}.
Set to \code{NA} if \code{P_r = 0}.}

\item{Omega}{The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
Set to \code{NA} if \code{P_r = 0}.}

\item{z}{The vector of the allocation variables of length \code{N}.
Set to \code{NA} if \code{P_r = 0}.}

\item{m}{[\code{numeric(C)}]\cr
The vector of current class frequencies.}

\item{Sigma_b_0_inv}{[\code{matrix(P_r, P_r)}]\cr
The prior precision of the class mean.}
}
\value{
A matrix of updated means for each class in columns.
}
\description{
Update class means
}
\examples{
N <- 100
b <- cbind(c(0, 0), c(1, 1))
Omega <- matrix(c(1, 0.3, 0.3, 0.5, 1, -0.3, -0.3, 0.8), ncol = 2)
z <- c(rep(1, N / 2), rep(2, N / 2))
m <- as.numeric(table(z))
beta <- sapply(
  z, function(z) oeli::rmvnorm(n = 1, b[, z], matrix(Omega[, z], 2, 2))
)
update_b(
  beta = beta, Omega = Omega, z = z, m = m,
  Sigma_b_0_inv = diag(2), mu_b_0 = c(0, 0)
)

}
\keyword{gibbs_sampler}
