% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_b_c}
\alias{update_b_c}
\title{Update mean of a single class}
\usage{
update_b_c(bar_b_c, Omega_c, m_c, Sigma_b_0_inv, mu_b_0)
}
\arguments{
\item{bar_b_c}{[\code{numeric(P_r)}]\cr
The average observation of this class.}

\item{Omega_c}{[\code{matrix(P_r, P_r)}]\cr
The class covariance matrix.}

\item{m_c}{[\code{integer(1)}]\cr
The number of observations in this class.}

\item{Sigma_b_0_inv}{[\code{matrix(P_r, P_r)}]\cr
The prior precision of the class mean.}
}
\value{
An update for \code{b_c}.
}
\description{
Update mean of a single class
}
\examples{
update_b_c(
  bar_b_c = c(0, 0), Omega_c = diag(2), m_c = 10,
  Sigma_b_0_inv = diag(2), mu_b_0 = c(0, 0)
)

}
\keyword{gibbs_sampler}
