% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_classes}
\alias{update_classes}
\title{Weight-based update of latent classes}
\usage{
update_classes(
  rep,
  Cmax,
  epsmin,
  epsmax,
  distmin,
  s,
  m,
  b,
  Omega,
  print_progress
)
}
\description{
This function updates the latent classes based on their class weights.
}
\examples{
rep <- 1
Cmax <- 10
epsmin <- 0.1
epsmax <- 0.9
distmin <- 0.1
s <- 0.92
m <- 94
b <- matrix(1:3,ncol=1)
Omega <- matrix(1:9,ncol=1)
update_classes(rep = rep, Cmax = Cmax, epsmin = epsmin, epsmax = epsmax, distmin = distmin,
               s = s, m = m, b = b, Omega = Omega, print_progress = TRUE)

}
\keyword{internal}
