% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_s}
\alias{update_s}
\title{Update class weight vector}
\usage{
update_s(delta, m)
}
\arguments{
\item{delta}{A numeric for the concentration parameter vector \code{rep(delta,C)} of the
Dirichlet prior for \code{s}.
Per default, \code{delta = 1}. In case of Dirichlet process-based updates of the
latent classes, \code{delta = 0.1} per default.}

\item{m}{The vector of current class frequencies.}
}
\value{
A draw from the Dirichlet posterior distribution for \code{s}.
}
\description{
Update class weight vector
}
\examples{
update_s(delta = 1, m = sample.int(4))

}
\keyword{gibbs_sampler}
