% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_timer}
\alias{update_timer}
\title{Routine to update the timer for Gibbs sampling}
\usage{
update_timer(rep, R)
}
\arguments{
\item{rep}{An integer, current iteration}

\item{R}{An integer, total number of iterations}
}
\value{
No return value, called for on-screen information.
}
\description{
Routine to update the timer for Gibbs sampling
}
