% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_z}
\alias{update_z}
\title{Update class allocation vector}
\usage{
update_z(s, beta, b, Omega)
}
\arguments{
\item{s}{[\code{numeric(C)}]\cr
The vector of class weights.}

\item{beta}{[\code{matrix(nrow = P_r, ncol = N)}]\cr
The matrix of the decider-specific coefficient vectors.}

\item{b}{[\code{matrix(nrow = P_r, ncol = C)}]\cr
The matrix of class means as columns.}

\item{Omega}{[\code{matrix(nrow = P_r * P_r, ncol = C)}]\cr
The matrix of vectorized class covariance matrices as columns.}
}
\value{
An update for \code{z}.
}
\description{
Update class allocation vector
}
\examples{
update_z(
  s = c(0.6, 0.4), beta = matrix(c(-2, 0, 2), ncol = 3),
  b = cbind(0, 1), Omega = cbind(1, 1)
)

}
\keyword{gibbs_sampler}
