% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_alternatives.R
\name{validate_RprobitB_alternatives}
\alias{validate_RprobitB_alternatives}
\title{Validator for \code{RprobitB_alternatives} object}
\usage{
validate_RprobitB_alternatives(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_alternatives}.}
}
\value{
The input \code{x}.
}
\description{
This function validates an \code{RprobitB_alternatives} object.
}
