% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_formula.R
\name{validate_RprobitB_formula}
\alias{validate_RprobitB_formula}
\title{Validator for \code{RprobitB_formula} object}
\usage{
validate_RprobitB_formula(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_formula}.}
}
\value{
The input \code{x}.
}
\description{
This function validates an \code{RprobitB_formula} object.
}
