% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parm.R
\name{check_parm}
\alias{check_parm}
\title{Check \code{parm}}
\usage{
check_parm(parm, model, data, norm)
}
\arguments{
\item{parm}{A list of true parameter values.}

\item{model}{A list of model information.}

\item{data}{A list of data information.}

\item{norm}{A list of normalization information.}
}
\value{
\code{parm}
}
\description{
Function that checks the input \code{parm} and draws missing parameter values.
}
