% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice_probabilities.R
\name{choice_probabilities}
\alias{choice_probabilities}
\title{Compute choice probabilities of an \code{RprobitB_model}.}
\usage{
choice_probabilities(object, at_true = TRUE)
}
\arguments{
\item{object}{An object of class \code{RprobitB_model}.}

\item{at_true}{If \code{TRUE}, choice probabilities are computed for the true parameter
values.}
}
\value{
A data frame, one row per choice situation and one column per alternative.
}
\description{
This function computes the choice probabilities of an \code{RprobitB_model}.
}
\examples{
p = simulate(form = choice ~ var | 0, N = 100, T = 10, J = 2, seed = 1)
m1 = mcmc(data = p, seed = 1)
}
