% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.RprobitB_fit}
\alias{coef.RprobitB_fit}
\title{Linear coefficients}
\usage{
\method{coef}{RprobitB_fit}(object, scale = 1, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{scale}{A factor by which the coefficients are multiplied.}

\item{...}{Ignored.}
}
\value{
An object of class \code{RprobitB_coef}.
}
\description{
This function returns the estimated linear coefficients.
}
