% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.RprobitB_fit.R
\name{coef.RprobitB_fit}
\alias{coef.RprobitB_fit}
\alias{print.RprobitB_coef}
\alias{plot.RprobitB_coef}
\title{Extract model effects}
\usage{
\method{coef}{RprobitB_fit}(object, ...)

\method{print}{RprobitB_coef}(x, ...)

\method{plot}{RprobitB_coef}(x, sd = 1, het = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Ignored.}

\item{x}{An object of class \code{RprobitB_coef}.}

\item{sd}{The number of standard deviations to display.}

\item{het}{Set to \code{FALSE} to show the standard deviation of the estimate.
Set to \code{TRUE} to show the standard deviation of the mixing distribution.}
}
\value{
An object of class \code{RprobitB_coef}.
}
\description{
This function extracts the estimated model effects.
}
