% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare RprobitB models}
\usage{
compare(...)
}
\arguments{
\item{...}{Models from \link[RprobitB]{fit}.}
}
\value{
WAIC value (numeric).
}
\description{
Function that compares RprobitB models based on their
Widely Applicable Information Criterion (WAIC) value.
}
\details{
For more details see the vignette "Model comparison":
\code{vignette("model_comparison", package = "RprobitB")}
}
\examples{
### compare the same model with different number of Gibbs samples
}
