% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare fitted models.}
\usage{
compare(...)
}
\arguments{
\item{...}{One or more objects of class \code{RprobitB_model}.}
}
\value{
A matrix with the following model selection criteria:
\itemize{
\item log-likelihood
\item AIC
\item BIC
\item WAIC
\item effective model parameters
}
}
\description{
This function computes model selection criteria.
}
\details{
For more details see the vignette "Model comparison":
\code{vignette("model_comparison", package = "RprobitB")}.
}
\examples{
### example missing here
}
