% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_choice_probabilities.R
\name{compute_choice_probabilities}
\alias{compute_choice_probabilities}
\title{Compute probit choice probabilities for a single choice situation.}
\usage{
compute_choice_probabilities(X, parameter)
}
\arguments{
\item{X}{A matrix of covariates with \code{J} rows and \code{P_f + P_r} columns, where
the first \code{P_f} columns are connected to fixed coefficients and the last
\code{P_r} columns are connected to random coefficients.}

\item{parameter}{An object of class \code{RprobitB_parameter}.}
}
\value{
A probability vector of length \code{J}.
}
\description{
This function computes the probit choice probabilities for a single choice
situation with \code{J} alternatives.
}
\keyword{internal}
