% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_point_estimates.R
\name{compute_point_estimates}
\alias{compute_point_estimates}
\title{Compute point estimates of an \code{RprobitB_model}.}
\usage{
compute_point_estimates(x, FUN = mean)
}
\arguments{
\item{x}{An object of class \code{\link{RprobitB_model}}.}

\item{FUN}{A function that computes a single numeric value out of a vector of numeric
values.}
}
\value{
An object of class \code{\link{RprobitB_parameter}}.
}
\description{
This function computes the point estimates of an \code{\link{RprobitB_model}}.
Per default, the \code{mean} of the Gibbs samples is used as a point estimate.
However, any statistic that computes a single numeric value out of a vector of
Gibbs samples can be specified for \code{FUN}.
}
\keyword{internal}
