% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sufficient_statistics.R
\name{compute_sufficient_statistics}
\alias{compute_sufficient_statistics}
\title{Compute sufficient statistics}
\usage{
compute_sufficient_statistics(data, scale)
}
\arguments{
\item{data}{An object of class \code{RprobitB_data}, which can be obtained from
\code{\link{simulate}} or \code{\link{prepare}}.}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}
}
\value{
A list of sufficient statistics.
}
\description{
Function that computes sufficient statistics for estimation.
}
\examples{
data = simulate(form = choice ~ var, N = 10, T = 10, J = 3, re = "ASC")
scale = check_scale(scale = NULL, P_f = 2, J = 3)
compute_sufficient_statistics(data = data, scale = scale)
}
