% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic.R
\name{compute_waic}
\alias{compute_waic}
\title{Compute WAIC value}
\usage{
compute_waic(gibbs_samples)
}
\arguments{
\item{gibbs_samples}{A list of gibbs samples}
}
\value{
The WAIC value
}
\description{
Function that computes WAIC value
}
