% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic.R
\name{compute_waic}
\alias{compute_waic}
\title{Compute WAIC value.}
\usage{
compute_waic(x, no_samples, pis = NULL)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}
}
\value{
The WAIC value.
}
\description{
This function computes the WAIC value of an \code{RprobitB_model}.
}
\examples{
model = mcmc(data = simulate(form = choice ~ var | 0, N = 100, T = 10, J = 2))
compute_waic(model)
}
