% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{create_labels_b}
\alias{create_labels_b}
\title{Create labels for \code{b}}
\usage{
create_labels_b(P_r, C)
}
\arguments{
\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}
}
\value{
A vector of labels for the model parameter \code{b} of length \code{P_r * C}
if \code{P_r > 0} and \code{NULL} otherwise.
}
\description{
This function creates labels for the model parameter \code{b}.
}
\details{
The labels are of the form \code{"c.p"}, where \code{c} is the latent class
number and \code{p} the index of the random coefficient.
}
\examples{
RprobitB:::create_labels_b(2,3)
}
\keyword{internal}
