% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{create_labels_Omega}
\alias{create_labels_Omega}
\title{Create labels for \code{Omega}.}
\usage{
create_labels_Omega(P_r, C, cov_sym)
}
\value{
A vector of labels for the model parameter \code{Omega}.
}
\description{
This function creates labels for the model parameter \code{Omega}.
}
\keyword{internal}
