% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_labels.R
\name{create_labels_s}
\alias{create_labels_s}
\title{Create labels for \code{s}.}
\usage{
create_labels_s(P_r, C)
}
\arguments{
\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{C}{The number of latent classes.}
}
\value{
A vector of labels for the model parameter \code{s}.
}
\description{
This function creates labels for the model parameter \code{s}.
}
\keyword{internal}
