% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_labels.R
\name{create_parameter_labels}
\alias{create_parameter_labels}
\title{Create parameters labels.}
\usage{
create_parameter_labels(
  P_f,
  P_r,
  J,
  C,
  cov_sym,
  keep_par = c("s", "alpha", "b", "Omega", "Sigma"),
  drop_par = NULL
)
}
\arguments{
\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{C}{The number of latent classes.}

\item{cov_sym}{If \code{TRUE} considers symmetric covariance matrix elements.}

\item{keep_par}{A vector of parameter names which are kept.}

\item{drop_par}{A vector of parameter names which get dropped.}
}
\value{
A list of labels for the model parameters.
}
\description{
This function creates labels for the model parameters.
}
\keyword{internal}
