% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Difference operator.}
\usage{
delta(J, i)
}
\arguments{
\item{J}{The total number of alternatives.}

\item{i}{The alternative number to which respect utility differences are computed.}
}
\value{
A matrix of dimension \code{J-1} x \code{J}.
}
\description{
This function computes the difference operator matrix for computing utility
differences.
}
\details{
Given a \code{J} x \code{P} matrix \code{X} of choice characteristics,
then \code{delta(i,J)\%*\%X} computes differences with respect to alternative
\code{i}.
}
\examples{
X = matrix(1:9,3,3, byrow = TRUE)
delta(3,3) \%*\% X
}
