% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance}
\usage{
euc_dist(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{Another numeric vector of the same length as \code{a}.}
}
\value{
The euclidean distance.
}
\description{
This function computes the euclidean distance between two vectors.
}
\keyword{utils}
