% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a (latent class) (mixed) (multinomial) probit model}
\usage{
fit(
  data,
  scale = list(parameter = "s", index = 1, value = 1),
  R = 10000,
  B = R/2,
  Q = 10,
  print_progress = TRUE,
  latent_classes = NULL,
  prior = NULL
)
}
\arguments{
\item{data}{An object of class \code{RprobitB_data}, which can be obtained from
\code{\link{simulate}} or \code{\link{prepare}}.}

\item{scale}{A list of three elements, determining the parameter normalization with
respect to the scale of utility:
\itemize{
\item \code{parameter}:
either \code{"a"} (for a linear coefficient in \code{"alpha"}) or
\code{"s"} (for a variance in the error-term covariance matrix
\code{"Sigma"})
\item \code{index}:
the index of the parameter that gets fixed
\item \code{value}:
the value for the fixed parameter
}}

\item{R}{The number of Gibbs sampler iterations.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{print_progress}{A boolean, determining whether to print the progress.}

\item{latent_classes}{A list of parameters specifying the number and updating scheme of latent
classes:
\itemize{
\item \code{C}:
The number (greater or equal 1) of latent classes. Set to 1 per default.
\item \code{update}:
A boolean, determining whether to update \code{C}.
\item \code{Cmax}:
The maximum number of latent classes. Ignored if \code{update = FALSE}.
\item \code{buffer}:
The updating buffer (number of iterations to wait before the next update).
Ignored if \code{update = FALSE}.
\item \code{epsmin}:
The threshold weight for removing latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{epsmax}:
The threshold weight for splitting latent classes (between 0 and 1).
Ignored if \code{update = FALSE}.
\item \code{distmin}:
The threshold difference in means for joining latent classes
(non-negative). Ignored if \code{update = FALSE}.
}}

\item{prior}{A list of parameters for the prior distributions of the normalized
parameters:
\itemize{
\item \code{eta}:
The mean vector of length \code{P_f} of the normal prior for
\code{alpha}.
\item \code{Psi}:
The covariance matrix of dimension \code{P_f} x \code{P_f} of the
normal prior for \code{alpha}.
\item \code{delta}:
The concentration parameter of length 1 of the Dirichlet prior for
\code{s}.
\item \code{xi}:
The mean vector of length \code{P_r} of the normal prior for each
\code{b_c}.
\item \code{D}:
The covariance matrix of dimension \code{P_r} x \code{P_r} of the
normal prior for each \code{b_c}.
\item \code{nu}:
The degrees of freedom (a natural number greater than \code{P_r}) of
the Inverse Wishart prior for each \code{Omega_c}.
\item \code{Theta}:
The scale matrix of dimension \code{P_r} x \code{P_r} of the
Inverse Wishart prior for each \code{Omega_c}.
\item \code{kappa}:
The degrees of freedom (a natural number greater than \code{J-1}) of
the Inverse Wishart prior for \code{Sigma}.
\item \code{E}:
The scale matrix of dimension \code{J-1} x \code{J-1} of the
Inverse Wishart prior for \code{Sigma}.
}}
}
\value{
An object of class \code{RprobitB_model}
}
\description{
Function that fits a (latent class) (mixed) (multinomial) probit model via
Bayesian estimation.
}
\details{
For more details see the vignette "Model fitting":
\code{vignette("model_fitting", package = "RprobitB")}
}
\examples{
data = simulate(form = choice ~ var, N = 10, T = 10, J = 3, re = "ASC")
model = fit(data = data)
}
