% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{formula.RprobitB_fit}
\alias{formula.RprobitB_fit}
\title{Extract the fitted formula}
\usage{
\method{formula}{RprobitB_fit}(x, ...)
}
\arguments{
\item{x}{[\code{RprobitB_fit}]\cr
Fitted choice model.}

\item{...}{Currently not used.}
}
\value{
A \code{formula} object.
}
\description{
Returns the normalized three-part model formula stored in a fitted model.
}
\examples{
model <- fit(choice ~ x, chains = 1)
formula(model)
}
\keyword{models}
