% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_body}
\alias{function_body}
\title{Extract function body as character}
\usage{
function_body(fun, braces = FALSE, nchar = 100)
}
\arguments{
\item{fun}{An object of class \code{function}.}

\item{braces}{A boolean, if \code{FALSE} (default) removes \code{"{"} and \code{"}"} at
start and end (if any).}

\item{nchar}{An integer, the maximum number of characters before abbreviation.}
}
\value{
A character.
}
\description{
This function extracts the body of a function as a character.
}
\examples{
add1 <- function(x) {
  stopifnot(is.numeric(x))
  x + 1
}
RprobitB:::function_body(add1)
RprobitB:::function_body(add1, braces = TRUE)
RprobitB:::function_body(add1, nchar = 20)
}
\keyword{utils}
