% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{function_body}
\alias{function_body}
\title{Extract function body as \code{character}}
\usage{
function_body(fun, braces = FALSE, nchar = getOption("width") - 4)
}
\arguments{
\item{fun}{A \code{function}.}

\item{braces}{A \code{logical}, if \code{FALSE} (default) removes \code{"{"} and \code{"}"}
at start and end (if any).}

\item{nchar}{An \code{integer}, the maximum number of characters before abbreviation.
Must be at least \code{3}.
By default, \code{nchar = getOption("width")}.}
}
\value{
A \code{character}, the body of \code{f}.
}
\description{
This function extracts the body of a function as a \code{character}.
}
\examples{
\donttest{
fun <- is_covariance_matrix
function_body(fun)
function_body(fun, braces = TRUE)
function_body(fun, nchar = 30)
}

}
\keyword{internal}
\keyword{utils}
