% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{get_cov}
\alias{get_cov}
\title{Get covariates of choice situation}
\usage{
get_cov(x, id, idc, idc_label)
}
\arguments{
\item{x}{Either an object of class \code{RprobitB_data} or \code{RprobitB_fit}.}

\item{id}{A numeric (vector), that specifies the decider(s).}

\item{idc}{A numeric (vector), that specifies the choice occasion(s).}

\item{idc_label}{The name of the column that contains the choice occasion identifier.}
}
\value{
A subset of the \code{choice_data} data frame specified in \code{prepare_data()}.
}
\description{
This convenience function returns the covariates and the choices of specific
choice occasions.
}
\examples{
data("model_train", package = "RprobitB")
get_cov(model_train, id = 1:2, idc = 1:2, idc_label = "choiceid")

}
