% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampling}
\alias{gibbs_sampling}
\title{Gibbs sampler}
\usage{
gibbs_sampling(
  sufficient_statistics,
  prior,
  latent_classes,
  init,
  R,
  B,
  print_progress
)
}
\arguments{
\item{sufficient_statistics}{The output of \code{\link{sufficient_statistics}}.}

\item{prior}{A named list of parameters for the prior distributions. See the documentation
of \code{\link{check_prior}} for details about which parameters can be specified.}

\item{latent_classes}{Either \code{NULL} or a list of parameters specifying the number and the
latent classes:
\itemize{
\item \code{C}:
The number (greater or equal 1) of latent classes, which is set to 1 per
default and is ignored if \code{P_r = 0}.
If \code{update = TRUE}, \code{C} equals the initial number of classes.
\item \code{update}:
A boolean, determining whether to update \code{C}.
Ignored if \code{P_r = 0}.
If \code{update = FALSE}, all of the following elements are ignored.
\item \code{Cmax}:
The maximum number of latent classes.
\item \code{buffer}:
The updating buffer (number of iterations to wait before the next update).
\item \code{epsmin}:
The threshold weight for removing latent classes (between 0 and 1).
\item \code{epsmax}:
The threshold weight for splitting latent classes (between 0 and 1).
\item \code{distmin}:
The threshold difference in means for joining latent classes
(non-negative).
}}

\item{init}{The output of \code{\link{set_initial_gibbs_values}}.}

\item{R}{The number of iterations of the Gibbs sampler.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{print_progress}{A boolean, determining whether to print the Gibbs sampler progress and the
estimated remaining computation time.}
}
\value{
A list of Gibbs samples for \code{Sigma}, \code{alpha} (if \code{P_f>0})
and \code{s}, \code{b}, \code{Omega} and a vector of classifications
(if \code{P_r>0}).
}
\description{
This function draws Gibbs samples from the posterior distribution of the
multinomial probit model parameters.
}
\keyword{internal}
