% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_bool}
\alias{is_bool}
\title{Check boolean}
\usage{
is_bool(x)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is either \code{TRUE} or
\code{FALSE}, \code{FALSE} else.
}
\description{
This function checks whether the input is a single boolean.
}
\examples{
RprobitB:::is_bool("TRUE")
RprobitB:::is_bool(FALSE)

}
\keyword{internal}
\keyword{utils}
