% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_covariance_matrix}
\alias{is_covariance_matrix}
\title{Check covariance matrix properties.}
\usage{
is_covariance_matrix(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A boolean, \code{TRUE} if \code{x} is a proper covariance matrix.
}
\description{
This function checks if the input is a proper covariance matrix, i.e. a
symmetric, numeric matrix with non-negative eigenvalues.
}
\examples{
x <- diag(2)
is_covariance_matrix(x)

}
\keyword{utils}
