% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_single_numeric}
\alias{is_single_numeric}
\title{Check single numeric}
\usage{
is_single_numeric(i)
}
\arguments{
\item{x}{Any element.}
}
\value{
A \code{logical}, \code{TRUE} if \code{x} is a single numeric, \code{FALSE}
else.
}
\description{
This function checks whether the input is a single numeric value.
}
\examples{
RprobitB:::is_single_numeric(1)
RprobitB:::is_single_numeric("1")
RprobitB:::is_single_numeric(NA_real_)

}
\keyword{internal}
\keyword{utils}
