% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_ordered}
\alias{ll_ordered}
\title{Choice probability in the ordered probit model}
\usage{
ll_ordered(d, y, mu, Tvec)
}
\arguments{
\item{d}{A numeric vector of threshold increments.}

\item{y}{TBA}

\item{mu}{TBA}

\item{Tvec}{TBA}
}
\value{
TBA
}
\description{
This function computes the conditional probability of one choice occasion
given the threshold increments \code{d}.
}
\details{
TBA
}
\examples{
ll_ordered(c(0,0,0), 1, 1, FALSE)
}
\keyword{posterior}
