% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_ordered}
\alias{ll_ordered}
\title{Log-likelihood in the ordered probit model}
\usage{
ll_ordered(d, y, mu, Tvec)
}
\arguments{
\item{d}{A numeric vector of threshold increments.}

\item{y}{A matrix of the choices.}

\item{mu}{A matrix of the systematic utilities.}

\item{Tvec}{The element \code{Tvec} in \code{\link{sufficient_statistics}}.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood value given the threshold
increments \code{d}.
}
\examples{
ll_ordered(c(0,0,0), 1, 1, FALSE)
}
\keyword{internal}
\keyword{posterior}
