% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute log-likelihood of an \code{RprobitB_fit}.}
\usage{
log_likelihood(x, at_true = FALSE)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{at_true}{If \code{TRUE}, choice probabilities are computed at the true parameter
values (if they are available).}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood of an \code{RprobitB_fit}.
}
\examples{
data <- simulate_choices(form = choice ~ covariate, N = 10, T = 10, J = 2)
x <- mcmc(data)
log_likelihood(x)
}
