% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{logLik}
\alias{logLik}
\title{Log-likelihood value}
\usage{
logLik(object, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{...}{Ignored.}
}
\value{
A numeric.
}
\description{
This function computes the log-likelihood value of an \code{RprobitB_fit}
object.
}
\examples{
data("model_train", package = "RprobitB")
logLik(model_train)

}
