% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.frame.RprobitB_fit}
\alias{model.frame.RprobitB_fit}
\title{Extract the fitted data}
\usage{
\method{model.frame}{RprobitB_fit}(formula, ...)
}
\arguments{
\item{formula}{[\code{RprobitB_fit}]\cr
Fitted choice model.}

\item{...}{Currently not used.}
}
\value{
A \code{data.frame} containing the fitted choice data.
}
\description{
Returns the choice data stored in a fitted model as a \code{data.frame}.
}
\examples{
model <- fit(choice ~ x, chains = 1)
head(model.frame(model))
}
\keyword{models}
