% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_elec}
\alias{model_elec}
\title{Mixed probit model for multivariate choice between electricity suppliers}
\format{
An object of class \code{RprobitB_fit}.
}
\usage{
data(model_elec)
}
\description{
This object is a fitted mixed probit model to the Electricity dataset of the
{mlogit} package with the model formula \code{choice ~ pf + cl + loc + wk + tod + seas | 0}.
}
\details{
The model was derived via
\preformatted{
data("Electricity", package = "mlogit")
Electricity <- as_cov_names(Electricity, c("pf","cl","loc","wk","tod","seas"), 1:4)
data <- prepare_data(
  form = choice ~ pf + cl + loc + wk + tod + seas | 0,
  choice_data = Electricity,
  re = c("cl","loc","wk","tod","seas")
)
model <- mcmc(
  data = data,
  R = 5000,
  Q = 10,
  scale = list(parameter = "a", index = 1, value = -1))
}
}
\keyword{model}
