% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{overview_effects}
\alias{overview_effects}
\title{Effect overview}
\usage{
overview_effects(form, re = NULL, alternatives, base = NULL)
}
\arguments{
\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative due to identifiability).
They can be removed by adding \code{+0} in the second spot.
See the vignette on choice data for more details.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{base}{A character, the name of the base alternative for covariates that are not
alternative specific (type 2 covariates).}
}
\value{
A data frame, each row is a effect, columns are the effect name
\code{"effect"}, and booleans whether the covariate is alternative-specific
\code{"as_value"}, whether the coefficient is alternative-specific
\code{"as_coef"}, and whether it is a random effect \code{"random"}.
}
\description{
This function gives an overview of the effect names, whether the covariate
is alternative-specific, whether the coefficient is alternative-specific,
and whether it is a random effect.
}
\examples{
overview_effects(
  form = choice ~ price + time + comfort + change | 1,
  re = c("price", "time"),
  alternatives = c("A", "B"),
  base = "A"
)

}
\seealso{
\code{\link[=check_form]{check_form()}} for checking the model formula specification.
}
