% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{parameter_labels}
\alias{parameter_labels}
\title{Create parameters labels}
\usage{
parameter_labels(
  P_f,
  P_r,
  J,
  N,
  C,
  cov_sym,
  keep_par = c("s", "alpha", "b", "Omega", "Sigma"),
  drop_par = NULL
)
}
\arguments{
\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{cov_sym}{Set to \code{TRUE} for labels of symmetric covariance elements.}

\item{keep_par}{A vector of parameter names which are kept.}

\item{drop_par}{A vector of parameter names which get dropped.}
}
\value{
A list of labels for the selected model parameters.
}
\description{
This function model parameter labels.
}
\examples{
parameter_labels()

}
\keyword{internal}
