% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RprobitB_data.R
\name{plot.RprobitB_data}
\alias{plot.RprobitB_data}
\title{Plot method for \code{RprobitB_data}.}
\usage{
\method{plot}{RprobitB_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_data}.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{RprobitB_data}.
}
