% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RprobitB_data.R
\name{plot.RprobitB_data}
\alias{plot.RprobitB_data}
\title{Plot method for \code{RprobitB_data}.}
\usage{
\method{plot}{RprobitB_data}(x, alpha = 0.8, bins = 30, position = "dodge", ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_data}.}

\item{alpha}{Passed to \link[ggplot2]{geom_histogram}.}

\item{bins}{Passed to \link[ggplot2]{geom_histogram}.}

\item{position}{Passed to \link[ggplot2]{geom_histogram}.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class
\code{RprobitB_data}.
}
