% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.RprobitB_fit}
\alias{plot.RprobitB_fit}
\title{Plot method for \code{RprobitB_fit}}
\usage{
\method{plot}{RprobitB_fit}(x, type, ignore = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{RprobitB_fit}}.}

\item{type}{The type of plot, which can be one of:
\itemize{
\item \code{"mixture"} for visualizing the mixture distribution.
\item \code{"acf"} for autocorrelation plots of the Gibbs samples.
\item \code{"trace"} for trace plots of the Gibbs samples.
\item \code{"class_seq"} for visualizing the sequence of class numbers.
\item \code{"class_allocation"} for visualizing the class allocation
(only if \code{P_r = 2}) at the final Gibbs sampler iteration.
See the details section for visualization options.
}}

\item{ignore}{A character (vector) of covariate or parameter names that do not get
visualized.}

\item{...}{Additional parameters, see the details section.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{RprobitB_fit}.
}
\details{
Some plot types have additional options that can be specified via
submitting the following parameters as ellipsis arguments.
\subsection{\code{"type = class_allocation"}}{

\itemize{
\item A numeric vector \code{iterations} for plotting the class allocation
at different iterations of the Gibbs sampler.
\item A numeric \code{perc} between 0 and 1 to draw the \code{perc} percentile
ellipsoids for the underlying Gaussian distributions
(\code{perc = 0.95} per default).
\item A numeric \code{sleep}, the number of seconds to pause after plotting.
The default is 1.
}
}
}
