% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.RprobitB_fit}
\alias{plot.RprobitB_fit}
\title{Visualize fitted probit model}
\usage{
\method{plot}{RprobitB_fit}(x, type, ignore = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{RprobitB_fit}}.}

\item{type}{[\code{character(1)}]\cr
The type of plot, which can be one of:
\itemize{
\item \code{"mixture"} to visualize the mixing distribution,
\item \code{"acf"} for autocorrelation plots of the Gibbs samples,
\item \code{"trace"} for trace plots of the Gibbs samples,
\item \code{"class_seq"} to visualize the sequence of class numbers.
}}

\item{ignore}{[\code{character()}]\cr
Covariate or parameter names that do not get visualized.}

\item{...}{Currently not used.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{RprobitB_fit}.
}
