% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RprobitB_model.R
\name{plot.RprobitB_model}
\alias{plot.RprobitB_model}
\title{Plot method for \code{RprobitB_model}.}
\usage{
\method{plot}{RprobitB_model}(x, type = "effects", ignore = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{RprobitB_model}}.}

\item{type}{The type of plot, which can be one or more of:
\itemize{
\item \code{"effects"} (the default) for visualizing the linear effects,
\item \code{"mixture"} for visualizing the mixture distribution,
\item \code{"acf"} for autocorrelation plots of the Gibbs samples,
\item \code{"trace"} for trace plots of the Gibbs samples.
}}

\item{ignore}{A character (vector) of covariate or parameter names that do not get
visualized.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class
\code{RprobitB_model}.
}
