% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RprobitB_model.R
\name{plot.RprobitB_model}
\alias{plot.RprobitB_model}
\title{Plot method for an \code{RprobitB_model}}
\usage{
\method{plot}{RprobitB_model}(x, type, pars, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}}

\item{type}{The type of plot}

\item{pars}{A vector of parameters to plot}

\item{...}{Additional parameters passed on to \code{plot}}
}
\description{
Plot method for an \code{RprobitB_model}
}
