% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Autocorrelation plots}
\usage{
plot_acf(gibbs_samples, names)
}
\arguments{
\item{gibbs_samples}{A list of Gibbs samples.}
}
\value{
...
}
\description{
Function that plots the autocorrelation of the Gibbs samples.
}
\details{
Function computes the \href{https://mc-stan.org/docs/2_18/reference-manual/effective-sample-size-section.html}{effective sample size}.
}
