% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Autocorrelation plot of Gibbs samples.}
\usage{
plot_acf(gibbs_samples, par_labels)
}
\arguments{
\item{gibbs_samples}{A matrix of Gibbs samples.}

\item{par_labels}{A character vector of length equal to the number of columns of
\code{gibbs_samples}, containing labels for the Gibbs samples.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots the autocorrelation of the Gibbs samples, including the
total sample size \code{SS}, effective sample size \code{ESS} and the factor
\code{SS/ESS}.
}
\keyword{internal}
