% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Visualizing the linear effects.}
\usage{
plot_effects(gibbs_samples, coeff_names)
}
\arguments{
\item{gibbs_samples}{An object of class \code{RprobitB_gibbs_samples}.}

\item{coeff_names}{A character vector of coefficient names.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function visualizes the linear effects of the covariates on the choices
together with an uncertainty interval of plus / minus one standard deviation.
}
\keyword{internal}
