% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Visualizing the effects.}
\usage{
plot_effects(x, restrict = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}

\item{restrict}{A character (vector) of covariate or parameter names, which can be used to
restrict the visualizations to a subset of covariates and parameters,
respectively. Per default, a combined plot of all parameters is produced.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function visualizes the effects of the covariates on the choices (i.e.
the linear coefficients) along with an uncertainty interval of plus and minus
one standard deviation.
}
