% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture_contour.R
\name{plot_mixture_contour}
\alias{plot_mixture_contour}
\title{Plotting mixing distribution contours.}
\usage{
plot_mixture_contour(
  mean_est,
  weight_est,
  cov_est,
  beta_true = NULL,
  cov_names = NULL
)
}
\arguments{
\item{mean_est}{A list of length \code{C}, where each element is a vector of two
estimated class means.}

\item{weight_est}{A numeric vector of length \code{C} with estimated class weights.}

\item{cov_est}{A list of length \code{C}, where each element is an estimated class
covariance matrix.}

\item{beta_true}{Either \code{NULL} or a matrix of \code{C} rows with true \code{beta} values.}

\item{cov_names}{Either \code{NULL} or a vector of two covariate names.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots contours of the estimated mixing distributions and adds
the true beta values for comparison if available.
}
