% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mixture_contour}
\alias{plot_mixture_contour}
\title{Plot bivariate contour of mixing distributions}
\usage{
plot_mixture_contour(means, covs, weights, names)
}
\arguments{
\item{means}{The class means.}

\item{covs}{The class covariances.}

\item{weights}{The class weights.}

\item{names}{The covariate names.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function plots an estimated bivariate contour mixing distributions.
}
\keyword{internal}
