% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mixture_contour}
\alias{plot_mixture_contour}
\title{Plot bivariate contour of mixing distributions}
\usage{
plot_mixture_contour(means, covs, weights, names)
}
\arguments{
\item{means}{The class means.}

\item{covs}{The class covariances.}

\item{weights}{The class weights.}

\item{names}{The covariate names.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function plots an estimated bivariate contour mixing distributions.
}
\examples{
means <- list(c(0, 0), c(1, 1))
covs <- list(diag(2), 0.5 * diag(2))
weights <- c(0.3, 0.7)
names <- c("A", "B")
plot_mixture_contour(means, covs, weights, names)

}
