% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixture_contour.R
\name{plot_mixture_contour}
\alias{plot_mixture_contour}
\title{Plot bivariate contour of Gaussian mixture}
\usage{
plot_mixture_contour(means, covs, weights, names = c("1", "2"))
}
\arguments{
\item{means}{[\code{list(C)}]\cr
The class means. Each list element must be a vector of length 2.}

\item{covs}{[\code{list(C)}]\cr
The class covariances. Each list element must be covariance matrix of
dimension 2.}

\item{weights}{[\code{numeric(C)}]\cr
The class weights.}

\item{names}{[\code{character(2)}]\cr
The covariate names.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plot bivariate contour of Gaussian mixture
}
