% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mixture_marginal}
\alias{plot_mixture_marginal}
\title{Plot marginal mixing distributions}
\usage{
plot_mixture_marginal(mean, cov, weights, name)
}
\arguments{
\item{mean}{The class means.}

\item{cov}{The class covariances.}

\item{weights}{The class weights.}

\item{name}{The covariate name.}
}
\value{
An object of class \code{ggplot}.
}
\description{
This function plots an estimated marginal mixing distributions.
}
\examples{
mean <- list(1, 2)
cov <- list(0.1, 1)
weights <- c(0.3, 0.7)
name <- "test"
RprobitB:::plot_mixture_marginal(mean, cov, weights, name)

}
\keyword{internal}
