% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Visualizing the trace of Gibbs samples.}
\usage{
plot_trace(gibbs_samples, par_labels)
}
\arguments{
\item{gibbs_samples}{A matrix of Gibbs samples.}

\item{par_labels}{A character vector of length equal to the number of columns of
\code{gibbs_samples}, containing labels for the parameters.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots traces of the Gibbs samples.
}
