% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{posterior_pars}
\alias{posterior_pars}
\title{Parameter sets from posterior samples}
\usage{
posterior_pars(x, S)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{S}{The size of parameter sets.}
}
\value{
A list of \code{RprobitB_parameter}-objects.
}
\description{
This function builds parameter sets based on posterior samples.
}
\keyword{internal}
