% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_pars.R
\name{posterior_pars}
\alias{posterior_pars}
\title{Parameter sets from posterior samples}
\usage{
posterior_pars(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}
}
\value{
A list of \code{RprobitB_parameter} objects.
}
\description{
This function builds parameter sets from the normalized, burned and
thinned posterior samples.
}
\keyword{internal}
