% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprint.R
\name{pprint}
\alias{pprint}
\title{Print abbreviated vectors and matrices}
\usage{
pprint(
  x,
  rowdots = 4,
  coldots = 4,
  digits = 4,
  label = NULL,
  details = TRUE,
  onerow = FALSE
)
}
\arguments{
\item{x}{A (numeric or character) vector or matrix.}

\item{rowdots}{The row number which is replaced by \code{...}.}

\item{coldots}{The column number which is replaced by \code{...}.}

\item{digits}{The number of decimal places, if \code{x} is numeric.}

\item{label}{A character, a label for \code{x}.
Only printed if \code{details = TRUE} and \code{onerow = FALSE}.}

\item{details}{A boolean, set to \code{TRUE} (default) to print the name and the dimension
of \code{x}.}

\item{onerow}{A boolean, set to \code{TRUE} to print \code{x} in a single row.}
}
\value{
Invisibly returns \code{x}.
}
\description{
This function prints abbreviated vectors and matrices.
}
\examples{
pprint(x = 1, label = "single numeric")
pprint(x = LETTERS[1:26], label = "letters")
pprint(x = 1:3, coldots = 2)
pprint(x = matrix(rnorm(100), ncol = 1), label = "single column matrix")
pprint(x = matrix(1:100, nrow = 1), label = "single row matrix")
pprint(x = matrix(LETTERS[1:24], ncol = 6), label = "big matrix")
pprint(x = diag(5), coldots = 2, rowdots = 2, onerow = TRUE)

}
\references{
This function is a modified version of \code{\link[ramify]{pprint}}.
}
\keyword{internal}
\keyword{utils}
