% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict choices.}
\usage{
predict(x, overview = TRUE)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}

\item{overview}{If \code{TRUE}, aggregate the actual and predicted choices in a table.}
}
\value{
Either a table if \code{overview = TRUE} or a data frame otherwise.
}
\description{
This function predicts the choices of decision makers.
}
