% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict choices.}
\usage{
predict(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}
}
\value{
value
}
\description{
This function predicts the choices of decision makers based on a fitted
\code{RprobitB_model} and choice characteristics.
}
\details{
For more details see the vignette "Prediction":
\code{vignette("prediction", package = "RprobitB")}.
}
