% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.RprobitB_fit.R
\name{predict.RprobitB_fit}
\alias{predict.RprobitB_fit}
\title{Predict choices}
\usage{
\method{predict}{RprobitB_fit}(object, data = NULL, overview = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{RprobitB_fit}.}

\item{data}{Either \code{NULL} or an object of class \code{RprobitB_data}.}

\item{overview}{If \code{TRUE}, aggregate the prediction in a table.}

\item{...}{Ignored.}
}
\value{
Either a table if \code{overview = TRUE} or a data frame otherwise.
}
\description{
This function predicts the choices of decision makers.
}
\examples{
data <- simulate_choices(form = choice ~ covariate, N = 10, T = 10, J = 2)
object <- mcmc(data)
predict(object, overview = TRUE)
predict(object, overview = FALSE)

}
