% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R, R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare choice data for estimation}
\usage{
prepare_data()

prepare_data()
}
\arguments{
\item{choice_data}{A \code{data.frame} of choice data in wide format, i.e. each row represents
one choice occasion.}

\item{id}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each decision maker. The default is \code{"id"}.}

\item{idc}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each choice situation of each decision maker.
Per default, these identifier are generated by the order of appearance.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function prepares choice data for estimation.

This function prepares empirical choice data for model fitting, see details.
}
\details{
Requirements for the \code{data.frame} \code{choice_data}:
\itemize{
\item It \strong{must} contain a column named \code{id} which contains unique
identifier for each decision maker.
\item It \strong{can} contain a column named \code{idc} which contains unique
identifier for each choice situation of each decision maker.
If this information is missing, these identifier are generated
automatically by the appearance of the choices in the data set.
\item It \strong{can} contain a column named \code{choice} with the observed
choices, where \code{choice} must match the name of the dependent
variable in \code{form}.
Such a column is required for model fitting but not for prediction.
\item It \strong{must} contain a numeric column named \emph{p_j} for each alternative
specific covariate \emph{p} in \code{form} and each choice alternative \emph{j}
in \code{alternatives}.
\item It \strong{must} contain a numeric column named \emph{q} for each covariate \emph{q}
in \code{form} that is constant across alternatives.
}

In the ordered case (\code{ordered = TRUE}), the column \code{choice} must
contain the full ranking of the alternatives in each choice occasion as a
character, where the alternatives are separated by commas, see the examples.

See \href{https://loelschlaeger.de/RprobitB/articles/v02_choice_data.html}{the vignette on choice data}
for more details.
}
\section{Choice data}{
TODO
}

\examples{
data("Train", package = "mlogit")
data <- prepare_data(
  form = choice ~ price + time + comfort + change | 0,
  choice_data = Train,
  re = c("price", "time"),
  id = "id",
  idc = "choiceid",
  standardize = c("price", "time")
)

### ranked case
choice_data <- data.frame(
  "id" = 1:3, "choice" = c("A,B,C","A,C,B","B,C,A"), "cov" = 1
)
data <- prepare_data(
  form = choice ~ 0 | cov + 0,
  choice_data = choice_data,
  ranked = TRUE
)

}
\seealso{
\itemize{
\item \code{\link[=check_form]{check_form()}} for checking the model formula
\item \code{\link[=overview_effects]{overview_effects()}} for an overview of the model effects
\item \code{\link[=create_lagged_cov]{create_lagged_cov()}} for creating lagged covariates
\item \code{\link[=as_cov_names]{as_cov_names()}} for re-labeling alternative-specific covariates
\item \code{\link[=simulate_choices]{simulate_choices()}} for simulating choice data
\item \code{\link[=train_test]{train_test()}} for splitting choice data into a train and test subset
}
}
