% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare empirical choice data.}
\usage{
prepare_data(
  form,
  choice_data,
  re = NULL,
  alternatives = NULL,
  id = "id",
  idc = NULL,
  standardize = NULL
)
}
\arguments{
\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative).
They can be removed by adding \code{+0} in the second spot.
See the vignette \code{vignette("data_management", package = "RprobitB")}
for more details.}

\item{choice_data}{A data frame of choice data with the following requirements:
\itemize{
\item It \strong{must} be in "wide" format, i.e. each row represents one choice
occasion.
\item It \strong{must} contain a column named \code{id} which contains unique
identifier for each decision maker.
\item It \strong{can} contain a column named \code{idc} which contains unique
identifier for each choice situation of each decision maker.
If this information is missing, these identifier are generated
automatically by the appearance of the choices in the data set.
\item It \strong{can} contain a column named \code{choice} with the observed
choices, where \code{choice} must match the name of the dependent
variable in \code{form}.
Such a column is required for model fitting but not for prediction.
\item It \strong{must} contain a column named \emph{p_j} for each alternative
specific covariate \emph{p} in \code{form} and each choice alternative \emph{j}
in \code{alternatives}.
\item It \strong{must} contain a column named \emph{q} for each covariate \emph{q} in
\code{form} that is constant across alternatives.
}}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{id}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each decision maker. The default is \code{"id"}.}

\item{idc}{A character, the name of the column in \code{choice_data} that contains
unique identifier for each choice situation of each decision maker.
The default is \code{NULL}, in which case these identifier are generated
automatically.}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{<covariate>_<alternative>}.
If \code{standardize = "all"}, all covariates get standardized.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function prepares empirical choice data for the RprobitB package.
}
\details{
See the vignette "Choice data" for more details:
\code{vignette("choice_data", package = "RprobitB")}.
}
\examples{
data("Train", package = "mlogit")
data <- prepare_data(
  form = choice ~ price + time + comfort + change | 1,
  choice_data = Train,
  re = c("price", "time"),
  id = "id",
  idc = "choiceid"
)
}
