% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RprobitB_normalization.R
\name{print.RprobitB_normalization}
\alias{print.RprobitB_normalization}
\title{Print method for \code{RprobitB_normalization}.}
\usage{
\method{print}{RprobitB_normalization}(x, ...)
}
\arguments{
\item{x}{An object of class \code{RprobitB_normalization}.}

\item{...}{Ignored.}
}
\description{
This function is the print method for an object of class
\code{RprobitB_normalization}.
}
