% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.RprobitB_model}
\alias{print.summary.RprobitB_model}
\title{Print method for the summary of \code{RprobitB_model}}
\usage{
\method{print}{summary.RprobitB_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.RprobitB_model}}

\item{...}{ignored}
}
\description{
Print method for the summary of \code{RprobitB_model}
}
