% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Draw from multivariate normal distribution}
\usage{
rmvnorm(mean, Sigma, log = FALSE)
}
\arguments{
\item{mean}{A \code{numeric}, the mean vector of length \code{n}.}

\item{Sigma}{A \code{matrix}, the covariance matrix of dimension \code{n} x \code{n}.}

\item{log}{A \code{logical}, if \code{TRUE} the draw is taken from the log-normal
distribution.
By default, \code{log = FALSE}.}
}
\value{
A \code{numeric} of length \code{n}, the random draw.
}
\description{
This function draws from a multivariate normal distribution.
}
\examples{
mean <- c(0,0)
Sigma <- diag(2)
rmvnorm(mean = mean, Sigma = Sigma)
rmvnorm(mean = mean, Sigma = Sigma, log = TRUE)

}
\keyword{internal}
\keyword{utils}
