% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB-package.R
\docType{package}
\name{RprobitB-package}
\alias{RprobitB}
\alias{RprobitB-package}
\title{RprobitB: Bayesian Probit Choice Modeling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayes estimation of probit choice models, both in the cross-sectional and panel setting. The package can analyze binary, multivariate, ordered, and ranked choices, and places a special focus on modeling heterogeneity of choice behavior among deciders. The main functionality includes model fitting via Markov chain Monte Carlo methods, tools for convergence diagnostic, choice data simulation, in-sample and out-of-sample choice prediction, and model selection using information criteria and Bayes factors. The latent class model extension facilitates preference-based decider classification, where the number of latent classes can be inferred via the Dirichlet process or a weight-based updating scheme. This allows for flexible modeling of choice behavior without the need to impose structural constraints. For a reference on the method see Oelschlaeger and Bauer (2021) \url{https://trid.trb.org/view/1759753}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/RprobitB/}
  \item \url{https://github.com/loelschlaeger/RprobitB}
  \item Report bugs at \url{https://github.com/loelschlaeger/RprobitB/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Authors:
\itemize{
  \item Dietmar Bauer \email{dietmar.bauer@uni-bielefeld.de} (\href{https://orcid.org/0000-0003-2920-7032}{ORCID})
}

Other contributors:
\itemize{
  \item Sebastian Büscher \email{sebastian.buescher@uni-bielefeld.de} [contributor]
  \item Manuel Batram \email{manuel.batram@uni-bielefeld.de} [contributor]
}

}
\keyword{internal}
