% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_package.R
\docType{package}
\name{RprobitB}
\alias{RprobitB}
\title{RprobitB: A package for Bayes estimation of probit models}
\description{
This package provides tools for Bayes estimation of probit models.
}
\keyword{internal}
