% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_alternatives.R
\name{RprobitB_alternatives}
\alias{RprobitB_alternatives}
\alias{is.RprobitB_alternatives}
\alias{validate_RprobitB_alternatives}
\alias{print.RprobitB_alternatives}
\title{Define choice alternatives}
\usage{
RprobitB_alternatives(
  J,
  alternatives = LETTERS[1:J],
  base = alternatives[1],
  ordered = FALSE
)

is.RprobitB_alternatives(x)

validate_RprobitB_alternatives(x)

\method{print}{RprobitB_alternatives}(x, ...)
}
\arguments{
\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least 2.
If \code{ordered = TRUE}, must be at least 3.}

\item{alternatives}{A \code{character}, the vector of names of the choice alternatives.
Its length must be \code{J}.
By default, \code{alternatives = LETTERS[1:J]}.}

\item{base}{A \code{character}, the name of the base alternative for covariates that are
not alternative specific, see details.
\code{base} must be contained in \code{alternatives}.
Ignored if the model has no alternative specific covariates (e.g., in the
ordered probit case).
By default, \code{base} is the first element of \code{alternatives}.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{x}{An \code{RprobitB_alternatives} object.}
}
\value{
An \code{RprobitB_alternatives} object.
}
\description{
This function constructs an object of class \code{RprobitB_alternatives},
which contains the choice alternatives.
}
\section{Base alternative}{
The full collection of coefficients for covariates that are constant across
alternatives (including ASCs) is not identified, one coefficient is a linear
coefficient of the others, respectively. To achieve identifiability, the
coefficient of one alternative \code{base} is fixed to zero. The other
coefficients then have to be interpreted with respect to \code{base}.
}

\examples{
\dontrun{
RprobitB_alternatives(3)
}

}
\keyword{internal}
