% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_covariates.R
\name{RprobitB_covariates}
\alias{RprobitB_covariates}
\alias{is.RprobitB_covariates}
\alias{simulate_RprobitB_covariates}
\alias{validate_RprobitB_covariates}
\alias{print.RprobitB_covariates}
\title{Define choice covariates}
\usage{
RprobitB_covariates(RprobitB_data, RprobitB_formula, RprobitB_alternatives)

is.RprobitB_covariates(x)

simulate_RprobitB_covariates(
  formula,
  N,
  J,
  T = 1,
  alternatives = LETTERS[1:J],
  re = NULL,
  ordered = FALSE,
  seed = NULL,
  covariates = list(.default = function() stats::rnorm(n = 1, mean = 0, sd = 9))
)

validate_RprobitB_covariates(
  x = RprobitB_covariates(),
  formula,
  N,
  J,
  T = 1,
  alternatives = LETTERS[1:J],
  re = NULL,
  ordered = FALSE,
  base = alternatives[1]
)

\method{print}{RprobitB_covariates}(x, ...)
}
\arguments{
\item{RprobitB_data}{An \code{\link{RprobitB_data}} object.}

\item{RprobitB_formula}{An \code{\link{RprobitB_formula}} object.}

\item{RprobitB_alternatives}{An \code{\link{RprobitB_alternatives}} object.}

\item{x}{An \code{\link{RprobitB_covariates}} object.}

\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the model to be
fitted, see details.}

\item{N}{A positive \code{integer}, the number of deciders.}

\item{J}{An \code{integer}, the number of choice alternatives.
Must be at least \code{2}.
If \code{ordered = TRUE}, must be at least \code{3}.}

\item{T}{A positive \code{integer}, the number of choice occasions per decider.
Can also be a \code{vector} of length \code{N} for a variable number of
choice occasions per decider.
By default, \code{T = 1}.}

\item{alternatives}{A \code{character}, the vector of names of the choice alternatives.
Its length must be \code{J}.
By default, \code{alternatives = LETTERS[1:J]}.}

\item{re}{A \code{character}, the vector of names of covariates in \code{formula} that
should have a random effect, see details.
By default, \code{re = NULL}, i.e., no random effects.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{covariates}{A named \code{list}, each element is a \code{function} that draws covariates,
see the details.
By default, \code{sampler = list()}.
Covariates that are not specified in \code{sampler} are drawn
via \code{"default" = function() stats::rnorm(n = 1, mean = 0, sd = 9)}.}

\item{base}{A \code{character}, the name of the base alternative for covariates that are
not alternative specific, see details.
\code{base} must be contained in \code{alternatives}.
Ignored if the model has no alternative specific covariates (e.g., in the
ordered probit case).
By default, \code{base} is the first element of \code{alternatives}.}

\item{...}{Currently not used.}
}
\value{
An \code{\link{RprobitB_covariates}} object.

It contains the elements:
\describe{
\item{\code{}}{}
}
}
\description{
These functions create and validate an object of class
\code{\link{RprobitB_covariates}}, which contains the covariate matrices,
see the details.

\code{\link{simulate_RprobitB_covariates}} simulates covariates.
}
\section{Covariate matrices}{
TODO
}

\section{Sample covariates}{
TODO
}

\keyword{object}
