% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_data.R
\name{RprobitB_data}
\alias{RprobitB_data}
\title{Construct object of class \code{RprobitB_data}}
\usage{
RprobitB_data(
  data,
  choice_data,
  N,
  T,
  J,
  P_f,
  P_r,
  alternatives,
  form,
  re,
  vars,
  ASC,
  standardize,
  simulated,
  parm,
  distr
)
}
\arguments{
\item{data}{A list with the choice data.
The list has \code{N} elements.
Each element is a list with two elements, \code{X} and \code{y}, which are
the covariates and decisions for a decision maker. More precisely,
\code{X} is a list of \code{T} elements, where each element is a matrix of
dimension \code{J}x(\code{P_f}+\code{P_r}) and contains the characteristics
for one choice occasion.
\code{y} is a vector of length \code{T} and contains the labels for the
chosen alternatives.}

\item{choice_data}{A data frame of choice data in "wide" format (i.e. each row represents
one choice occasion) with the following requirements:
\itemize{
\item It must contain columns named \code{id} (a unique identifier for each
decision maker) and \code{choice} (the chosen alternatives), where
\code{choice} is the name of the dependent variable in \code{form}.
\item For each alternative specific covariate \emph{p} (covariate of type 1 or
3) in \code{form} and each choice alternative \emph{j}, \code{choice_data}
must contain a column named \emph{p_j}.
\item For each covariate \emph{q} that is constant across covariates (covariate
of type 2), \code{choice_data} must contain a column named \emph{q}.
}}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{alternatives}{A character vector with the names of the choice alternatives.}

\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative).
They can be removed by adding \code{+0} in the second spot.
See the vignette \code{vignette("data_management", package = "RprobitB")}
for more details.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{vars}{The element \code{vars} in the output of \link{check_form}.}

\item{ASC}{A boolean, determining whether the model has ASCs.}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{covariate_alternative}.
If \code{standardize = "all"}, all covariates get standardized.}

\item{simulated}{A boolean, if \code{TRUE} then \code{data} is simulated, otherwise
\code{data} is empirical.}

\item{parm}{A named list of true parameter values:
\itemize{
\item \code{alpha}:
The fixed coefficient vector of length \code{P_f}.
\item \code{C:}
The number (greater or equal 1) of latent classes of decision makers.
If \code{P_r = 0}, then \code{C} is ignored.
\item \code{s}:
The vector of class weights of length \code{C}.
\item \code{b}:
The matrix of class means as columns of dimension \code{P_r} x \code{C}.
\item \code{Omega}:
The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
\item \code{Sigma}:
The error term covariance matrix of dimension \code{J} x \code{J}.
Internally, \code{Sigma} gets differenced with respect to alternative
\code{J}, so it becomes a covariance matrix of dimension
\code{J-1} x \code{J-1}.
}}

\item{distr}{A named list of number generation functions from which the covariates are
drawn. Each element of \code{distr} must be of the form
\code{"cov" = list("name" = "<name of the number generation function>", ...)},
where \code{cov} is the name of the covariate and \code{...} are required
parameters for the number generation function.
Covariates for which no distribution is specified are drawn from a standard
normal distribution.
Possible number generation functions are
\itemize{
\item functions of the type \code{r*} from base R (e.g. \code{rnorm}) where
all required parameters (except for \code{n}) must be specified,
\item the function \code{sample}, where all required parameters
(except for \code{size}) must be specified.
}}
}
\value{
An object of class \code{RprobitB_data}, which is a list of
\itemize{
\item the arguments of this functions as elements, and
\item the element \code{covs}, which is a data frame of the two
columns \code{name} (the covariate names) and \code{random} (a boolean
determining whether the covariates are random effects).
}

list of the arguments
of this function and the element \code{covs}, which is a data frame
containing the parameter names.
In case of empirical data, the elements \code{parm} and \code{distr} equal
\code{NULL}.
}
\description{
Function that constructs an object of class \code{RprobitB_data}.
}
