% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_data.R, R/data_management.R
\name{RprobitB_data}
\alias{RprobitB_data}
\alias{is.RprobitB_data}
\alias{summary.RprobitB_data}
\alias{print.RprobitB_data}
\alias{plot.RprobitB_data}
\title{Define choice data for probit model}
\usage{
RprobitB_data(
  data,
  choice_data,
  N,
  T,
  J,
  P_f,
  P_r,
  alternatives,
  ordered,
  ranked,
  base,
  form,
  re,
  ASC,
  effects,
  standardize,
  simulated,
  choice_available,
  true_parameter,
  res_var_names
)

is.RprobitB_data(x)

\method{summary}{RprobitB_data}(object, ...)

\method{print}{RprobitB_data}(x, ...)

\method{plot}{RprobitB_data}(x, by_choice = FALSE, alpha = 1, position = "dodge", ...)

RprobitB_data(
  data,
  choice_data,
  N,
  T,
  J,
  P_f,
  P_r,
  alternatives,
  ordered,
  ranked,
  base,
  form,
  re,
  ASC,
  effects,
  standardize,
  simulated,
  choice_available,
  true_parameter,
  res_var_names
)
}
\arguments{
\item{data}{A list with the choice data.
The list has \code{N} elements.
Each element is a list with two elements, \code{X} and \code{y}, which are
the covariates and decisions for a decision maker. More precisely,
\code{X} is a list of \code{T} elements, where each element is a matrix of
dimension \code{J}x(\code{P_f}+\code{P_r}) and contains the characteristics
for one choice occasion.
\code{y} is a vector of length \code{T} and contains the labels for the
chosen alternatives.}

\item{choice_data}{A \code{data.frame} of choice data in wide format, i.e. each row represents
one choice occasion.}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).
Per default, \code{T = 1}.}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{P_f}{The number of covariates connected to a fixed coefficient (can be 0).}

\item{P_r}{The number of covariates connected to a random coefficient (can be 0).}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.
If \code{ordered = TRUE}, \code{alternatives} is assumed to be specified with
the alternatives ordered from worst to best.}

\item{ordered}{A boolean, \code{FALSE} per default. If \code{TRUE}, the choice set
\code{alternatives} is assumed to be ordered from worst to best.}

\item{ranked}{TBA}

\item{base}{A character, the name of the base alternative for covariates that are not
alternative specific (i.e. type 2 covariates and ASCs). Ignored and set to
\code{NULL} if the model has no alternative specific covariates (e.g. in the
ordered probit model).
Per default, \code{base} is the last element of \code{alternatives}.}

\item{ASC}{A boolean, determining whether the model has ASCs.}

\item{effects}{A data frame with the effect names and booleans indicating whether
they are connected to random effects.}

\item{standardize}{A character vector of names of covariates that get standardized.
Covariates of type 1 or 3 have to be addressed by
\code{<covariate>_<alternative>}.
If \code{standardize = "all"}, all covariates get standardized.}

\item{simulated}{A boolean, if \code{TRUE} then \code{data} is simulated, otherwise
\code{data} is empirical.}

\item{choice_available}{A boolean, if \code{TRUE} then \code{data} contains observed choices.}

\item{true_parameter}{An object of class \code{RprobitB_parameters}.}

\item{res_var_names}{A names list of reserved variable names in \code{choice_data}.}

\item{x}{An \code{RprobitB_data} object.}

\item{object}{An \code{RprobitB_data} object.}

\item{...}{Not used.}
}
\value{
An \code{RprobitB_data} object.

An object of class \code{RprobitB_data} with the arguments of this function
as elements.
}
\description{
These functions create, validate, summarize and visualize an object of class
\code{RprobitB_data}, which contains the choice data for a probit model,
see details.

This function constructs an object of class \code{RprobitB_data}.
}
\section{Choice simulation}{
TODO
}

\section{Choice data}{
TODO
}

\keyword{internal}
