% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_latent_classes.R
\name{RprobitB_latent_classes}
\alias{RprobitB_latent_classes}
\title{Create object of class \code{RprobitB_latent_classes}.}
\usage{
RprobitB_latent_classes(latent_classes = NULL)
}
\arguments{
\item{latent_classes}{Either \code{NULL} or a list of parameters specifying the number and the
latent classes:
\itemize{
\item \code{C}:
The number (greater or equal 1) of latent classes, which is set to 1 per
default and is ignored if \code{P_r = 0}.
If \code{update = TRUE}, \code{C} equals the initial number of classes.
\item \code{update}:
A boolean, determining whether to update \code{C}.
Ignored if \code{P_r = 0}.
If \code{update = FALSE}, all of the following elements are ignored.
\item \code{Cmax}:
The maximum number of latent classes.
\item \code{buffer}:
The updating buffer (number of iterations to wait before the next update).
\item \code{epsmin}:
The threshold weight for removing latent classes (between 0 and 1).
\item \code{epsmax}:
The threshold weight for splitting latent classes (between 0 and 1).
\item \code{distmin}:
The threshold difference in means for joining latent classes
(non-negative).
}}
}
\value{
An object of class \code{RprobitB_latent_classes}.
}
\description{
This function creates an object of class \code{RprobitB_latent_classes}.
}
\keyword{constructor}
