% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{RprobitB_normalization}
\alias{RprobitB_normalization}
\title{Create object of class \code{RprobitB_normalization}}
\usage{
RprobitB_normalization(
  level,
  scale = Sigma_1 ~ 1,
  form,
  re = NULL,
  alternatives,
  base
)
}
\arguments{
\item{level}{The alternative name with respect to which utility differences are computed.
Currently, only differences with respect to the last alternative can be
computed.}

\item{scale}{A formula object which determines the utility scale. It is of the form
\verb{<parameter> ~ <value>}, where \verb{<parameter>} is either the name of a fixed
effect or \verb{Sigma_<j>} for the \verb{<j>}th diagonal element of \code{Sigma}, and
\verb{<value>} is the value of the fixed parameter.}

\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative due to identifiability).
They can be removed by adding \code{+0} in the second spot.
See the vignette on choice data for more details.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{base}{A character, the name of the base alternative for covariates that are not
alternative specific (type 2 covariates).}
}
\value{
An object of class \code{RprobitB_normalization}, which is a list of
\itemize{
\item \code{level}, a list with the elements \code{level} (the number of the
alternative specified by the input \code{level}) and \code{name} (the name of
the alternative, i.e. the input \code{level}),
\item and \code{scale}, a list with the elements \code{parameter} (either \code{"s"} for
an element of \code{Sigma} or \code{"a"}for an element of \code{alpha}), the
parameter \code{index}, and the fixed \code{value}. If \code{parameter = "a"}, also
the \code{name} of the fixed effect.
}
}
\description{
This function creates an object of class \code{RprobitB_normalization},
which determines the utility scale and level.
}
\details{
Any choice model has to be normalized with respect to the utility level and
scale.
\itemize{
\item For level normalization, {RprobitB} takes utility differences with
respect to one alternative.
\item For scale normalization, RprobitB fixes a model parameter. Per
default, the first error-term variance is fixed to \code{1}, i.e.
\code{scale = Sigma_1 ~ 1}. Alternatively, any error-term variance or
any non-random coefficient can be fixed.
}
}
\examples{
RprobitB:::RprobitB_normalization(
  level = "B",
  scale = price ~ -1,
  form = choice ~ price + time + comfort + change | 1,
  re = "time",
  alternatives = c("A", "B"),
  base = "A"
)

}
\keyword{internal}
