% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior}
\alias{RprobitB_prior}
\alias{is.RprobitB_prior}
\alias{print.RprobitB_prior}
\title{Define probit model prior}
\usage{
RprobitB_prior(formula, re = NULL, J, C = 1, ordered = FALSE, ...)

is.RprobitB_prior(x)

\method{print}{RprobitB_prior}(x, ...)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}}, a symbolic description of the model to be
fitted, see details.}

\item{re}{A \code{character}, the vector of names of covariates in \code{formula} that
should have a random effect, see details.
By default, \code{re = NULL}, i.e., no random effects.}

\item{ordered}{A \code{logical}, \code{TRUE} if the choice alternatives are ordered and
\code{FALSE} (default) else.}

\item{...}{
  Arguments passed on to \code{\link[=RprobitB_prior_alpha]{RprobitB_prior_alpha}}, \code{\link[=RprobitB_prior_s]{RprobitB_prior_s}}, \code{\link[=RprobitB_prior_b]{RprobitB_prior_b}}, \code{\link[=RprobitB_prior_Omega]{RprobitB_prior_Omega}}, \code{\link[=RprobitB_prior_Sigma]{RprobitB_prior_Sigma}}, \code{\link[=RprobitB_prior_Sigma_diff]{RprobitB_prior_Sigma_diff}}, \code{\link[=RprobitB_prior_d]{RprobitB_prior_d}}
  \describe{
    \item{\code{P_f}}{An \code{integer}, the number of fixed model effects.
Can be computed via the function \code{\link{compute_P_f}}.}
    \item{\code{alpha_prior_mean}}{A \code{numeric} of length \code{P_f}, the mean vector for the conjugate
normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_mean = numeric(P_f)}.}
    \item{\code{alpha_prior_Sigma}}{A \code{matrix} of dimension \code{P_f} x \code{P_f}, the covariance
matrix for the conjugate normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_Sigma = 10 * diag(P_f)}.}
    \item{\code{alpha_prior_custom}}{A \code{function}, a custom prior density function for \code{alpha},
see details.
By default, \code{alpha_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{alpha_prior_custom_test_par}}{A \code{numeric} of length \code{P_f}, a test input for the custom prior
density function \code{alpha_prior_custom}.
By default, \code{alpha_prior_custom_test_par = numeric(P_f)}.
Ignored if \code{alpha_prior_custom = NULL}.}
    \item{\code{s_prior_concentration}}{A \code{numeric} of length \code{C}, the concentration vector for the
conjugate Dirichlet prior distribution of \code{s}.
By default, \code{s_prior_concentration = rep(1, C)}.}
    \item{\code{s_prior_custom}}{A \code{function}, a custom prior density function for \code{s},
see details.
By default, \code{s_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{s_prior_custom_test_par}}{A \code{numeric} of length \code{C}, a test input for the custom prior
density function \code{s_prior_custom}.
By default, \code{s_prior_custom_test_par = numeric(C)}.
Ignored if \code{s_prior_custom = NULL}.}
    \item{\code{C}}{An \code{integer}, the number (greater or equal 1) of latent classes of
decision makers.
By default, \code{C = 1}.}
    \item{\code{P_r}}{An \code{integer}, the number of random model effects.
Can be computed via the function \code{\link{compute_P_r}}.}
    \item{\code{b_prior_mean}}{A \code{numeric} of length \code{P_r}, the mean vector for the conjugate
normal prior distribution of \code{b}.
By default, \code{b_prior_mean = numeric(P_r)}.}
    \item{\code{b_prior_Sigma}}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, the covariance
matrix for the conjugate normal prior distribution of \code{b}.
By default, \code{b_prior_Sigma = 10 * diag(P_r)}.}
    \item{\code{b_prior_custom}}{A \code{function}, a custom prior density function for \code{b},
see details.
By default, \code{b_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{b_prior_custom_test_par}}{A \code{numeric} of length \code{P_r}, a test input for the custom prior
density function \code{b_prior_custom}.
By default, \code{b_prior_custom_test_par = numeric(P_r)}.
Ignored if \code{b_prior_custom = NULL}.}
    \item{\code{Omega_prior_df}}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Omega}.
It must be greater or equal than \code{P_r + 2} (to ensure a proper prior).
By default, \code{Omega_prior_df = P_r + 2}.}
    \item{\code{Omega_prior_scale}}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of \code{Omega}.
By default, \code{Omega_prior_scale = diag(P_r)}.}
    \item{\code{Omega_prior_custom}}{A \code{function}, a custom prior density function for \code{Omega},
see details.
By default, \code{Omega_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{Omega_prior_custom_test_par}}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, a test input for the
custom prior density function \code{Omega_prior_custom}.
By default, \code{Omega_prior_custom_test_par = diag(P_r)}.
Ignored if \code{Omega_prior_custom = NULL}.}
    \item{\code{Sigma_prior_df}}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Sigma}.
It must be greater or equal than \code{3} (to ensure a proper prior).
By default, \code{Sigma_prior_df = 3}.}
    \item{\code{Sigma_prior_scale}}{A \code{matrix} of dimension \code{1} x \code{1}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of
\code{Sigma}.
Can also be a single \code{numeric}.
By default, \code{Sigma_prior_scale = 1}.}
    \item{\code{Sigma_prior_custom}}{A \code{function}, a custom prior density function for \code{Sigma},
see details.
By default, \code{Sigma_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{Sigma_prior_custom_test_par}}{A \code{matrix} of dimension \code{1} x \code{1}, a test input for
the custom prior density function \code{Sigma_prior_custom}.
Can also be a single \code{numeric}.
By default, \code{Sigma_prior_custom_test_par = 1}.
Ignored if \code{Sigma_prior_custom = NULL}.}
    \item{\code{J}}{An \code{integer}, the number of choice alternatives.
Must be at least 2.
If \code{ordered = TRUE}, must be at least 3.}
    \item{\code{Sigma_diff_prior_df}}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Sigma_diff}.
It must be greater or equal than \code{J + 1} (to ensure a proper prior).
By default, \code{Sigma_diff_prior_df = J + 1}.}
    \item{\code{Sigma_diff_prior_scale}}{A \code{matrix} of dimension \code{(J-1)} x \code{(J-1)}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of
\code{Sigma_diff}.
By default, \code{Sigma_diff_prior_scale = diag(J-1)}.}
    \item{\code{Sigma_diff_prior_custom}}{A \code{function}, a custom prior density function for \code{Sigma_diff},
see details.
By default, \code{Sigma_diff_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{Sigma_diff_prior_custom_test_par}}{A \code{matrix} of dimension \code{(J-1)} x \code{(J-1)}, a test input for
the custom prior density function \code{Sigma_diff_prior_custom}.
By default, \code{Sigma_diff_prior_custom_test_par = diag(J-1)}.
Ignored if \code{Sigma_diff_prior_custom = NULL}.}
    \item{\code{d_prior_mean}}{A \code{numeric} of length \code{J-2}, the mean vector for the conjugate
normal prior distribution of \code{d}.
By default, \code{d_prior_mean = numeric(J-2)}.}
    \item{\code{d_prior_Sigma}}{A \code{matrix} of dimension \code{(J-2)} x \code{(J-2)}, the covariance
matrix for the conjugate normal prior distribution of \code{d}.
By default, \code{d_prior_Sigma = 10 * diag(J-2)}.}
    \item{\code{d_prior_custom}}{A \code{function}, a custom prior density function for \code{d},
see details.
By default, \code{d_prior_custom = NULL}, i.e., no custom prior.}
    \item{\code{d_prior_custom_test_par}}{A \code{numeric} of length \code{J-2}, a test input for the custom prior
density function \code{d_prior_custom}.
By default, \code{d_prior_custom_test_par = numeric(J-2)}.
Ignored if \code{d_prior_custom = NULL}.}
  }}

\item{x}{An \code{RprobitB_prior} object.}
}
\value{
An \code{RprobitB_prior} object.
}
\description{
This function defines an object of class \code{RprobitB_prior}, which
contains the prior specification for a probit model.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item Only if \code{P_f > 0}:
\code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item Only if \code{P_r > 0}:
\code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item Only if \code{P_r > 0}:
\code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item Only in the unordered case (\code{ordered = FALSE}):
\code{Sigma_diff ~ Inverse-Wishart(df = Sigma_diff_prior_df, scale = Sigma_diff_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha} (only if \code{P_f > 0})
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b} (only if \code{P_r > 0})
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega} (only if \code{P_r > 0})
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma} (only if \code{ordered = TRUE})
\itemize{
\item \code{Sigma_prior_df = 3}
\item \code{Sigma_prior_scale = 1}
}
\item \code{Sigma_diff} (only if \code{ordered = FALSE})
\itemize{
\item \code{Sigma_diff_prior_df = J + 1}
\item \code{Sigma_diff_prior_scale = diag(J-1)}
}
\item \code{d} (only if \code{ordered = TRUE})
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are applied, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in {RprobitB}) can be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\section{Model formula}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need any
separation via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{
Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (for sign-restriction) can be specified via appending \code{+}
to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
### default conjugate prior distributions
RprobitB_prior(formula = choice ~ A, J = 2)

### conjugate prior distributions with custom parameters
RprobitB_prior(
  formula = choice ~ A | B + 0, re = "A", J = 4, C = 2,
  alpha_prior_mean = c(-10, 0, 10),
  s_prior_concentration = c(1, 2),
  b_prior_Sigma = matrix(2)
)

### custom prior distributions
RprobitB_prior(
  formula = choice ~ A | B | C, re = "C", J = 3,
  alpha_prior_custom = function(x) {
    stats::dnorm(x[1], mean = 0, sd = 1) * dunif(x[2], min = 0, max = 1)
  }
)

}
\keyword{object}
