% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_alpha}
\alias{RprobitB_prior_alpha}
\alias{is.RprobitB_prior_alpha}
\alias{print.RprobitB_prior_alpha}
\title{Define \code{alpha} prior}
\usage{
RprobitB_prior_alpha(
  P_f,
  alpha_prior_mean = numeric(P_f),
  alpha_prior_Sigma = 10 * diag(P_f),
  alpha_prior_custom = NA,
  alpha_prior_custom_test_par = numeric(P_f)
)

is.RprobitB_prior_alpha(x)

\method{print}{RprobitB_prior_alpha}(x, ...)
}
\arguments{
\item{P_f}{An \code{integer}, the number of fixed model effects.
Can be computed via the function \code{\link{P_f}}.}

\item{alpha_prior_mean}{A \code{numeric} of length \code{P_f}, the mean vector for the conjugate
normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_mean = numeric(P_f)}.}

\item{alpha_prior_Sigma}{A \code{matrix} of dimension \code{P_f} x \code{P_f}, the covariance
matrix for the conjugate normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_Sigma = 10 * diag(P_f)}.}

\item{alpha_prior_custom}{A \code{function}, a custom prior density function for \code{alpha},
see details.
By default, \code{alpha_prior_custom = NULL}, i.e., no custom prior.}

\item{alpha_prior_custom_test_par}{A \code{numeric} of length \code{P_f}, a test input for the custom prior
density function \code{alpha_prior_custom}.
By default, \code{alpha_prior_custom_test_par = numeric(P_f)}.
Ignored if \code{alpha_prior_custom = NULL}.}

\item{x}{An \code{RprobitB_prior_alpha} object.}

\item{...}{Not used.}
}
\value{
An \code{RprobitB_prior_alpha} object or \code{NA} if \code{P_f = 0}.
}
\description{
This function defines the prior distributions for the probit model parameter
\code{alpha}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item \code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item \code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item \code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item \code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item \code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha}
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b}
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega}
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma}
\itemize{
\item \code{Sigma_prior_df = if (ordered) 4 else J + 1}
\item \code{Sigma_prior_scale = if (ordered) diag(1) else diag(J-1)}
}
\item \code{d}
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are used, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
}
}

\examples{
### conjugate prior: alpha ~ Normal(0, 10*I)
RprobitB_prior_alpha(
  P_f = 2,
  alpha_prior_mean = numeric(P_f),
  alpha_prior_Sigma = 10 * diag(P_f)
)

### custom prior: alpha_1 ~ Normal(0,1), alpha_2 ~ Uniform(0,1)
RprobitB_prior_alpha(
  P_f = 2,
  alpha_prior_custom = function(x) {
    dnorm(x[1], mean = 0, sd = 1) * dunif(x[2], min = 0, max = 1)
  }
)

}
\keyword{internal}
