% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_alpha}
\alias{RprobitB_prior_alpha}
\alias{is.RprobitB_prior_alpha}
\alias{print.RprobitB_prior_alpha}
\title{Define \code{alpha} prior}
\usage{
RprobitB_prior_alpha(
  P_f,
  alpha_prior_mean = numeric(P_f),
  alpha_prior_Sigma = 10 * diag(P_f),
  alpha_prior_custom = NA,
  alpha_prior_custom_test_par = numeric(P_f)
)

is.RprobitB_prior_alpha(x)

\method{print}{RprobitB_prior_alpha}(
  x,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = TRUE,
  details = FALSE,
  ...
)
}
\arguments{
\item{P_f}{An \code{integer}, the number of fixed model effects.
Can be computed via the function \code{\link{compute_P_f}}.}

\item{alpha_prior_mean}{A \code{numeric} of length \code{P_f}, the mean vector for the conjugate
normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_mean = numeric(P_f)}.}

\item{alpha_prior_Sigma}{A \code{matrix} of dimension \code{P_f} x \code{P_f}, the covariance
matrix for the conjugate normal prior distribution of \code{alpha}.
By default, \code{alpha_prior_Sigma = 10 * diag(P_f)}.}

\item{alpha_prior_custom}{A \code{function}, a custom prior density function for \code{alpha},
see details.
By default, \code{alpha_prior_custom = NULL}, i.e., no custom prior.}

\item{alpha_prior_custom_test_par}{A \code{numeric} of length \code{P_f}, a test input for the custom prior
density function \code{alpha_prior_custom}.
By default, \code{alpha_prior_custom_test_par = numeric(P_f)}.
Ignored if \code{alpha_prior_custom = NULL}.}

\item{x}{An \code{\link{RprobitB_prior_alpha}} object.}

\item{rowdots}{An \code{integer}, the row number which is replaced by \code{...}.
By default, \code{rowdots = 4}.}

\item{coldots}{An \code{integer}, the column number which is replaced by \code{...}.
By default, \code{coldots = 4}.}

\item{digits}{An \code{integer}, the number of printed decimal places.
Only relevant if input \code{x} is numeric.
By default, \code{digits = 2}.}

\item{simplify}{A \code{logical}, set to \code{TRUE} to simplify the output.
By default, \code{simplify = FALSE}.}

\item{details}{A \code{logical}, set to \code{TRUE} to print the type and
dimension of \code{x}.
By default, \code{details = !simplify}.}

\item{...}{Currently not used.}
}
\value{
An \code{\link{RprobitB_prior_alpha}} object or \code{NA} if \code{P_f = 0}.

It contains the elements:
\describe{
\item{\code{conjugate}}{Is the prior conjugated?}
\item{\code{alpha_prior_mean}}{The prior mean vector.}
\item{\code{alpha_prior_Sigma}}{The prior covariance matrix.}
\item{\code{alpha_prior_custom}}{The custom prior function.}
}
}
\description{
This function defines the prior distributions for the probit model parameter
\code{alpha}. Only relevant if \code{P_f > 0}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item Only if \code{P_f > 0}:
\code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item Only if \code{P_r > 0}:
\code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item Only if \code{P_r > 0}:
\code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item Only in the unordered case (\code{ordered = FALSE}):
\code{Sigma_diff ~ Inverse-Wishart(df = Sigma_diff_prior_df, scale = Sigma_diff_prior_scale)}
\item Only in the ordered case (\code{ordered = TRUE}):
\code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha} (only if \code{P_f > 0})
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b} (only if \code{P_r > 0})
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega} (only if \code{P_r > 0})
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma} (only if \code{ordered = TRUE})
\itemize{
\item \code{Sigma_prior_df = 3}
\item \code{Sigma_prior_scale = 1}
}
\item \code{Sigma_diff} (only if \code{ordered = FALSE})
\itemize{
\item \code{Sigma_diff_prior_df = J + 1}
\item \code{Sigma_diff_prior_scale = diag(J-1)}
}
\item \code{d} (only if \code{ordered = TRUE})
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are applied, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in \{RprobitB\}) can be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\examples{
\dontrun{
### conjugate prior: alpha ~ Normal(0, 10*I)
RprobitB_prior_alpha(P_f = 2)

### custom prior: alpha_1 ~ Normal(0,1), alpha_2 ~ Uniform(0,1)
RprobitB_prior_alpha(
  P_f = 2,
  alpha_prior_custom = function(x) {
    stats::dnorm(x[1]) * stats::dunif(x[2])
  }
)
}

}
\keyword{internal}
\keyword{object}
