% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RprobitB_prior.R
\name{RprobitB_prior_Omega}
\alias{RprobitB_prior_Omega}
\alias{is.RprobitB_prior_Omega}
\alias{print.RprobitB_prior_Omega}
\title{Define \code{Omega} prior}
\usage{
RprobitB_prior_Omega(
  P_r,
  Omega_prior_df = P_r + 2,
  Omega_prior_scale = diag(P_r),
  Omega_prior_custom = NA,
  Omega_prior_custom_test_par = diag(P_r)
)

is.RprobitB_prior_Omega(x)

\method{print}{RprobitB_prior_Omega}(
  x,
  rowdots = 4,
  coldots = 4,
  digits = 2,
  simplify = TRUE,
  details = FALSE,
  ...
)
}
\arguments{
\item{P_r}{An \code{integer}, the number of random model effects.
Can be computed via the function \code{\link{compute_P_r}}.}

\item{Omega_prior_df}{An \code{integer}, the degrees of freedom for the conjugate
Inverse-Wishart prior distribution of \code{Omega}.
It must be greater or equal than \code{P_r + 2} (to ensure a proper prior).
By default, \code{Omega_prior_df = P_r + 2}.}

\item{Omega_prior_scale}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, the scale
matrix for the conjugate Inverse-Wishart prior distribution of \code{Omega}.
By default, \code{Omega_prior_scale = diag(P_r)}.}

\item{Omega_prior_custom}{A \code{function}, a custom prior density function for \code{Omega},
see details.
By default, \code{Omega_prior_custom = NULL}, i.e., no custom prior.}

\item{Omega_prior_custom_test_par}{A \code{matrix} of dimension \code{P_r} x \code{P_r}, a test input for the
custom prior density function \code{Omega_prior_custom}.
By default, \code{Omega_prior_custom_test_par = diag(P_r)}.
Ignored if \code{Omega_prior_custom = NULL}.}

\item{x}{An \code{RprobitB_prior_Omega} object.}

\item{rowdots}{An \code{integer}, the row number which is replaced by \code{...}.
By default, \code{rowdots = 4}.}

\item{coldots}{An \code{integer}, the column number which is replaced by \code{...}.
By default, \code{coldots = 4}.}

\item{digits}{An \code{integer}, the number of printed decimal places.
Only relevant if input \code{x} is numeric.
By default, \code{digits = 2}.}

\item{simplify}{A \code{logical}, set to \code{TRUE} to simplify the output.
By default, \code{simplify = FALSE}.}

\item{details}{A \code{logical}, set to \code{TRUE} to print the type and
dimension of \code{x}.
By default, \code{details = !simplify}.}

\item{...}{Not used.}
}
\value{
An \code{RprobitB_prior_Omega} object or \code{NA} if \code{P_r = 0}.
}
\description{
This function defines the prior distributions for the probit model parameter
\code{Omega}.
}
\section{Model priors}{
\subsection{Conjugate priors}{

A priori, the following
\href{https://en.wikipedia.org/wiki/Conjugate_prior}{conjugate prior distributions}
are assumed for the model parameters:
\itemize{
\item \code{alpha ~ Normal(mean = alpha_prior_mean, Sigma = alpha_prior_Sigma)}
\item \code{s ~ Dirichlet(concentration = s_prior_concentration)}
\item \code{b ~ Normal(mean = b_prior_mean, Sigma = b_prior_Sigma)}
\item \code{Omega ~ Inverse-Wishart(df = Omega_prior_df, scale = Omega_prior_scale)}
\item \code{Sigma ~ Inverse-Wishart(df = Sigma_prior_df, scale = Sigma_prior_scale)}
\item \code{d ~ Normal(mean = d_prior_mean, Sigma = d_prior_Sigma)}
}
}

\subsection{Default parameters for the conjugate priors}{

By default, the following parameters are set for the conjugate priors:
\itemize{
\item \code{alpha}
\itemize{
\item \code{alpha_prior_mean = numeric(P_f)}
\item \code{alpha_prior_Sigma = 10*diag(P_f)}
}
\item \code{s}
\itemize{
\item \code{s_prior_concentration = rep(1, C)}
}
\item \code{b}
\itemize{
\item \code{b_prior_mean = numeric(P_r)}
\item \code{b_prior_Sigma = 10*diag(P_r)}
}
\item \code{Omega}
\itemize{
\item \code{Omega_prior_df = P_r + 2}
\item \code{Omega_prior_scale = diag(P_r)}
}
\item \code{Sigma}
\itemize{
\item \code{Sigma_prior_df = if (ordered) 4 else J + 1}
\item \code{Sigma_prior_scale = if (ordered) diag(1) else diag(J-1)}
}
\item \code{d}
\itemize{
\item \code{d_prior_mean = numeric(J - 2)}
\item \code{d_prior_Sigma = diag(J - 2)}
These parameters can be overwritten by submitting eponymous inputs.
}
}
}

\subsection{Custom priors}{

By default, conjugate priors are used, but custom priors can be
specified as well. To specify a custom prior for parameter \code{<par>},
use the input \code{<par>_prior_custom}, which should compute the density
of the prior (or a value that is proportional), see \sQuote{Examples}.
The following density functions (implemented in {RprobitB}) could be helpful:
\itemize{
\item \code{\link{ddirichlet}}
\item \code{\link{dmvnorm}}
\item \code{\link{dwishart}}
}
}
}

\examples{
### conjugate prior: Omega ~ Inverse-Wishart(P_r + 2, I)
RprobitB_prior_Omega(P_r = 2)

### custom prior: Omega ~ Inverse-Wishart(P_r + 2, I), Omega_11 = Omega_22
RprobitB_prior_Omega(
  P_r = 2,
  Omega_prior_custom = function(x) {
    dwishart(x, df = 4, scale = diag(2), inv = TRUE) * (x[1,1] == x[2,2])
  }
)

}
\keyword{internal}
