% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwishart}
\alias{rwishart}
\title{Draw from a Wishart}
\usage{
rwishart(nu, V)
}
\arguments{
\item{nu}{A double, the degrees of freedom.}

\item{V}{A matrix, the scale matrix.}
}
\value{
A list, the draw from the Wishart (W), inverted Wishart (IW), and
corresponding Cholesky decomposition (C and CI)
}
\description{
Function to draw from Wishart and inverted Wishart distribution.
}
\keyword{utils}
