% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwishart}
\alias{rwishart}
\title{Draw from Wishart distribution}
\usage{
rwishart(nu, V)
}
\arguments{
\item{nu}{A double, the degrees of freedom.}

\item{V}{A matrix, the scale matrix.}
}
\value{
A list, the draws from the Wishart (W), inverted Wishart (IW), and
corresponding Cholesky decompositions (C and CI).
}
\description{
This function draws from a Wishart and inverted Wishart distribution.
}
\keyword{dist}
