% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R, R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choice data}
\usage{
simulate_choices(
  RprobitB_covariates = simulate_RprobitB_covariates(formula, N, J, T = 1, alternatives =
    LETTERS[1:J], re = NULL, ordered = FALSE),
  true_parameter = RprobitB_parameter(),
  ranked = FALSE,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)

simulate_choices(
  RprobitB_covariates = simulate_RprobitB_covariates(formula, N, J, T = 1, alternatives =
    LETTERS[1:J], re = NULL, ordered = FALSE),
  true_parameter = RprobitB_parameter(),
  ranked = FALSE,
  seed = NULL,
  column_choice = "choice",
  column_decider = "id",
  column_occasion = "idc"
)
}
\arguments{
\item{RprobitB_covariates}{An \code{\link{RprobitB_covariates}} object, which contains the covariate
matrices used for the choice data simulation.}

\item{true_parameter}{An \code{\link{RprobitB_parameter}} object, which contains the model
parameters used for the choice data simulation.
By default, \code{RprobitB_parameter = RprobitB_parameter()}, i.e. default
parameters are used.}

\item{ranked}{TODO}

\item{seed}{Set a seed for the simulation.}

\item{covariates}{A named list of covariate values. Each element must be a vector of length
equal to the number of choice occasions and named according to a covariate.
Covariates for which no values are supplied are drawn from a standard normal
distribution.}
}
\value{
An object of class \code{RprobitB_data}.

An \code{\link{RprobitB_data}} object.
}
\description{
This function simulates choice data from a probit model.

This function simulates choice data from a probit model. It helps to create
an \code{\link{RprobitB_data}} object.
}
\details{
See \href{https://loelschlaeger.de/RprobitB/articles/v02_choice_data.html}{the vignette on choice data}
for more details.
}
\section{Choice simulation}{
TODO
}

\section{Model formula}{
The structure of \code{formula} should be
\code{choice ~ A | B | C}, where
\itemize{
\item \code{choice} is the name of the dependent variable (the choices),
\item \code{A} are names of \strong{alternative specific covariates} with
\strong{a coefficient that is constant across alternatives},
\item \code{B} are names of \strong{covariates that are constant across
alternatives},
\item and \code{C} are names of \strong{alternative specific covariates}
with \strong{alternative specific coefficients}.
}

Multiple covariates of one type are separated by a \code{+} sign, i.e.,
\code{choice ~ A1 + A2}.

By default, alternative specific constants (ASCs) are added to the model.
They can be removed by adding \code{+ 0} in the second spot, i.e.,
\code{choice ~ A | B + 0 | C}. To not include any covariates of
the second category but to estimate ASCs, add \code{1} in the second
spot, e.g., \code{choice ~ A | 1 | C}. The expression
\code{choice ~ A | 0 | C} is interpreted as no covariates of the second
category and no ASCs.

In the ordered probit model (\code{ordered = TRUE}), covariates are not
alternative specific, i.e., there exists only one type of covariate.
Therefore, the \code{formula} object does not need the special
separation form via \code{|}, and hence has the simple structure
\code{choice ~ A + B + C}.
ASCs cannot be estimated in the ordered case.
}

\section{Random effects}{
Covariates can have random effects, i.e., their coefficients can follow a
random distribution. Per default, the distribution is normal. The log-normal
distribution (for sign-restriction) can be specified via appending \code{+}
to the corresponding name in \code{re}.
To have random effects for the ASCs, add \code{ASC} (or \code{ASC+}) to
\code{re}.
}

\examples{
### simulate data from a binary probit model with two latent classes
data <- simulate_choices(
  form = choice ~ cost | income | time,
  N = 100,
  T = 10,
  J = 2,
  re = c("cost", "time"),
  alternatives = c("car", "bus"),
  seed = 1,
  true_parameter = list(
    "alpha" = c(-1, 1),
    "b" = matrix(c(-1, -1, -0.5, -1.5, 0, -1), ncol = 2),
    "C" = 2
  )
)

### simulate data from an ordered probit model
data <- simulate_choices(
  form = opinion ~ age + gender,
  N = 10,
  T = 1:10,
  J = 5,
  alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
  ordered = TRUE,
  covariates = list(
    "gender" = rep(sample(c(0,1), 10, replace = TRUE), times = 1:10)
    ),
  seed = 1
)

### simulate data from a ranked probit model
data <- simulate_choices(
  form = product ~ price,
  N = 10,
  T = 1:10,
  J = 3,
  alternatives = c("A", "B", "C"),
  ranked = TRUE,
  seed = 1
)

### simulate data from a binary probit model with two latent classes
data <- simulate_choices(
  formula = choice ~ cost | income | time, N = 10, J = 2, T = 1:10,
  alternatives = c("car", "bus"), re = c("cost", "time"),
  true_parameter = RprobitB_parameter(C = 2)
)

### simulate data from an ordered probit model
data <- simulate_choices(
  formula = opinion ~ age + gender, N = 50, J = 5,
  alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
  ordered = TRUE
)

### simulate data from a ranked probit model
data <- simulate_choices(
  formula = product ~ price, N = 10, J = 3, T = 1:10, ranked = TRUE
)

}
\seealso{
\itemize{
\item ... for checking the model formula TODO
\item \code{\link[=overview_effects]{overview_effects()}} for an overview of the model effects
\item \code{\link[=create_lagged_cov]{create_lagged_cov()}} for creating lagged covariates
\item \code{\link[=as_cov_names]{as_cov_names()}} for re-labeling alternative-specific covariates
\item \code{\link[=prepare_data]{prepare_data()}} for preparing empirical choice data
\item \code{\link[=train_test]{train_test()}} for splitting choice data into a train and test subset
}

\itemize{
\item TODO
}
}
