% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choice data.}
\usage{
simulate_choices(
  form,
  N,
  T,
  J,
  re = NULL,
  alternatives = NULL,
  covariates = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative).
They can be removed by adding \code{+0} in the second spot.
See the vignette \code{vignette("data_management", package = "RprobitB")}
for more details.}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{covariates}{A named list of covariate values. Each element must be a vector of length
equal to the number of choice occasions and named according to a covariate.
Covariates for which no values are supplied are drawn from a standard normal
distribution.}

\item{seed}{Set a seed for the simulation.}

\item{...}{Optionally specify \code{alpha}, \code{C}, \code{s}, \code{b}, \code{Omega},
\code{Sigma}, \code{Sigma_full}, \code{beta}, or \code{z} for the simulation.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function simulates choice data for the RprobitB package.
}
\details{
See the vignette "Choice data" for more details:
\code{vignette("choice_data", package = "RprobitB")}.
}
\examples{
data <- simulate_choices(
  form = choice ~ cost | income | time,
  N = 100,
  T = 10,
  J = 2,
  re = c("cost", "time"),
  alternatives = c("car", "bus"),
  seed = 1,
  alpha = c(-1, 1),
  C = 2
)

}
