% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R, R/simulate_choices.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choice data}
\usage{
simulate_choices(
  formula,
  N,
  T = 1,
  J = 3,
  alternatives = LETTERS[1:J],
  re = NULL,
  ordered = FALSE,
  ranked = FALSE,
  covariates = function(n, t) {
     P <- P(formula, re, J)
     matrix(rnorm(P * J, mean
    = 0, sd = 9), nrow = P, ncol = J)
 },
  seed = NULL,
  true_parameter = RprobitB_parameter(formula = formula, re = re, N = N, J = J, C = 1,
    ordered = ordered, seed = seed)
)

simulate_choices(
  formula,
  N,
  T = 1,
  J = 3,
  alternatives = LETTERS[1:J],
  re = NULL,
  ordered = FALSE,
  ranked = FALSE,
  covariates = function(n, t) {
     P <- P(formula, re, J)
     matrix(rnorm(P * J, mean
    = 0, sd = 9), nrow = P, ncol = J)
 },
  seed = NULL,
  true_parameter = RprobitB_parameter(formula = formula, re = re, N = N, J = J, C = 1,
    ordered = ordered, seed = seed)
)
}
\arguments{
\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).
Per default, \code{T = 1}.}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.
If \code{ordered = TRUE}, \code{alternatives} is assumed to be specified with
the alternatives ordered from worst to best.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the ASCs, include \code{"ASC"} in \code{re}.}

\item{ordered}{A boolean, \code{FALSE} per default. If \code{TRUE}, the choice set
\code{alternatives} is assumed to be ordered from worst to best.}

\item{ranked}{TBA}

\item{covariates}{A named list of covariate values. Each element must be a vector of length
equal to the number of choice occasions and named according to a covariate.
Covariates for which no values are supplied are drawn from a standard normal
distribution.}

\item{seed}{Set a seed for the simulation.}

\item{true_parameter}{Optionally specify a named list with true parameter values for \code{alpha},
\code{C}, \code{s}, \code{b}, \code{Omega}, \code{Sigma}, \code{Sigma_full},
\code{beta}, \code{z}, or \code{d} for the simulation.
See \href{https://loelschlaeger.de/RprobitB/articles/v01_model_definition.html}{the vignette on model definition}
for definitions of these variables.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function simulates choice data from a probit model.

This function simulates choice data from a probit model.
}
\details{
See \href{https://loelschlaeger.de/RprobitB/articles/v02_choice_data.html}{the vignette on choice data}
for more details.
}
\section{Choice simulation}{
TODO
}

\examples{
### simulate data from a binary probit model with two latent classes
data <- simulate_choices(
  form = choice ~ cost | income | time,
  N = 100,
  T = 10,
  J = 2,
  re = c("cost", "time"),
  alternatives = c("car", "bus"),
  seed = 1,
  true_parameter = list(
    "alpha" = c(-1, 1),
    "b" = matrix(c(-1, -1, -0.5, -1.5, 0, -1), ncol = 2),
    "C" = 2
  )
)

### simulate data from an ordered probit model
data <- simulate_choices(
  form = opinion ~ age + gender,
  N = 10,
  T = 1:10,
  J = 5,
  alternatives = c("very bad", "bad", "indifferent", "good", "very good"),
  ordered = TRUE,
  covariates = list(
    "gender" = rep(sample(c(0,1), 10, replace = TRUE), times = 1:10)
    ),
  seed = 1
)

### simulate data from a ranked probit model
data <- simulate_choices(
  form = product ~ price,
  N = 10,
  T = 1:10,
  J = 3,
  alternatives = c("A", "B", "C"),
  ranked = TRUE,
  seed = 1
)

}
\seealso{
\itemize{
\item \code{\link[=check_form]{check_form()}} for checking the model formula
\item \code{\link[=overview_effects]{overview_effects()}} for an overview of the model effects
\item \code{\link[=create_lagged_cov]{create_lagged_cov()}} for creating lagged covariates
\item \code{\link[=as_cov_names]{as_cov_names()}} for re-labeling alternative-specific covariates
\item \code{\link[=prepare_data]{prepare_data()}} for preparing empirical choice data
\item \code{\link[=train_test]{train_test()}} for splitting choice data into a train and test subset
}
}
