% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{simulate_choices}
\alias{simulate_choices}
\title{Simulate choice data}
\usage{
simulate_choices(
  form,
  N,
  T,
  J,
  re = NULL,
  alternatives = NULL,
  base = NULL,
  covariates = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{form}{A formula object that is used to specify the probit model.
The structure is \code{choice ~ A | B | C}, where
\itemize{
\item \code{A} are names of alternative and choice situation specific
covariates with a generic coefficient,
\item \code{B} are names of choice situation specific covariates with
alternative specific coefficients,
\item and \code{C} are names of alternative and choice situation specific
covariates with alternative specific coefficients.
}
Separate multiple covariates of one type by a \code{+} sign.
By default, alternative specific constants (ASCs) are added to the model
(for all except for the last alternative due to identifiability).
They can be removed by adding \code{+0} in the second spot.
See the vignette on choice data for more details.}

\item{N}{The number (greater or equal 1) of decision makers.}

\item{T}{The number (greater or equal 1) of choice occasions or a vector of choice
occasions of length \code{N} (i.e. a decision maker specific number).}

\item{J}{The number (greater or equal 2) of choice alternatives.}

\item{re}{A character (vector) of covariates of \code{form} with random effects.
If \code{re = NULL} (the default), there are no random effects.
To have random effects for the alternative specific constants, include
\code{"ASC"} in \code{re}.}

\item{alternatives}{A character vector with the names of the choice alternatives.
If not specified, the choice set is defined by the observed choices.}

\item{base}{A character, the name of the base alternative for covariates that are not
alternative specific (type 2 covariates).}

\item{covariates}{A named list of covariate values. Each element must be a vector of length
equal to the number of choice occasions and named according to a covariate.
Covariates for which no values are supplied are drawn from a standard normal
distribution.}

\item{seed}{Set a seed for the simulation.}

\item{...}{Optionally specify \code{alpha}, \code{C}, \code{s}, \code{b}, \code{Omega},
\code{Sigma}, \code{Sigma_full}, \code{beta}, or \code{z} for the simulation.}
}
\value{
An object of class \code{RprobitB_data}.
}
\description{
This function simulates choice data.
}
\details{
See \href{https://loelschlaeger.de/RprobitB/articles/v02_choice_data.html}{the vignette on choice data}
for more details.
}
\examples{
data <- simulate_choices(
  form = choice ~ cost | income | time,
  N = 100,
  T = 10,
  J = 2,
  re = c("cost", "time"),
  alternatives = c("car", "bus"),
  seed = 1,
  alpha = c(-1, 1),
  b = matrix(c(-1, -1, -0.5, -1.5, 0, -1), ncol = 2),
  C = 2
)
}
\seealso{
\itemize{
\item \code{\link[=check_form]{check_form()}} for checking the model formula
\item \code{\link[=overview_effects]{overview_effects()}} for an overview of the model effects
\item \code{\link[=create_lagged_cov]{create_lagged_cov()}} for creating lagged covariates
\item \code{\link[=as_cov_names]{as_cov_names()}} for renaming alternative-specific covariates
\item \code{\link[=prepare_data]{prepare_data()}} for preparing empirical choice data
\item \code{\link[=train_test]{train_test()}} for splitting choice data into a train and test subset
}
}
