% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform fitted \code{RprobitB_model}}
\usage{
transform(RprobitB_model, B = NULL, Q = NULL, scale = NULL)
}
\arguments{
\item{RprobitB_model}{An object of class \code{RprobitB_model}. Such an object is generated by the
function \code{\link{fit}}.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{scale}{A list of three elements, determining the parameter normalization with
respect to the scale of utility:
\itemize{
\item \code{parameter}:
either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"})
\item \code{index}:
the index of the parameter that gets fixed
\item \code{value}:
the value for the fixed parameter
}}
}
\value{
An object of class \code{RprobitB_model}.
}
\description{
Function that transforms a fitted \code{RprobitB_model} in three ways:
\itemize{
\item Change the length \code{B} of the burn-in period
\item Change the thinning factor \code{Q}
\item Change the utility normalization \code{scale}
}
}
\details{
For more details see the vignette "Model fitting":
\code{vignette("model_fitting", package = "RprobitB")}
}
