% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform fitted \code{RprobitB_model}}
\usage{
transform(model, B = NULL, Q = NULL, scale = NULL)
}
\arguments{
\item{model}{An object of class \code{RprobitB_model}. Such an object is generated by the
function \code{\link{fit}}.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{scale}{A list of three elements, determining the parameter normalization with
respect to the scale of utility:
\itemize{
\item \code{parameter}:
either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"})
\item \code{index}:
the index of the parameter that gets fixed
\item \code{value}:
the value for the fixed parameter
}}
}
\value{
An object of class \code{RprobitB_model}.
}
\description{
For an \code{RprobitB_model}, this function can:
\itemize{
\item change the length \code{B} of the burn-in period,
\item change the thinning factor \code{Q},
\item change the utility normalization \code{scale}.
}
}
\details{
For more details see the vignette "Model fitting":
\code{vignette("model_fitting", package = "RprobitB")} or "Model checking and
model selection":
\code{vignette("model_checking_and_model_selection", package = "RprobitB")}
}
\examples{
### fit a model to simulated data
data = simulate(form = choice ~ var | 1, N = 100, T = 10, J = 2)
scale = list("parameter" = "s", "index" = 1, "value" = 1)
model = fit(data = data, R = 1e4, B = 1, Q = 1)

### change 'B' and 'Q'
model = transform(model = model, B = 1e4/2, Q = 10)

### change 'scale'
scale = list("parameter" = "a", "index" = 1, "value" = -1)
model = transform(model = model, scale = scale)
}
