% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Change the length of the burn-in period, the thinning factor and the scale
of a fitted \code{RprobitB_model}.}
\usage{
transform(x, B = NULL, Q = NULL, scale = NULL, check_preference_flip = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{RprobitB_model}}.}

\item{B}{The length of the burn-in period, i.e. a non-negative number of samples to
be discarded.}

\item{Q}{The thinning factor for the Gibbs samples, i.e. only every \code{Q}th
sample is kept.}

\item{scale}{A named list of three elements, determining the parameter normalization with
respect to the utility scale:
\itemize{
\item \code{parameter}:
Either \code{"a"} (for a linear coefficient of \code{"alpha"}) or
\code{"s"} (for a variance of the error-term covariance matrix
\code{"Sigma"}).
\item \code{index}:
The index of the parameter that gets fixed.
\item \code{value}:
The value for the fixed parameter.
}}

\item{check_preference_flip}{If \code{TRUE} check for flip in preferences with new scale.}
}
\value{
An object of class \code{RprobitB_model}.
}
\description{
Given a fitted \code{RprobitB_model}, this function can:
\itemize{
\item change the length \code{B} of the burn-in period and the thinning
factor \code{Q} of the Gibbs sampler,
\item change the model \code{scale}.
}
}
\details{
For more details see the vignette "Model fitting":
\code{vignette("model_fitting", package = "RprobitB")}.
}
\examples{
\dontrun{
### fit a model to simulated data
data = simulate(form = choice ~ var | 1, N = 100, T = 10, J = 2,
                Sigma = 2, alpha = c(0.5,2))
scale = list("parameter" = "s", "index" = 1, "value" = 1)
m1 = mcmc(data = data, R = 1e4, B = 1, Q = 1, scale = scale)

### change 'B' and 'Q'
m2 = transform(m1, B = 5e3, Q = 10)

### change 'scale'
scale_new = list("parameter" = "a", "index" = 1, "value" = 1)
m3 = transform(m2, scale = scale_new)
}
}
