% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_parameter.R
\name{transform_parameter}
\alias{transform_parameter}
\title{Transformation of parameter values.}
\usage{
transform_parameter(parameter, normalization)
}
\arguments{
\item{parameter}{An object of class \code{RprobitB_parameter}.}

\item{normalization}{An object of class \code{RprobitB_normalization}.}
}
\value{
An object of class \code{RprobitB_parameter}.
}
\description{
This function transforms parameter values based on \code{normalization}.
}
