% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_true_parameter}
\alias{transform_true_parameter}
\title{Transformation of true parameter values.}
\usage{
transform_true_parameter(true_parameter, normalization)
}
\arguments{
\item{true_parameter}{An object of class \code{RprobitB_true_parameter}.}

\item{normalization}{An object of class \code{RprobitB_normalization}.}
}
\value{
An object of class \code{RprobitB_parameter}.
}
\description{
This function transforms the true parameter values 'true_parameter' based
on 'normalization$scale'.
}
\examples{
true_parameter = RprobitB_parameter(P_f = 2, P_r = 2, J = 3, N = 100)
normalization = RprobitB_normalization(J = 3, P_f = 2)
transform_true_parameter(true_parameter = true_parameter,
                         normalization = normalization)
}
