% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_d}
\alias{update_d}
\title{Update utility threshold increments}
\usage{
update_d(d, y, mu, ll, zeta, Z, Tvec)
}
\value{

}
\description{
This function updates the utility threshold increments
}
\details{

}
\examples{

}
\keyword{posterior}
