% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_d}
\alias{update_d}
\title{Update utility threshold increments}
\usage{
update_d(d, y, mu, ll, zeta, Z, Tvec)
}
\arguments{
\item{d}{The current vector of utility threshold increments.}

\item{y}{A matrix of the choices.}

\item{mu}{A matrix of the systematic utilities.}

\item{ll}{Current log-likelihood value.}

\item{zeta}{The mean vector of the normal prior for \code{d}.}

\item{Z}{The covariance matrix of the normal prior for \code{d}.}

\item{Tvec}{The element \code{Tvec} in \code{\link{sufficient_statistics}}.}
}
\value{
The updated value for \code{d}.
}
\description{
Update utility threshold increments
}
\keyword{gibbs_sampler}
