% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_m}
\alias{update_m}
\title{Update class sizes}
\usage{
update_m(C, z, non_zero = FALSE)
}
\arguments{
\item{C}{[\code{integer(1)}]\cr
The number (greater or equal 1) of latent classes of decision makers.}

\item{z}{[\code{numeric(N)}]\cr
The decider class allocations.}

\item{non_zero}{[\code{logical(1)}]\cr
Enforce strictly positive values in \code{m} (for numerical stability)?}
}
\value{
An update for \code{m}.
}
\description{
Update class sizes
}
\examples{
update_m(C = 4, z = c(1, 1, 1, 2, 2, 3))

}
\keyword{gibbs_sampler}
