% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_m}
\alias{update_m}
\title{Update class sizes}
\usage{
update_m(C, z, nozero = FALSE)
}
\arguments{
\item{C}{The number (greater or equal 1) of latent classes of decision makers.
Set to \code{NA} if \code{P_r = 0}. Otherwise, \code{C = 1} per default.}

\item{z}{The vector of the allocation variables of length \code{N}.
Set to \code{NA} if \code{P_r = 0}.}

\item{nozero}{If \code{TRUE}, each element in the output vector \code{m} is at least one
(for numerical stability).}
}
\value{
An updated class size vector.
}
\description{
This function updates the class size vector.
}
\examples{
update_m(C = 3, z = c(0,1,1,2,2,2))
}
\keyword{posterior}
