% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_m}
\alias{update_m}
\title{Update class sizes}
\usage{
update_m(C, z, nozero)
}
\arguments{
\item{C}{An \code{integer}, the number (greater or equal 1) of latent classes of
decision makers.
By default, \code{C = 1}.}

\item{z}{A \code{numeric} of length \code{N}, the vector of the allocation variables.
Entry \code{n} of \code{z} is an integer from \code{1} to \code{C} and
denotes the allocated class for decider \code{n}.}

\item{nozero}{If \code{TRUE}, each element in the output vector \code{m} is at least one
(for numerical stability).}
}
\value{
An updated class size vector.
}
\description{
This function updates the class size vector.
}
\examples{
update_m(C = 3, z = c(1,1,1,2,2,3), FALSE)
}
\keyword{internal}
\keyword{posterior}
