% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_m}
\alias{update_m}
\title{Update class sizes \code{m}}
\usage{
update_m(C, z, non_zero = FALSE)
}
\arguments{
\item{C}{The number (greater or equal 1) of latent classes of decision makers.
Set to \code{NA} if \code{P_r = 0}. Otherwise, \code{C = 1} per default.}

\item{z}{The vector of the allocation variables of length \code{N}.
Set to \code{NA} if \code{P_r = 0}.}

\item{non_zero}{[\code{logical(1)}]\cr
Enforce strictly positive values in \code{m} (for numerical stability)?}
}
\value{
An update for \code{m}.
}
\description{
Update class sizes \code{m}
}
\examples{
update_m(C = 3, z = c(1, 1, 1, 2, 2, 3))

}
\keyword{gibbs_sampler}
