% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_Omega_c}
\alias{update_Omega_c}
\title{Update covariance of a single class}
\usage{
update_Omega_c(S_c, m_c, n_Omega_0, V_Omega_0)
}
\arguments{
\item{S_c}{[\code{matrix(P_r, P_r)}]\cr
The scatter matrix of this class.}

\item{m_c}{[\code{integer(1)}]\cr
The number of observations in this class.}

\item{n_Omega_0}{[\code{integer(1)}]\cr
The degrees of freedom of the Inverse Wishart prior for each \code{Omega_c}.}

\item{V_Omega_0}{[\code{matrix(P_r, P_r)}]\cr
The scale matrix of the Inverse Wishart prior for each \code{Omega_c}.}
}
\value{
An update for \code{Omega_c}.
}
\description{
Update covariance of a single class
}
\examples{
update_Omega_c(S_c = diag(2), m_c = 10, n_Omega_0 = 4, V_Omega_0 = diag(2))

}
\keyword{gibbs_sampler}
