% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_s}
\alias{update_s}
\title{Update class weight vector}
\usage{
update_s(delta, m)
}
\arguments{
\item{delta}{The concentration parameter of length 1 of the Dirichlet prior for \code{s}.}

\item{m}{The vector of current class sizes.}
}
\value{
A draw from the Dirichlet posterior for \code{s}.
}
\description{
This function updates the class weight vector by drawing from its posterior.
}
\keyword{posterior}
