% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_U_ranked}
\alias{update_U_ranked}
\title{Update ranked utility vector}
\usage{
update_U_ranked(U, sys, Sigma_inv)
}
\arguments{
\item{U}{[\code{numeric(J - 1)}]\cr
The current utility vector.}

\item{sys}{[\code{numeric(J - 1)}]\cr
The systematic utility.}

\item{Sigma_inv}{[\code{matrix(J - 1, J - 1)}]\cr
The inverted error covariance matrix.}
}
\value{
An update for (a single) ranked \code{U}.
}
\description{
Update ranked utility vector
}
\examples{
U <- sys <- c(0, 0)
Sigma_inv <- diag(2)
update_U_ranked(U, sys, Sigma_inv)

}
\keyword{gibbs_sampler}
