% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_z}
\alias{update_z}
\title{Update class allocation vector}
\usage{
update_z(s, beta, b, Omega)
}
\arguments{
\item{s}{The vector of class weights of length \code{C}.
Set to \code{NA} if \code{P_r = 0}.
For identifiability, the vector must be non-ascending.}

\item{beta}{The matrix of the decision-maker specific coefficient vectors of dimension
\code{P_r} x \code{N}.
Set to \code{NA} if \code{P_r = 0}.}

\item{b}{The matrix of class means as columns of dimension \code{P_r} x \code{C}.
Set to \code{NA} if \code{P_r = 0}.}

\item{Omega}{The matrix of class covariance matrices as columns of dimension
\code{P_r*P_r} x \code{C}.
Set to \code{NA} if \code{P_r = 0}.}
}
\value{
An updated class allocation vector.
}
\description{
This function updates the class allocation vector independently for all observations
by drawing from its conditional distribution.
}
\keyword{posterior}
