% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Compute WAIC value}
\usage{
waic(x, S = 1000, print_progress = TRUE, check_conv = TRUE, ncores = 1)
}
\arguments{
\item{x}{An object of class \code{RprobitB_fit}.}

\item{S}{The number of posterior samples used for the calculation of the WAIC value.
Must be greater or equal two for variance computation.}

\item{print_progress}{Set to \code{TRUE} to print computation progress.}

\item{check_conv}{Set to \code{TRUE} to plot the convergence behavior of the WAIC calculation.}

\item{ncores}{Computation is parallelized, set the number of cores.}
}
\value{
Invisibly returns a list of
\itemize{
\item \code{waic}, the WAIC value,
\item \code{se_waic}, the standard error of \code{waic},
\item \code{p_si}, the matrix of probabilities of each observation for the
\code{S} different posterior samples,
\item \code{S}, the number of posterior samples,
\item \code{lppd}, the log pointwise predictive density,
\item \code{p_waic}, the effective number of parameters,
\item \code{p_waic_i}, the vector of summands of \code{p_waic}.
}
}
\description{
This function computes the WAIC value of an \code{RprobitB_fit}-object.
}
\examples{
data <- simulate_choices(
  form = choice ~ cov | 0,
  N = 10,
  T = 10,
  J = 2,
  seed = 1
)
x <- mcmc(data, R = 1000)
RprobitB:::waic(x = x, S = 10, check_conv = TRUE)

}
\keyword{internal}
